% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{addZScores}
\alias{addZScores}
\title{Add Z scores to DE results}
\usage{
addZScores(df)
}
\arguments{
\item{df}{Data.frame with the columns "pval", "padj" and "log2FoldChange"}
}
\value{
Updated data.frame with Z scores
}
\description{
Add Z scores to DE results
}
\examples{
\dontrun{
df_adj <- addZScores(df)
}

}
