% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{appendStatisticsToDE}
\alias{appendStatisticsToDE}
\title{Append statistics to DE results}
\usage{
appendStatisticsToDE(de.list, expr.frac.per.type)
}
\arguments{
\item{de.list}{list with DE results}

\item{expr.frac.per.type}{}
}
\description{
Append statistics to DE results
}
\keyword{internal}
