% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_effects.R
\name{asSquaredDistance}
\alias{asSquaredDistance}
\title{Map a distance matrix to the matrix used as a squared Euclidean distance}
\usage{
asSquaredDistance(D, dist = c("cor", "l2", "l1"))
}
\arguments{
\item{D}{numeric sample x sample distance matrix (symmetric, zero diagonal)}

\item{dist}{distance type: \code{"cor"} (gene-centred cosine distance \code{1 - cos}, which already equals half a
squared Euclidean distance of unit-norm centred profiles), \code{"l2"} (Euclidean; squared here) or \code{"l1"}
(Manhattan; of negative type, used as is: location/dispersion separation is approximate)}
}
\value{
numeric matrix playing the role of \code{D^2}
}
\description{
Map a distance matrix to the matrix used as a squared Euclidean distance
}
