% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{brewerPalette}
\alias{brewerPalette}
\title{Helper function for creating color palettes}
\usage{
brewerPalette(name, n = NULL, rev = TRUE)
}
\arguments{
\item{name}{character A palette name; please refer to RColorBrewer::brewer.pal()}

\item{n}{integer (default=NULL) Number of different colors in the palette, minimum 3, maximum depending on palette. Please refer to RColorBrewer::brewer.pal() for more details}

\item{rev}{boolean (default=TRUE) Whether to reverse the palette order}
}
\value{
palette function
}
\description{
Helper function for creating color palettes
}
