% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{calcBalancesOnCellSets}
\alias{calcBalancesOnCellSets}
\title{Calculate balances based on the contrast defines by two sets of cells}
\usage{
calcBalancesOnCellSets(d.counts, cell.set1, cell.set2)
}
\arguments{
\item{d.counts}{Cell count matrix}

\item{cell.set1}{Set#1}

\item{cell.set2}{Set#1}
}
\value{
Balance values
}
\description{
Calculate balances based on the contrast defines by two sets of cells
}
