% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_checks.R
\name{checkDataAndCells}
\alias{checkDataAndCells}
\title{Check data and list of cells}
\usage{
checkDataAndCells(d.counts, cells)
}
\arguments{
\item{d.counts}{Cell count table}

\item{cells}{List of cell types of interest}
}
\description{
Check data and list of cells
}
\keyword{internal}
