% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{checkDataGroups}
\alias{checkDataGroups}
\title{Check cell count data and groups}
\usage{
checkDataGroups(d.counts, d.groups)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Groups variable for samples}
}
\description{
Check cell count data and groups
}
\keyword{internal}
