% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{checkDataSbp}
\alias{checkDataSbp}
\title{Check the agreement between sbp and cell count table}
\usage{
checkDataSbp(d.counts, sbp)
}
\arguments{
\item{d.counts}{Cell count table}

\item{sbp}{Sbp matrix: each row - partition}
}
\description{
Check the agreement between sbp and cell count table
}
\keyword{internal}
