% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{checkDataTree}
\alias{checkDataTree}
\title{Check agreement between Tree and the Cell count table}
\usage{
checkDataTree(d.counts, tree)
}
\arguments{
\item{d.counts}{Cell count table}

\item{tree}{phylo tree}
}
\description{
Check agreement between Tree and the Cell count table
}
\keyword{internal}
