% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_checks.R
\name{checkSbp}
\alias{checkSbp}
\title{Check sequential binary partitions(sbp)}
\usage{
checkSbp(sbp)
}
\arguments{
\item{sbp}{Sbp matrix: each row - partition}
}
\description{
Check sequential binary partitions(sbp)
}
\keyword{internal}
