% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{checkSbpWhole}
\alias{checkSbpWhole}
\title{Check sequential binary partitions(sbp) for the whole set of balances}
\usage{
checkSbpWhole(sbp)
}
\arguments{
\item{sbp}{Sbp matrix: each row - partition}
}
\description{
Check sequential binary partitions(sbp) for the whole set of balances
}
\keyword{internal}
