% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_shifts.R
\name{cluster.expression.distances}
\alias{cluster.expression.distances}
\title{expression distance between clusters}
\usage{
cluster.expression.distances(
  con,
  groups = NULL,
  dist = "JS",
  n.cores = con$n.cores,
  min.cluster.size = 1,
  min.samples = 1,
  max.n.cells = Inf,
  aggr = median,
  return.details = FALSE,
  use.aggregated.matrices = FALSE,
  use.single.cell.comparisons = FALSE,
  n.cells = 100
)
}
\arguments{
\item{con}{conos object}

\item{groups}{clustering factor}

\item{dist}{what distance measure to use: 'JS' - Jensen-Shannon divergence, 'cor' - Pearson's linear correlation on log transformed values}

\item{n.cores}{number of cores}

\item{min.cluster.size}{minimum number of cells in a cluster (in a sample) for the distance to be estimated. default: 1}

\item{min.samples}{minimum number of samples in which a given pair of clusters has been compared (i.e. the two clusters were present in sufficient size). default:1}

\item{max.n.cells}{maximum number of cells to take from a cluster (to avoid very large clusters,default: Inf)}

\item{aggr}{aggregation function (default: median)}

\item{return.details}{whether to return a list of individual sample distances}

\item{use.aggregated.matrices}{whether to simply aggregate all the molecules from all of the samples in which cluster appears (i.e. don't do aggregation of per-sample distance estimates, but just add all the data and measure the distance once)}

\item{use.single.cell.comparisons}{whether instead of adding up all molecules from a cluster in a given sample, instead compare n.cell draws of single cells (note: doesn't seem to work that well)}

\item{n.cells}{number of single cells to sample when use.single.cell.comparisons=T}
}
\value{
distance matrix (with possible NAs), or if return.details=T a list with a distance matrix ($mdist) and a list of distance matrices computed on the individual samples ($dc)
}
\description{
Evaluate expression distances between clusters, separately within each sample, combining, optionally controling for cluster sizes.
}
\details{
When using per-sample distance estimation (default, unless usr.aggregated.matrices=T), the distances for the cluster pairs that
are not seen sufficient number of times together (see min.samples) are set to NA.
}
