% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusteredOntologyDotplot}
\alias{clusteredOntologyDotplot}
\title{Performs ontology clustering by genes and then shows medoids of the clusters with
enrichplot::dotplot}
\usage{
clusteredOntologyDotplot(
  ont.res,
  p.adj = 0.05,
  min.genes = 1,
  cut.h = 0.66,
  top.n = Inf,
  ...
)
}
\arguments{
\item{ont.res}{something}

\item{p.adj}{numeric Adjusted p-value set (default=0.05)}

\item{min.genes}{integer Minimun number of genes (default=1)}

\item{cut.h}{numeric (default=0.66)}

\item{top.n}{integer Number of enriched terms to display (default=Inf). If Inf, no limit is set.}

\item{...}{additional parameters passed to enrichplot::dotplot()}
}
\value{
enrichplot::dotplot showing medoids of gene clustering
}
\description{
Performs ontology clustering by genes and then shows medoids of the clusters with
enrichplot::dotplot
}
