% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{clusterGOsPerType}
\alias{clusterGOsPerType}
\title{Cluster ontology terms per cell type}
\usage{
clusterGOsPerType(clust.df, cut.h)
}
\arguments{
\item{clust.df}{Cluster dataframe}

\item{cut.h}{numeric scalar or vector with heights where the tree should be cut.}
}
\description{
Cluster ontology terms per cell type
}
\keyword{internal}
