% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{clusterIndividualGOs}
\alias{clusterIndividualGOs}
\title{Cluster individual ontology terms}
\usage{
clusterIndividualGOs(genes.per.go.per.type, cut.h)
}
\arguments{
\item{genes.per.go.per.type}{Genes per term per cell type}

\item{cut.h}{numeric scalar or vector with heights where the tree should be cut.}
}
\value{

}
\description{
Cluster individual ontology terms
}
\keyword{internal}
