% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{collapseFamilies}
\alias{collapseFamilies}
\title{Collapse ontology families}
\usage{
collapseFamilies(ont.res)
}
\arguments{
\item{ont.res}{Results from identifyFamilies}
}
\value{
List of collapsed families and ontology data
}
\description{
Collapse ontology families based on overlaps between parents and/or children of significant ontology terms
}
\keyword{internal}
