% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_distance.R
\name{consensusShiftDistances}
\alias{consensusShiftDistances}
\title{Calculate consensus change direction and distances between samples along this axis}
\usage{
consensusShiftDistances(
  tcm,
  sample.groups,
  use.median = FALSE,
  mean.trim = 0,
  use.cpp = TRUE
)
}
\description{
Calculate consensus change direction and distances between samples along this axis
}
\keyword{internal}
