% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_trees.R
\name{constructCanonicalTree}
\alias{constructCanonicalTree}
\title{Construct the canonical tree}
\usage{
constructCanonicalTree(d.counts, d.groups)
}
\arguments{
\item{d.counts}{Table with cell type counts}

\item{d.groups}{Groups variable for samples}
}
\value{
phylo tree
}
\description{
Construct the canonical tree
}
\keyword{internal}
