% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{constructTreeScore}
\alias{constructTreeScore}
\title{Construct a supervised SBP tree using a continuous contrast score (top-down)
This replaces legacy constructTree() but uses a continuous score.}
\usage{
constructTreeScore(
  cnts,
  score,
  partition.thresh = 0,
  zero.pseudocount = 0.1,
  basis.type = c("default")
)
}
\description{
Construct a supervised SBP tree using a continuous contrast score (top-down)
This replaces legacy constructTree() but uses a continuous score.
}
\keyword{internal}
