% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_plots.R
\name{createDendrogram}
\alias{createDendrogram}
\title{helper function for creating dendograms}
\usage{
createDendrogram(
  dend.data,
  angle = 90,
  plot.theme = theme_get(),
  font.size = 3,
  hjust = 1
)
}
\description{
helper function for creating dendograms
}
\keyword{internal}
