% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{diffCellDensity}
\alias{diffCellDensity}
\title{Estimate differential cell density}
\usage{
diffCellDensity(
  density.mat,
  sample.groups,
  ref.level,
  target.level,
  type = "subtract"
)
}
\arguments{
\item{density.mat}{estimated cell density matrix with estimateCellDensity}

\item{sample.groups}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{ref.level}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{target.level}{target/disease level for sample.group vector}

\item{type}{method to calculate differential cell density of each bin; subtract: target density minus ref density; entropy: estimated kl divergence entropy between sample groups ; t.test: zscore of t-test,
global variance is setting for t.test}
}
\description{
Estimate differential cell density
}
\keyword{internal}
