% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{diffCellDensityPermutations}
\alias{diffCellDensityPermutations}
\title{Estimate differential cell density with permutation testing}
\usage{
diffCellDensityPermutations(
  density.mat,
  sample.model,
  perm.method = c("block", "freedman-lane"),
  n.permutations = 999,
  robust.method = c("none", "huber", "winsor"),
  na.mode = c("drop", "impute_weak"),
  alternative = c("two-sided", "greater", "less"),
  return.residuals = FALSE,
  return.sampled.stats = TRUE,
  n.cores = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{density.mat}{estimated cell density matrix with estimateCellDensity}

\item{sample.meta}{sample metadata data.frame}

\item{contrast}{contrast matrix for the design}

\item{sample.groups}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{ref.level}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{target.level}{target/disease level for sample.group vector}

\item{type}{method to calculate differential cell density of each bin;}

\item{block.id}{block variable for permutation test}
}
\description{
Estimate differential cell density with permutation testing
}
\keyword{internal}
