% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthology.R
\name{distanceBetweenTerms}
\alias{distanceBetweenTerms}
\title{Distance between terms}
\usage{
distanceBetweenTerms(type = NULL, ont.res)
}
\arguments{
\item{type}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{pathway.data}{Results from preparePathwayData (default: stored list)}
}
\value{
Distance matrix
}
\description{
Calculate distance matrix between onthology terms
}
