% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_inference.R
\name{drawPermutations}
\alias{drawPermutations}
\title{Draw permutations according to a plan}
\usage{
drawPermutations(plan, B = plan$n.permutations)
}
\arguments{
\item{plan}{output of \code{\link[=permutationPlan]{permutationPlan()}}}

\item{B}{number of permutations (ignored when the plan is exhaustive)}
}
\value{
integer matrix (n x B); column b is \code{p} such that sample i takes the labels of sample \code{p[i]}
(i.e. the permuted design is \code{X[p, ]}); identity for samples outside the permuted set
}
\description{
Draw permutations according to a plan
}
