% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{enrichGOOpt}
\alias{enrichGOOpt}
\title{Enrich gene ontologies using enricher_internal}
\usage{
enrichGOOpt(
  gene,
  org.db,
  go.environment,
  keyType = "ENTREZID",
  ont = "BP",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  qvalueCutoff = 0.2,
  minGSSize = 5,
  maxGSSize = 500,
  readable = FALSE
)
}
\arguments{
\item{gene}{a vector of entrez gene IDs}

\item{org.db}{organism db, e.g. org.Hs.eg.db::org.Hs.eg.db}

\item{go.environment}{GO environment, see DOSE:::enricher_internal, parameter USER_DATA}

\item{keyType}{Gene key type (default="ENTREZID")}

\item{ont}{type of ontology, one of BP, CC, or MF (default="BP")}

\item{readable}{See DOSE:::enricher_internal (default=FALSE)}
}
\description{
Enrich gene ontologies using enricher_internal
}
\keyword{internal}
