% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{enrichGSEGOOpt}
\alias{enrichGSEGOOpt}
\title{Estimate enriched GSEA terms using GSEA_internal}
\usage{
enrichGSEGOOpt(
  gene.ids,
  org.db,
  organism,
  keyType = "ENTREZID",
  go.environment,
  ont = "BP",
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  minGSSize = 5,
  maxGSSize = 500,
  readable = FALSE,
  eps = 0,
  exponent = 1,
  seed = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gene.ids}{a vector of gene IDs}

\item{org.db}{organism db, e.g. org.Hs.eg.db::org.Hs.eg.db}

\item{organism}{Organism, e.g. "Hs"}

\item{keyType}{(default="ENTREZID")}

\item{go.environment}{GO environment, see DOSE:::enricher_internal, parameter USER_DATA}

\item{ont}{type of ontology, one of BP, CC, MF}

\item{pvalueCutoff}{(default=1)}

\item{pAdjustMethod}{(default="BH")}

\item{minGSSize}{(default=5)}

\item{maxGSSize}{(default=500)}

\item{readable}{(default=FALSE)}

\item{eps}{(default=0)}

\item{exponent}{(default=1)}

\item{seed}{(default=FALSE)}

\item{verbose}{Print progress (default=FALSE)}

\item{...}{additional parameters parsed to DOSE::GSEA_internal}
}
\description{
Estimate enriched GSEA terms using GSEA_internal
}
\keyword{internal}
