% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{estimateCdaSpace}
\alias{estimateCdaSpace}
\title{Internal function for estimating CDA space}
\usage{
estimateCdaSpace(d.counts, d.groups, thresh.pc.var = 0.95, n.dim = 2)
}
\description{
Internal function for estimating CDA space
}
\keyword{internal}
