% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{estimateCellDensityKde}
\alias{estimateCellDensityKde}
\title{Estimate cell density in giving embedding, density will be estimated for individual samples}
\usage{
estimateCellDensityKde(
  emb,
  sample.per.cell,
  sample.weights,
  bins,
  bandwidth = 0.05,
  expansion.mult = 0.05
)
}
\arguments{
\item{emb}{cell embedding matrix}

\item{sample.per.cell}{Named sample factor with cell names (default: stored vector)}

\item{sample.weights}{Named numeric vector: per-sample signed weights.}

\item{bins}{number of bins for density estimation, default 400}
}
\description{
Estimate cell density in giving embedding, density will be estimated for individual samples
}
\keyword{internal}
