% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{estimateClusterFreeDE_LM}
\alias{estimateClusterFreeDE_LM}
\title{Estimate cluster-free differential expression using linear model and permutations}
\usage{
estimateClusterFreeDE_LM(
  genes,
  de.inp,
  sample.per.cell,
  design,
  perm.method = c("block", "freedman-lane"),
  robust.method = c("none", "huber", "winsor"),
  na.mode = c("drop", "impute_weak"),
  alternative = c("two-sided", "greater", "less"),
  n.permutations = 1000,
  max.z = 20,
  min.n.obs.per.samp = 2,
  min.n.samp.per.cond = 2,
  lfc.pseudocount = 1e-05,
  keep.means = FALSE,
  return.residuals = FALSE,
  verbose = TRUE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{genes}{vector of gene names to analyze}

\item{de.inp}{list with elements:
- cm: dgCMatrix (cells x genes) count matrix
- nns.per.cell: list of integer neighbor indices per cell}

\item{sample.per.cell}{vector/factor of sample IDs for each cell (length = nrow(cm))}

\item{design}{output of buildDesignPairMatrices (must include $X, $F, $contrast.X, $contrast.F)}

\item{perm.method}{"freedman-lane" or "block"}

\item{robust.method}{"none","huber","winsor"}

\item{na.mode}{"drop" or "impute_weak"}

\item{alternative}{"two-sided","greater","less"}

\item{n.permutations}{number of permutations to perform}

\item{max.z}{maximum absolute z-score to store}

\item{min.n.obs.per.samp}{minimum number of observed cells per sample to include in response}

\item{min.n.samp.per.cond}{minimum number of samples per condition to include response column}

\item{lfc.pseudocount}{pseudocount to add when computing log-fold changes}

\item{keep.means}{logical; if TRUE, compute and return per-gene means per condition}

\item{return.residuals}{logical; if TRUE, return residuals per gene}

\item{verbose}{logical}

\item{n.cores}{number of cores for parallel processing}
}
\description{
Estimate cluster-free differential expression using linear model and permutations
}
\keyword{internal}
