% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{estimateClusterFreeExpressionShiftsLM}
\alias{estimateClusterFreeExpressionShiftsLM}
\title{Estimate cluster-free shifts using linear model and permutations}
\usage{
estimateClusterFreeExpressionShiftsLM(
  cm,
  sample.per.cell,
  nns.per.cell,
  x,
  dist = "cor",
  log.vecs = TRUE,
  min.n.obs.per.samp = 2L,
  n.cores = 1,
  perm.method = c("freedman-lane", "block"),
  robust.method = c("none", "huber", "winsor"),
  n.permutations = 999,
  adjust = TRUE,
  na.mode = c("drop", "impute_weak"),
  alternative = c("two-sided", "greater", "less"),
  smooth = TRUE,
  wins = 0.025,
  return.sampled.stats = FALSE,
  return.residuals = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cm}{dgCMatrix (genes x cells)}

\item{sample.per.cell}{vector/factor of sample IDs for each cell (length = ncol(cm))}

\item{nns.per.cell}{list of integer neighbor indices per cell (same length as ncol(cm))}

\item{x}{output of buildDesignPairMatrices (must include $X, $Z, $contrast, $pairs)}

\item{dist, log.vecs, }{min.n.obs.per.samp as before}

\item{perm.method}{"freedman-lane" or "block"}

\item{robust.method}{"none","huber","winsor"}

\item{na.mode}{"drop" or "impute_weak"}

\item{alternative}{"two-sided","greater","less"}

\item{smooth}{logical; if TRUE, median-smooth effect sizes (and optionally z)}

\item{wins}{winsorization fraction for permutation adjustment (e.g., 0.01)}

\item{return.sampled.stats}{set TRUE if you want permutation-adjusted z's}

\item{return.residuals}{pass-through to cpp}

\item{verbose}{logical}
}
\description{
Estimate cluster-free shifts using linear model and permutations
}
\keyword{internal}
