% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{estimateClusterFreeZScores}
\alias{estimateClusterFreeZScores}
\title{Estimate Local Z Scores}
\usage{
estimateClusterFreeZScores(
  graph,
  count.matrix.transposed,
  is.ref,
  genes = NULL,
  max.z = 20,
  ...
)
}
\arguments{
\item{graph}{Alignment graph adjacency matrix}

\item{count.matrix.transposed}{Joint count matrix with cells by rows and genes by columns}

\item{is.ref}{Boolean vector per cell indicating whether it came from the reference condition}

\item{genes}{Genes to be tested. Procedure is slow, so it's reasonable to restrict amount of genes.}

\item{ref.level}{Reference condition level, e.g., wt, ctrl, or healthy}
}
\description{
Estimate Local Z Scores
}
