% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateClusterPerGO}
\alias{estimateClusterPerGO}
\title{Estimate ontology clusters}
\usage{
estimateClusterPerGO(genes.per.go, cut.h)
}
\arguments{
\item{genes.per.go}{Genes per term}

\item{cut.h}{numeric scalar or vector with heights where the tree should be cut.}
}
\description{
Estimate ontology clusters
}
\keyword{internal}
