% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeDESeq}
\alias{estimateDEForTypeDESeq}
\title{Estimate DE using DESeq2}
\usage{
estimateDEForTypeDESeq(
  cm,
  sample.meta,
  model,
  test.type = "wald",
  cooksCutoff = FALSE,
  independentFiltering = TRUE,
  ...
)
}
\arguments{
\item{cm}{count matrix}

\item{sample.meta}{sample metadata data frame}

\item{model}{list returned by buildDesignMatrices (must contain F and contrast.F)}

\item{test.type}{test type incorporated in DESeq2, either "wald" or "LRT"}

\item{...}{additional parameters forwarded to DESeq2}
}
\description{
Estimate DE using DESeq2
}
\keyword{internal}
