% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeEdgeR}
\alias{estimateDEForTypeEdgeR}
\title{Estimate DE using edgeR (design + contrast from buildDesignMatrices)}
\usage{
estimateDEForTypeEdgeR(cm, model)
}
\arguments{
\item{cm}{count matrix (genes x samples)}

\item{model}{list returned by buildDesignMatrices (must contain F and contrast.F)}
}
\description{
Estimate DE using edgeR (design + contrast from buildDesignMatrices)
}
\keyword{internal}
