% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeLimma}
\alias{estimateDEForTypeLimma}
\title{Estimate DE using limma-voom (design + contrast from buildDesignMatrices)}
\usage{
estimateDEForTypeLimma(cm, model)
}
\arguments{
\item{cm}{count matrix (genes x samples)}

\item{model}{list returned by buildDesignMatrices (must contain F and contrast.F)}
}
\description{
Estimate DE using limma-voom (design + contrast from buildDesignMatrices)
}
\keyword{internal}
