% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEPerCellTypeInner}
\alias{estimateDEPerCellTypeInner}
\title{Differential expression using different methods (DESeq2, edgeR, wilcoxon, ttest) with various covariates}
\usage{
estimateDEPerCellTypeInner(
  raw.mats,
  cell.groups = NULL,
  s.groups = NULL,
  ref.level = NULL,
  target.level = NULL,
  common.genes = FALSE,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  max.cell.count = Inf,
  independent.filtering = TRUE,
  n.cores = 4,
  return.matrix = TRUE,
  fix.n.samples = NULL,
  verbose = TRUE,
  test = "Wald",
  meta.info = NULL,
  gene.filter = NULL
)
}
\arguments{
\item{raw.mats}{list of counts matrices; column for gene and row for cell}

\item{cell.groups}{factor specifying cell types (default=NULL)}

\item{s.groups}{list of two character vector specifying the app groups to compare (default=NULL)}

\item{ref.level}{reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{target.level}{target level in 'sample.groups' (default=NULL)}

\item{common.genes}{boolean Only investigate common genes across cell groups (default=FALSE)}

\item{cooks.cutoff}{boolean cooksCutoff for DESeq2 (default=FALSE)}

\item{min.cell.count}{numeric Minimum cell count (default=10)}

\item{max.cell.count}{numeric Maximum cell count (default=Inf). If Inf, there is no limit set.}

\item{independent.filtering}{boolean independentFiltering for DESeq2 (default=FALSE)}

\item{n.cores}{numeric Number of cores (default=1)}

\item{return.matrix}{Return merged matrix of results (default=TRUE)}

\item{fix.n.samples}{Number of samples to fix (default=NULL). If greater the the length of the s.groups, an error is thrown.}

\item{verbose}{boolean Whether to output verbose messages (default=TRUE)}

\item{test}{DE method: DESeq2, edgeR, wilcoxon, ttest}

\item{meta.info}{dataframe with possible covariates; for example, sex or age}

\item{gene.filter}{matrix/boolean Genes to omit (rows) per cluster (cols) (default=NULL)}
}
\value{
differential expression for each cell type
}
\description{
Differential expression using different methods (DESeq2, edgeR, wilcoxon, ttest) with various covariates
}
