% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateEnrichedGO}
\alias{estimateEnrichedGO}
\title{Estimate enriched GOs}
\usage{
estimateEnrichedGO(de.gene.ids, go.environment, ...)
}
\arguments{
\item{de.gene.ids}{vector containing gene IDs}

\item{go.environment}{GO environment, see DOSE:::enricher_internal, parameter USER_DATA}

\item{...}{additional parameters parsed to enrichGOOpt}
}
\description{
Estimate enriched GOs
}
\keyword{internal}
