% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateEnrichedGSEGO}
\alias{estimateEnrichedGSEGO}
\title{Estimate enriched GSEA ontologies}
\usage{
estimateEnrichedGSEGO(go.environment, ...)
}
\arguments{
\item{go.environment}{List with GO environment}

\item{...}{additional parameters passed to enrichGSEGOOpt}
}
\value{

}
\description{
Estimate enriched GSEA ontologies
}
\keyword{internal}
