% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R, R/expression_shifts.R
\name{estimateExpressionShiftMagnitudes.Conos}
\alias{estimateExpressionShiftMagnitudes.Conos}
\alias{estimateExpressionShiftMagnitudes.default}
\title{Estimate Expression Shift Magnitudes}
\usage{
\method{estimateExpressionShiftMagnitudes}{Conos}(con, groups = NULL, ...)

\method{estimateExpressionShiftMagnitudes}{default}(
  count.matrices,
  sample.groups,
  groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  n.cores = 1,
  verbose = FALSE,
  transposed.matrices = FALSE
)
}
\arguments{
\item{con}{conos object}

\item{groups}{cell cluster factor}

\item{...}{
  Arguments passed on to \code{\link[=estimateExpressionShiftMagnitudes.default]{estimateExpressionShiftMagnitudes.default}}
  \describe{
    \item{\code{}}{}
  }}

\item{count.matrices}{raw count matrices for each of the samples}

\item{sample.groups}{a two-level factor on the sample names describing the conditions being compared}

\item{dist}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance}

\item{within.group.normalization}{normalize the shift magnitude by the mean magnitude of within-group variation}

\item{valid.comparisons}{a logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons}

\item{n.cells}{number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{n.top.genes}{number of top highest-expressed genes to consider (default: all genes)}

\item{n.subsamples}{number of samples to draw (default:100)}

\item{min.cells}{minimum number of cells per cluster/per sample to be included in the analysis}

\item{n.cores}{number of cores to use}

\item{verbose}{}

\item{transposed.matrices}{are count matrices transposed (i.e. rows are cells and genes are cols)?}
}
\value{
a list include 1. df - a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells; 2. ctdml - a list of cluster distance matrices; 3. sample.groups; 4. valid.comparisons
}
\description{
Calculate expression shift magnitudes of different clusters between conditions
}
