% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_shifts.R
\name{estimateExpressionShiftZScores}
\alias{estimateExpressionShiftZScores}
\title{Expression shift Z scores per cluster between conditions}
\usage{
estimateExpressionShiftZScores(
  pca,
  sample.per.cell,
  sample.groups,
  cell.groups,
  ref.level
)
}
\arguments{
\item{pca}{Principal components}

\item{sample.per.cell}{Named factor with cell names indicating sample origin per cell}

\item{sample.groups}{Named factor with cell names indicating condition/sample, e.g., ctrl/disease}

\item{cell.groups}{Named factor with cell names and clustering/annotation}

\item{ref.level}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt}
}
\description{
Expression shift Z scores per cluster between conditions
}
