% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{estimateGeneProgramsFabia}
\alias{estimateGeneProgramsFabia}
\title{Estimate Gene Programs with FABIA}
\usage{
estimateGeneProgramsFabia(
  z.scores,
  n.programs,
  n.sampled.cells = 15000,
  cyc = 1500,
  alpha = 0.2,
  random = -1,
  ...
)
}
\arguments{
\item{n.programs}{maximal number of gene programs to find (parameter \code{p} for fabia).}

\item{n.sampled.cells}{number of sub-sampled cells for estimating the gene programs. If 0, all cells are used.
it is interpreted as a vector of cell}

\item{...}{further arguments passed to \code{fabia::fabia()} (other than \code{p}, \code{X}, \code{cyc}, \code{alpha}, \code{random})}
}
\description{
Estimate Gene Programs with FABIA
}
\keyword{internal}
