% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthology.R
\name{estimateOnthology}
\alias{estimateOnthology}
\title{Estimate onthology}
\usage{
estimateOnthology(
  type,
  org = "human",
  de.gene.ids,
  go.environment = NULL,
  p.adj = 0.05,
  p.adjust.method = "BH",
  readable = T,
  verbose = T,
  n.cores = 1,
  ...
)
}
\arguments{
\item{type}{Onthology type, either GO (gene onthology) or DO (disease onthology). Please see DOSE package for more information.}

\item{org}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{p.adj}{Adjusted P cutoff (default=0.05)}

\item{p.adjust.method}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{readable}{Mapping gene ID to gene name (default=T)}

\item{verbose}{Print progress (default=T)}

\item{n.cores}{Number of cores used (default: stored vector)}

\item{...}{Additional parameters for sccore:::plapply function}

\item{ont.data}{List containing DE gene IDs, and filtered DE genes}
}
\value{
A list containing a list of onthologies per type of onthology, and a data frame with merged results
}
\description{
Calculate onthologies based on DEs
}
