% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntologyClusterName}
\alias{estimateOntologyClusterName}
\title{Estimate names for ontology clusters}
\usage{
estimateOntologyClusterName(
  descriptions,
  method = c("medoid", "consensus"),
  n.words = 5,
  exclude.words = NULL
)
}
\arguments{
\item{descriptions}{Descriptions}

\item{method}{One of "medoid" or "consensus"}

\item{n.words}{(default=5)}

\item{exclude.words}{(default=NULL)}
}
\description{
Estimate names for ontology clusters
}
\keyword{internal}
