% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntologyClusterNames}
\alias{estimateOntologyClusterNames}
\title{Estimate ontology cluster names}
\usage{
estimateOntologyClusterNames(
  ont.df,
  clust.naming = c("medoid", "consensus", "min.pvalue")
)
}
\arguments{
\item{ont.df}{Ontology dataframe}

\item{clust.naming}{One of "medoid", "consensus", or "min.pvalue"}
}
\value{

}
\description{
Estimate ontology cluster names
}
\keyword{internal}
