% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntologyFamilies}
\alias{estimateOntologyFamilies}
\title{Estimate ontology families}
\usage{
estimateOntologyFamilies(ont.list, type)
}
\arguments{
\item{ont.list}{List of results from estimateOntology}

\item{type}{Type of ontology result, i.e., GO, GSEA, or DO}
}
\value{
List of families and ontology data per cell type
}
\description{
Estimate ontology families based on ontology results
}
