% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntologyFromIds}
\alias{estimateOntologyFromIds}
\title{Calculate ontologies based on DEs}
\usage{
estimateOntologyFromIds(
  de.gene.scores,
  go.environment,
  type = "GO",
  org.db = NULL,
  n.top.genes = 500,
  keep.gene.sets = FALSE,
  verbose = TRUE,
  n.cores = 1,
  ...
)
}
\arguments{
\item{de.gene.scores}{input DE gene scores}

\item{go.environment}{GO environment set (Homo sapiens, Mus musculus, etc.)}

\item{type}{character string Ontology type, either GO (gene ontology) or DO (disease ontology) (default="GO")
Please see DOSE package for more information.}

\item{org.db}{organism db, e.g. org.Hs.eg.db::org.Hs.eg.db}

\item{n.top.genes}{numeric Number of most different genes to take as input. If less are left after filtering
for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}

\item{keep.gene.sets}{boolean Keep gene sets (default=FALSE)}

\item{verbose}{boolean Print progress (default=TRUE)}

\item{n.cores}{integer Number of cores (default=1)}

\item{...}{Additional parameters for DO/GO/GSEA functions. In case of GSEA, pass nPerm to call fgseaSimple
instead of fgseaMultilevel}
}
\value{
A list containing a list of ontologies per type of ontology, and a data frame with merged results
}
\description{
Calculate ontologies based on DEs
}
