% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimatePerCellTypeDE}
\alias{estimatePerCellTypeDE}
\title{Do differential expression for each cell type in a conos object between the specified subsets of apps}
\usage{
estimatePerCellTypeDE(
  raw.mats,
  cell.groups = NULL,
  sample.groups = NULL,
  ref.level = NULL,
  common.genes = FALSE,
  test = "LRT",
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  max.cell.count = Inf,
  independent.filtering = TRUE,
  n.cores = 1,
  cluster.sep.chr = "<!!>",
  return.matrix = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{raw.mats}{list of counts matrices; column for gene and row for cell}

\item{cell.groups}{factor specifying cell types (default=NULL)}

\item{sample.groups}{a list of two character vector specifying the app groups to compare (default=NULL)}

\item{ref.level}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{common.genes}{Only investigate common genes across cell groups (default=FALSE)}

\item{test}{which DESeq2 test to use (options: "LRT" (default), "Wald")}

\item{cooks.cutoff}{cooksCutoff for DESeq2 (default=FALSE)}

\item{min.cell.count}{(default=10)}

\item{max.cell.count}{maximal number of cells per cluster per sample to include in a comparison (useful for comparing the number of DE genes between cell types)}

\item{independent.filtering}{independentFiltering for DESeq2 (default=FALSE)}

\item{n.cores}{Number of cores (default=1)}

\item{cluster.sep.chr}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{return.matrix}{Return merged matrix of results (default=TRUE)}
}
\description{
Do differential expression for each cell type in a conos object between the specified subsets of apps
}
