% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimatePerCellTypeDEmeth}
\alias{estimatePerCellTypeDEmeth}
\title{Differential expression using DESeq2 or edgeR with various covariates}
\usage{
estimatePerCellTypeDEmeth(
  raw.mats,
  cell.groups = NULL,
  sample.groups = NULL,
  ref.level = NULL,
  common.genes = F,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  independent.filtering = T,
  n.cores = 1,
  cluster.sep.chr = "<!!>",
  return.matrix = T,
  verbose = T,
  meth = NULL,
  covariates = c(),
  meta.info = NULL,
  useT = F,
  minmu = 0.5,
  minReplicatesForReplace = 7,
  test.name = "Wald"
)
}
\arguments{
\item{raw.mats}{list of counts matrices; column for gene and row for cell}

\item{cell.groups}{factor specifying cell types (default=NULL)}

\item{sample.groups}{a list of two character vector specifying the app groups to compare (default=NULL)}

\item{ref.level}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{common.genes}{Only investigate common genes across cell groups (default=F)}

\item{cooks.cutoff}{cooksCutoff for DESeq2 (default=F)}

\item{min.cell.count}{(default=10)}

\item{independent.filtering}{independentFiltering for DESeq2 (default=F)}

\item{n.cores}{Number of cores (default=1)}

\item{cluster.sep.chr}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{return.matrix}{Return merged matrix of results (default=T)}

\item{covariates}{list of covariates to include; for example, cdr, sex or age}

\item{meta.info}{dataframe with possible covariates; for example, sex or age}
}
\description{
Differential expression using DESeq2 or edgeR with various covariates
}
