% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractCellGraph}
\alias{extractCellGraph}
\alias{extractCellGraph.Conos}
\alias{extractCellGraph.Seurat}
\title{Extract the cell groups from the object}
\usage{
extractCellGraph(object)

\method{extractCellGraph}{Conos}(object)

\method{extractCellGraph}{Seurat}(object)
}
\arguments{
\item{object}{object from which to extract the cell groups}
}
\description{
Extract the cell groups from the object
}
