% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractCellGroups}
\alias{extractCellGroups}
\alias{extractCellGroups.Conos}
\alias{extractCellGroups.Seurat}
\title{Extract the cell groups from the object}
\usage{
extractCellGroups(object)

\method{extractCellGroups}{Conos}(object)

\method{extractCellGroups}{Seurat}(object)
}
\arguments{
\item{object}{object from which to extract the cell groups}
}
\description{
Extract the cell groups from the object
}
