% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractEmbedding}
\alias{extractEmbedding}
\alias{extractEmbedding.Conos}
\alias{extractEmbedding.Seurat}
\title{Extract embeddings from the object}
\usage{
extractEmbedding(object)

\method{extractEmbedding}{Conos}(object)

\method{extractEmbedding}{Seurat}(object)
}
\arguments{
\item{object}{object from which to extract the embeddings}
}
\description{
Extract embeddings from the object
}
