% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractGeneExpression}
\alias{extractGeneExpression}
\alias{extractGeneExpression.Conos}
\alias{extractGeneExpression.Seurat}
\alias{extractGeneExpression.dgCMatrix}
\title{Extract the gene exrpession from the object}
\usage{
extractGeneExpression(object, gene)

\method{extractGeneExpression}{Conos}(object, gene)

\method{extractGeneExpression}{Seurat}(object, gene)

\method{extractGeneExpression}{dgCMatrix}(object, gene)
}
\arguments{
\item{object}{object from which to extract the cell groups}

\item{gene}{character vector of the specific gene names on which to subset}
}
\description{
Extract the gene exrpession from the object
}
