% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractJointCountMatrix}
\alias{extractJointCountMatrix}
\alias{extractJointCountMatrix.Conos}
\alias{extractJointCountMatrix.Seurat}
\alias{extractJointCountMatrix.dgCMatrix}
\title{Extract the joint count matrix from the object}
\usage{
extractJointCountMatrix(object, raw = TRUE, ...)

\method{extractJointCountMatrix}{Conos}(object, raw = TRUE)

\method{extractJointCountMatrix}{Seurat}(object, raw = TRUE, transposed = TRUE, sparse = TRUE)

\method{extractJointCountMatrix}{dgCMatrix}(object, raw = TRUE)
}
\arguments{
\item{object}{object from which to extract the cell groups}

\item{raw}{boolean If TRUE, return merged "raw" count matrices (default=TRUE)}

\item{...}{additional parameters to be passed to extractJointCountMatrix}

\item{transposed}{boolean If TRUE, return merged transposed count matrices (default=TRUE)}

\item{sparse}{boolean If TRUE, return merged the sparse dgCMatrix matrix (default=TRUE)}
}
\description{
Extract the joint count matrix from the object
}
