% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractOdGenes}
\alias{extractOdGenes}
\alias{extractOdGenes.Conos}
\alias{extractOdGenes.Seurat}
\title{Extract the top overdispersed genes from the object}
\usage{
extractOdGenes(object, n.genes = NULL)

\method{extractOdGenes}{Conos}(object, n.genes = NULL)

\method{extractOdGenes}{Seurat}(object, n.genes = NULL)
}
\arguments{
\item{object}{object from which to extract the top overdispersed genes}

\item{n.genes}{numeric Number of overdispersed genes to extract (default=NULL)}
}
\description{
Extract the top overdispersed genes from the object
}
