% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractRawCountMatrices}
\alias{extractRawCountMatrices}
\alias{extractRawCountMatrices.Conos}
\alias{extractRawCountMatrices.Seurat}
\alias{extractRawCountMatrices.dgCMatrix}
\title{Extract the cell groups from the object}
\usage{
extractRawCountMatrices(object, transposed = TRUE)

\method{extractRawCountMatrices}{Conos}(object, transposed = TRUE)

\method{extractRawCountMatrices}{Seurat}(object, transposed = TRUE)

\method{extractRawCountMatrices}{dgCMatrix}(object, transposed = TRUE)
}
\arguments{
\item{object}{object from which to extract the cell groups}

\item{transposed}{boolean Whether the raw count matrix should be transposed (default=TRUE)}
}
\description{
Extract the cell groups from the object
}
