% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wrappers.R
\name{extractSamplePerCell}
\alias{extractSamplePerCell}
\alias{extractSamplePerCell.Conos}
\alias{extractSamplePerCell.Seurat}
\title{Extract the sample/dataset per cell from the object}
\usage{
extractSamplePerCell(object)

\method{extractSamplePerCell}{Conos}(object)

\method{extractSamplePerCell}{Seurat}(object)
}
\arguments{
\item{object}{object from which to extract the sample/dataset per cell}
}
\description{
Extract the sample/dataset per cell from the object
}
