% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{extractSimpleTripletFromSpec}
\alias{extractSimpleTripletFromSpec}
\title{Extract simple triplet contrast from model contrast specification}
\usage{
extractSimpleTripletFromSpec(model)
}
\arguments{
\item{model}{model object returned by buildDesignMatrices}
}
\value{
character vector of length 3 with c("group", "num", "den")
}
\description{
Extract simple triplet contrast from model contrast specification
}
\keyword{internal}
