% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{filterDEMetadata}
\alias{filterDEMetadata}
\title{Filter DE metadata}
\usage{
filterDEMetadata(meta)
}
\arguments{
\item{meta}{data frame containing metadata in columns}
}
\value{
cleaned data frame with metadata
}
\description{
Filter DE metadata
}
\keyword{internal}
