% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free_de.R
\name{filterLocalZScores}
\alias{filterLocalZScores}
\title{Filter Local Z Scores}
\usage{
filterLocalZScores(
  z.scores,
  min.row.z = 1,
  min.individual.z = 0.25,
  n.cores = 1
)
}
\arguments{
\item{z.scores}{Output from localZScores function}

\item{min.row.z}{(default=1)}

\item{n.cores}{Cores for parallel processing (default=1)}

\item{min.inidividual}{z (default=0.25)}
}
\description{
Filter Local Z Scores
}
