% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filterOntologies}
\alias{filterOntologies}
\title{Filter ontologies by adj. P}
\usage{
filterOntologies(ont.list, p.adj)
}
\arguments{
\item{ont.list}{List of ontology terms}

\item{p.adj}{Adj. P cut-off}
}
\value{

}
\description{
Filter ontologies by adj. P
}
\keyword{internal}
