% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{filterOntologyDf}
\alias{filterOntologyDf}
\title{Filter ontology dataframe}
\usage{
filterOntologyDf(
  ont.df,
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  subtype = NULL,
  cell.subgroups = NULL
)
}
\arguments{
\item{ont.df}{Ontology dataframe}

\item{p.adj}{Adj. P cut-off (default=0.05)}

\item{q.value}{(default=0.2)}

\item{min.genes}{Min. number of significant terms per term (default=1)}

\item{subtype}{(default=NULL)}

\item{cell.subgroups}{(default=NULL)}
}
\description{
Filter ontology dataframe
}
\keyword{internal}
