% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_partitions.R
\name{gaPartitionAll}
\alias{gaPartitionAll}
\title{Genetic algorithm to obtaint the most contrast balance including all cell types}
\usage{
gaPartitionAll(d.counts, d.groups, n.pop = 1000, n.epoch = 100, n.seed = 239)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Group variable}

\item{n.pop}{number of individuals in the population}

\item{n.epoch}{number of epochs}

\item{n.seed}{seed id}
}
\value{
matrix with final population, rows - individuals, top individual - the best partition
}
\description{
Genetic algorithm to obtaint the most contrast balance including all cell types
}
\keyword{internal}
