% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_partitions.R
\name{getCanonicalContrasts}
\alias{getCanonicalContrasts}
\title{Get caconical decomposisiotn of balances based}
\usage{
getCanonicalContrasts(d.counts, d.groups, n.dim = 2)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Group variable}

\item{n.dim}{number of components to return}
}
\value{
Caconical components
}
\description{
Get caconical decomposisiotn of balances based
}
\keyword{internal}
