% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{getCdaLoadings}
\alias{getCdaLoadings}
\title{Get Loadings from the caconical data analysis}
\usage{
getCdaLoadings(d.counts, d.groups, n.seed = 239)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Group variable}

\item{n.seed}{Random seed}
}
\value{
Vector with cda loadings
}
\description{
Get Loadings from the caconical data analysis
}
\keyword{internal}
