% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{getCodaContrastScore}
\alias{getCodaContrastScore}
\title{Compute per-sample contrast score from lmCoda result}
\usage{
getCodaContrastScore(x)
}
\description{
Compute per-sample contrast score from lmCoda result
}
\keyword{internal}
