% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getCovariatesForTriplet}
\alias{getCovariatesForTriplet}
\title{Get covariates from model formula for triplet tests}
\usage{
getCovariatesForTriplet(model, group.var)
}
\arguments{
\item{model}{model object returned by buildDesignMatrices}

\item{group.var}{grouping variable in}
}
\value{
character vector with covariate names
}
\description{
Get covariates from model formula for triplet tests
}
\keyword{internal}
