% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{getDEEntrezIds}
\alias{getDEEntrezIds}
\title{Get DE ENTREZ IDs}
\usage{
getDEEntrezIds(de.raw, org.db, p.adj = 1, expr.cutoff = 0.05)
}
\arguments{
\item{de.raw}{List with differentially expressed genes per cell group}

\item{org.db}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{expr.cutoff}{Cutoff for cells per group expressing a DE gene, i.e., cutoff for highly-expressed genes (default=0.05)}

\item{p.adj.cutoff}{Adj. P cutoff for filtering DE genes (default=0.05)}
}
\value{
A list containing DE ENSEMBL gene IDs, and filtered DE genes
}
\description{
Filter and prepare DE genes for ontology calculations
}
