% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{getDensityContour}
\alias{getDensityContour}
\title{Extract contour from embedding}
\usage{
getDensityContour(
  emb,
  cell.groups,
  group,
  color = "black",
  linetype = 2,
  conf = "10\%",
  bandwidth = NULL,
  ...
)
}
\arguments{
\item{emb}{cell embedding matrix}

\item{cell.groups}{vector of cell type annotation}

\item{group}{specify cell types for contour, multiple cell types are also supported}

\item{conf}{confidence interval of contour}
}
\description{
Extract contour from embedding
}
\keyword{internal}
