% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getExpressionFractionPerGroup}
\alias{getExpressionFractionPerGroup}
\title{get expression fraction per cell group}
\usage{
getExpressionFractionPerGroup(cm, cell.groups)
}
\arguments{
\item{cm}{sparse count matrix}

\item{cell.groups}{factor containing cell groups with cell names as names}
}
\description{
get expression fraction per cell group
}
\keyword{internal}
