% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getGeneScale}
\alias{getGeneScale}
\title{Get Gene Scale}
\usage{
getGeneScale(
  genes = c("up", "down", "all"),
  type = c("fill", "color"),
  high = "gray80",
  ...
)
}
\arguments{
\item{genes}{type of genes ("up", "down" or "all")}

\item{type}{type of scale ("fill" or "color")}

\item{high}{color for the highest value}
}
\value{
ggplot2 fill or color scale
}
\description{
Get Gene Scale
}
