% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R, R/coda_balances.R
\name{getLogFreq}
\alias{getLogFreq}
\title{Get log frequencies from the cell count table}
\usage{
getLogFreq(d.counts)

getLogFreq(d.counts)
}
\description{
Get log frequencies from the cell count table

Get log frequencies from the cell count table
}
\keyword{internal}
