% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_trees.R
\name{getNodeOrder}
\alias{getNodeOrder}
\title{Get inner nodes on the tree from leaves to the root}
\usage{
getNodeOrder(t)
}
\arguments{
\item{t}{A phylo object}
}
\value{
Order of inner nodes
}
\description{
Get inner nodes on the tree from leaves to the root
}
\keyword{internal}
