% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthology.R
\name{getOnthologySummary}
\alias{getOnthologySummary}
\title{Get onthology summary}
\usage{
getOnthologySummary(type = NULL, ont.res)
}
\arguments{
\item{type}{Onthology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default=NULL)}

\item{pathway.data}{Results from preparePathwayData (default: stored list)}
}
\value{
Data frame
}
\description{
Get summary
}
