% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{getOntologyFamilyChildren}
\alias{getOntologyFamilyChildren}
\title{Get ontology family children}
\usage{
getOntologyFamilyChildren(ont.sum, fams, subtype, genes)
}
\arguments{
\item{ont.sum}{Dataframe containing summarized ontology terms}

\item{fams}{Ontology families}

\item{subtype}{Type of ontology, one of "BP", "CC", "MF"}

\item{genes}{Included genes}
}
\description{
Get ontology family children
}
\keyword{internal}
