% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{getOntologyListLevels}
\alias{getOntologyListLevels}
\title{Create list with ontology terms}
\usage{
getOntologyListLevels(type = c("GO", "GSEA", "DO"))
}
\arguments{
\item{type}{one of GO, GSEA or DO}
}
\description{
Create list with ontology terms
}
\keyword{internal}
