% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{getOntologySummary}
\alias{getOntologySummary}
\title{Get ontology summary}
\usage{
getOntologySummary(ont.res)
}
\arguments{
\item{ont.res}{Results from prepareOntologyData (default: stored list)}
}
\value{
Data frame
}
\description{
Get summary
}
