% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getPerCellTypeDEmat}
\alias{getPerCellTypeDEmat}
\title{Do differential expression for each cell type in a conos object between the specified subsets of apps}
\usage{
getPerCellTypeDEmat(
  raw.mats,
  cell.groups = NULL,
  sample.groups = NULL,
  cooks.cutoff = FALSE,
  ref.level = NULL,
  min.cell.count = 10,
  independent.filtering = FALSE,
  n.cores = 1,
  cluster.sep.chr = "<!!>",
  return.details = TRUE,
  verbose = T
)
}
\arguments{
\item{raw.mats}{list of counts matrix; row for gene and column for cell}

\item{cell.groups}{factor specifying cell types (default=NULL)}

\item{sample.groups}{a list of two character vector specifying the app groups to compare (default=NULL)}

\item{cooks.cutoff}{cooksCutoff for DESeq2 (default=F)}

\item{ref.level}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{min.cell.count}{(default=10)}

\item{independent.filtering}{independentFiltering for DESeq2 (default=F)}

\item{n.cores}{Number of cores (default=1)}

\item{cluster.sep.chr}{character string of length 1 specifying a delimiter to separate cluster and app names}

\item{return.details}{Return details}
}
\description{
Do differential expression for each cell type in a conos object between the specified subsets of apps
}
