% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getPerCellTypeGeneFilter}
\alias{getPerCellTypeGeneFilter}
\title{filter genes by expression per celltype not globally}
\usage{
getPerCellTypeGeneFilter(counts, cell.groups, threshold = 0.05)
}
\arguments{
\item{counts}{sparse count matrix}

\item{cell.groups}{factor containing cell groups with cell names as names}

\item{threshold}{minimum fraction of expressing cells per cell type (default=0.05)}
}
\description{
filter genes by expression per celltype not globally
}
\keyword{internal}
