% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getResDf}
\alias{getResDf}
\title{helper: get the DE table (data.frame) from either a df or a list(res=...)}
\usage{
getResDf(x)
}
\description{
helper: get the DE table (data.frame) from either a df or a list(res=...)
}
\keyword{internal}
