% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{getRndBalances}
\alias{getRndBalances}
\title{Get balances based on the random ilr basis}
\usage{
getRndBalances(d.counts, n.seed = 239)
}
\arguments{
\item{d.counts}{Cell count table}

\item{n.seed}{Random seed}
}
\value{
Matrix with balances in columns
}
\description{
Get balances based on the random ilr basis
}
\keyword{internal}
