% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{getRndSbp}
\alias{getRndSbp}
\title{Generate random sbp}
\usage{
getRndSbp(n.cell.types, n.seed = 239)
}
\arguments{
\item{n.cell.types}{number of cell types}
}
\value{
list with binary and normalized sbp matrices (partitions are in rows)
}
\description{
Generate random sbp
}
\keyword{internal}
