% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_partitions.R
\name{getSignifCellTypes}
\alias{getSignifCellTypes}
\title{Get the most contrast balance based on genetic algorithm}
\usage{
getSignifCellTypes(
  d.counts,
  d.groups,
  n.pop = 1000,
  n.epoch = 100,
  n.seed = 239,
  thresh.parent = 1/4
)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Group variable}

\item{n.pop}{number of individuals in the population}

\item{n.epoch}{number of epochs}

\item{n.seed}{seed id}

\item{thresh.parent}{fraction of parents for a next epoch}
}
\value{
list of significant cell types and two contrast sets of cells
}
\description{
Get the most contrast balance based on genetic algorithm
}
\keyword{internal}
