% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_partitions.R
\name{getSignifCellTypesIteratively}
\alias{getSignifCellTypesIteratively}
\title{Get the most contrast balance based on genetic algorithm iteratively(!)}
\usage{
getSignifCellTypesIteratively(
  d.counts,
  d.groups,
  n.pop = 1000,
  n.epoch = 100,
  n.seed = 239,
  thresh.parent = 1/4,
  p.val.thresh = 0.01
)
}
\arguments{
\item{d.counts}{Cell count table}

\item{d.groups}{Group variable}

\item{n.pop}{number of individuals in the population}

\item{n.epoch}{number of epochs}

\item{n.seed}{seed id}

\item{thresh.parent}{fraction of parents for a next epoch}

\item{p.val.thresh}{p-value threshold}
}
\value{
list of significant cell types and two contrast sets of cells
}
\description{
Get the most contrast balance based on genetic algorithm iteratively(!)
}
\keyword{internal}
