% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{getTopDEGenes}
\alias{getTopDEGenes}
\title{Get Top DE Genes}
\usage{
getTopDEGenes(
  z.scores,
  min.z = 0,
  max.z = 10,
  cell.subset = NULL,
  top.quantile = NULL
)
}
\arguments{
\item{z.scores}{Filtered z scores}

\item{min.z}{(default=0)}

\item{max.z}{(default=10)}

\item{cell.subset}{Cells to subset from z.scores (default=NULL)}

\item{top.quantile}{(default=NULL)}
}
\description{
Get Top DE Genes
}
