% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{groupOntologiesByCluster}
\alias{groupOntologiesByCluster}
\title{Group ontologies by clusters}
\usage{
groupOntologiesByCluster(
  ont.clust.df,
  field = "ClusterName",
  sign.field = "p.adjust"
)
}
\arguments{
\item{ont.clust.df}{Cluster dataframe per ontology}

\item{field}{(default="ClusterName")}

\item{sign.field}{(default="p.adjust")}
}
\description{
Group ontologies by clusters
}
\keyword{internal}
