% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{identifyFamilies}
\alias{identifyFamilies}
\title{Identify ontology families}
\usage{
identifyFamilies(ids)
}
\arguments{
\item{ids}{Data frame of cell type-specific ontology results from estimateOntology}
}
\value{
List of families and ontology data
}
\description{
Identify parents and/or children of significant ontology terms
}
\keyword{internal}
