% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{innodePosFromHclust}
\alias{innodePosFromHclust}
\title{Compute internal-node positions from hclust (matches ggdendro coords)}
\usage{
innodePosFromHclust(h)
}
\description{
Compute internal-node positions from hclust (matches ggdendro coords)
}
\keyword{internal}
