% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{mapGeneIds}
\alias{mapGeneIds}
\title{Map gene ENTREZ IDs to gene SYMBOLs}
\usage{
mapGeneIds(genes, org.db)
}
\arguments{
\item{genes}{vector containing gene ENTREZ IDs}

\item{org.db}{organism db, e.g. org.Hs.eg.db::org.Hs.eg.db}
}
\value{

}
\description{
Map gene ENTREZ IDs to gene SYMBOLs
}
\keyword{internal}
