% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel.preprocessed}
\alias{panel.preprocessed}
\title{Small preprocessed conos panel}
\format{
a list of four \code{Pagoda2} objects named by sample
}
\usage{
data(panel.preprocessed)
}
\description{
Four pagoda2-processed samples (two control, two disease) used in the examples and tests.
}
\keyword{datasets}
