% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{pickStatCol}
\alias{pickStatCol}
\title{helper: choose an available statistic column (with fallbacks)}
\usage{
pickStatCol(df, var.to.sort)
}
\description{
helper: choose an available statistic column (with fallbacks)
}
\keyword{internal}
