% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellNumbers}
\alias{plotCellNumbers}
\title{Plot proportions}
\usage{
plotCellNumbers(
  legend.position = "right",
  cell.groups,
  sample.per.cell,
  sample.groups
)
}
\arguments{
\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{sample.per.cell}{Vector indicating sample name with cell names (default: stored vector)}

\item{sample.groups}{Vector indicating sample groups with sample names (default: stored vector)}
}
\value{
A ggplot2 object
}
\description{
Plot the cell group proportions per sample
}
