% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellTypeSizeDep}
\alias{plotCellTypeSizeDep}
\title{show a scatter plot of cell-type values vs. number of cells per cell type}
\usage{
plotCellTypeSizeDep(
  df,
  cell.groups,
  palette = NULL,
  font.size = 4,
  ylab = "expression distance",
  yline = 1,
  show.regression = TRUE,
  show.whiskers = TRUE,
  plot.theme = theme_get()
)
}
\arguments{
\item{df}{a data frame with $value and $Type columns, just like plotMeanValuesPerCellType}

\item{cell.groups}{a cell groups vector for calculating number of cells per cell type}

\item{palette}{cell type palette}

\item{ylab}{y axis label}

\item{yline}{value at which a horizontal reference value should be plotted}

\item{show.whiskers}{whether se values should be plotted}
}
\value{
ggplot2 object
}
\description{
show a scatter plot of cell-type values vs. number of cells per cell type
}
