% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCodaGlobal}
\alias{plotCodaGlobal}
\title{Internal function to plot global statistics from CoDA analysis (NEW)}
\usage{
plotCodaGlobal(
  x,
  style = c("bar", "violin"),
  fdr.threshold = 0.05,
  drop.intercept = TRUE,
  use.contrast.title = FALSE
)
}
\description{
Internal function to plot global statistics from CoDA analysis (NEW)
}
\keyword{internal}
