% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCodaHeatmap}
\alias{plotCodaHeatmap}
\title{Internal function to plot heatmap of CoDA coefficients and contrast (NEW)}
\usage{
plotCodaHeatmap(x, coef.names = NULL, fdr.threshold = 0.05)
}
\description{
Internal function to plot heatmap of CoDA coefficients and contrast (NEW)
}
\keyword{internal}
