% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCodaSpaceInner}
\alias{plotCodaSpaceInner}
\title{Internal plotting function to plot CoDA space}
\usage{
plotCodaSpaceInner(
  df.space,
  df.loadings,
  d.groups,
  ref.level,
  target.level,
  font.size = 3,
  palette = NULL
)
}
\description{
Internal plotting function to plot CoDA space
}
\keyword{internal}
