% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_plots.R
\name{plotContrastTree}
\alias{plotContrastTree}
\title{helper function for creating dendograms}
\usage{
plotContrastTree(
  d.counts,
  d.groups,
  ref.level,
  target.level,
  plot.theme = theme_get(),
  label.angle = 90,
  p.threshold = 0.05,
  adjust.pvalues = TRUE,
  h.methods = "both",
  font.size = 3,
  label.hjust = 1,
  tree.order = NULL,
  loadings.mean = NULL,
  palette = NULL,
  verbose = FALSE
)
}
\description{
helper function for creating dendograms
}
\keyword{internal}
