% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCountBoxplotsPerType}
\alias{plotCountBoxplotsPerType}
\title{Plot Count Boxplots Per Type}
\usage{
plotCountBoxplotsPerType(
  count.df,
  y.lab = "count",
  x.lab = "",
  y.expand = c(0, 0),
  show.significance = FALSE,
  palette = NULL,
  notch = FALSE,
  legend.position = "right",
  alpha = 0.2,
  plot.theme = theme_get(),
  size = 0.5,
  jitter.width = 0.15,
  adjust.pvalues = T
)
}
\arguments{
\item{count.df}{data.frame with columns \code{group}, \code{variable} and \code{value}}

\item{notch}{Whether to show notch in the boxplots}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{alpha}{Transparency level on the data points (default: 0.2)}

\item{size}{marker size (default: 0.5)}

\item{jitter.width}{width of the point jitter (default: 0.15)}
}
\description{
Plot Count Boxplots Per Type
}
