% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDEGenes}
\alias{plotDEGenes}
\title{Plot DE genes}
\usage{
plotDEGenes(
  de.raw,
  de.genes.filtered,
  cell.groups,
  legend.position = "bottom",
  p.adjust.cutoff = 0.05,
  label.x.pos = 0.01,
  label.y.pos = 1,
  rel_heights = c(2.5, 0.5),
  scale = 0.93,
  show.legend = T
)
}
\arguments{
\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="bottom")}

\item{p.adjust.cutoff}{Adjusted P cutoff (default=0.05)}

\item{label.x.pos}{Plot label position on x axis (default=0.01)}

\item{label.y.pos}{Plot label position on y axis (default=1)}

\item{rel_heights}{Relative heights for plots. Only relevant if show.legend=T. See cowplot::plot_grid for more info (default=c(2.5, 0.5))}

\item{scale}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (defaul=0.93)}

\item{show.legend}{Include legend in plot (default=T)}
}
\value{
A ggplot2 object
}
\description{
Plot number of DE genes as a function of number of cells
}
