% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{plotDensityKde}
\alias{plotDensityKde}
\title{Plot cell density}
\usage{
plotDensityKde(
  mat,
  bins,
  cell.emb,
  show.grid = TRUE,
  lims = NULL,
  show.labels = FALSE,
  show.ticks = FALSE,
  palette = NULL,
  legend.title = NULL,
  ...
)
}
\arguments{
\item{bins}{number of bins for density estimation, should keep consistent with bins in estimateCellDensity}

\item{palette}{color palette function. Default: \code{YlOrRd}}
}
\description{
Plot cell density
}
\keyword{internal}
