% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_distance.R
\name{plotExpressionDistance}
\alias{plotExpressionDistance}
\title{Plot inter-sample expression distance}
\usage{
plotExpressionDistance(
  cluster.shifts,
  cell.groups = NULL,
  sample.groups = NULL,
  weighted.distance = FALSE,
  notch = TRUE,
  alpha = 0.2,
  min.cells = 10,
  palette = NULL,
  show.significance = FALSE
)
}
\arguments{
\item{cell.groups}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{sample.groups}{Named sample factor with cell names (default: stored vector)}

\item{weighted.distance}{whether to weigh the expression distance by the sizes of cell types (default: FALSE)}

\item{notch}{Show notches in plot, see ggplot2::geom_boxplot for more info (default=T)}

\item{alpha}{transparency level for the individual points (default: 0.2)}

\item{palette}{a set of colors to use for the conditions}

\item{show.significance}{whether to show statistical significance betwwen sample groups. wilcox.test was used; (* < 0.05; ** < 0.01; *** < 0.001)}

\item{name}{Test results to plot (default=expression.shifts)}
}
\value{
A ggplot2 object
}
\description{
Plot results from cao$estimateExpressionShiftMagnitudes()
}
