% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_distance.R
\name{plotExpressionDistancetSNE}
\alias{plotExpressionDistancetSNE}
\title{Plot sample-sample expression distance in tSNE}
\usage{
plotExpressionDistancetSNE(
  cluster.shifts,
  sample.groups,
  method = "MDS",
  cell.type = NULL,
  perplexity = 4,
  max_iter = 1000,
  palette = NULL
)
}
\arguments{
\item{sample.groups}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{method}{dimension reduction methods (MDS or tSNE) (default is MDS)}

\item{cell.type}{Named of cell type, default is null, it set plot sample-sample expression distance in tSNE for the cell type}

\item{perplexity}{tSNE perpexity (default: 4)}

\item{max_iter}{tSNE max_iter (default: 1e3)}
}
\value{
A ggplot2 object
}
\description{
Plot results from cao$estimateExpressionShiftMagnitudes()
}
