% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotExpressionShiftZScores}
\alias{plotExpressionShiftZScores}
\title{Plot Expression Shift Z Scores}
\usage{
plotExpressionShiftZScores(
  plot.df,
  size.norm = F,
  cell.groups = NULL,
  sample.per.cell = NULL
)
}
\arguments{
\item{plot.df}{Test results to plot}

\item{size.norm}{Plot size normalized results. Requires cell.groups, and sample.per.cell (default=F)}

\item{cell.groups}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{sample.per.cell}{Named sample factor with cell names (default: stored vector)}

\item{label}{Plot labels on size normalized plots (default=T)}
}
\value{
A ggplot2 object
}
\description{
Plot results from estimateExpressionShiftZScores
}
