% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFilteredDEGenes}
\alias{plotFilteredDEGenes}
\title{Plot filtered DE genes}
\usage{
plotFilteredDEGenes(de.filter, cell.groups, legend.position = "bottom")
}
\arguments{
\item{de.filter}{List of filtered differentially expressed genes, results from prepareOnthologyData (default: stored list)}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="bottom")}
}
\description{
Plot number of DE genes as a function of number of cells
}
