% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{plotGeneComparisonBetweenCondition}
\alias{plotGeneComparisonBetweenCondition}
\title{Plot Gene Comparison Between Conditions}
\usage{
plotGeneComparisonBetweenCondition(
  genes,
  con,
  condition.per.cell,
  z.scores = NULL,
  cur.scores = NULL,
  show.legend = TRUE,
  legend.pos = c(1, 1),
  size = 0.2,
  n.col = NULL,
  z.lims = NULL,
  max.expr = NULL,
  adj.list = NULL,
  ...
)
}
\arguments{
\item{genes}{Vector of genes to plot}

\item{con}{Conos object}

\item{condition.per.cell}{Named factor with cell names and condition}

\item{z.scores}{Z scores matrix. It is recommended to filter and adjust scores before plotting (default=NULL)}

\item{cur.scores}{Named numeric vector with gene names.}

\item{show.legend}{Plot legend (default=TRUE)}

\item{legend.pos}{Legend position, see ggplot2::theme}

\item{size}{Size of cells on plot}

\item{n.col}{Columns of plots. If NULL, will be number of conditions + 1 (default=NULL)}

\item{...}{Plotting variables propagated to conos:::embeddingPlot}
}
\description{
Plot Gene Comparison Between Conditions
}
