% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(
  df,
  color.per.type = NULL,
  row.order = NULL,
  col.order = F,
  legend.position = "right",
  legend.key.width = unit(8, "pt"),
  legend.title = "-log10(p-value)",
  x.axis.position = "top"
)
}
\arguments{
\item{df}{Data frame with plot data}

\item{color.per.type}{Colors per cell type (default=NULL)}

\item{row.order}{Forced row order (default=NULL)}

\item{col.order}{Forced column order (default=NULL)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{legend.key.width}{(default=unit(8, "pt))}

\item{legend.title}{Title on plot (default="-log10(p-value)")}

\item{x.axis.position}{Position of x axis (default="top")}
}
\value{
A ggplot2 object
}
\description{
Plot heatmap
}
