% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMeanMedValuesPerCellType}
\alias{plotMeanMedValuesPerCellType}
\title{Plot bar, point or boxplots showing mean/median values per cell type.
This is a generic function for plotting mean or median values per cell type
(used for expression shift distances and others).}
\usage{
plotMeanMedValuesPerCellType(
  df,
  pvalues = NULL,
  type = c("box", "point", "bar"),
  show.jitter = TRUE,
  notch = TRUE,
  jitter.alpha = 0.05,
  palette = NULL,
  ylab = "expression distance",
  yline = 1,
  plot.theme = theme_get(),
  jitter.size = 1,
  line.size = 0.75,
  trim = 0,
  order.x = TRUE,
  pvalue.y = NULL,
  y.max = NULL,
  y.offset = NULL,
  ns.symbol = ""
)
}
\arguments{
\item{df}{dataframe containing the results, including $value and $Type slots which will be summarized}

\item{type}{character vector (default=c('box', 'point', 'bar')) Type of a plot "bar" (default), "point" (mean + sd), or "box" for boxplot}

\item{show.jitter}{boolean (default = FALSE) Whether to show individual data points}

\item{notch}{\itemize{
\item whether to show notches in the boxplot version (default=TRUE)
}}

\item{jitter.alpha}{transparency value for the data points (default: 0.05)}

\item{palette}{\itemize{
\item cell type palette
}}
}
\value{
A ggplot2 object
}
\description{
Plot bar, point or boxplots showing mean/median values per cell type.
This is a generic function for plotting mean or median values per cell type
(used for expression shift distances and others).
}
\keyword{internal}
