% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNCellRegression}
\alias{plotNCellRegression}
\title{Plot number of cells regression}
\usage{
plotNCellRegression(
  n,
  n.total,
  x.lab = "Number of cells",
  y.lab = "N",
  legend.position = "right",
  label = TRUE,
  size = 4,
  palette = NULL,
  plot.line = TRUE,
  line.width = 0.5,
  plot.theme = ggplot2::theme_get()
)
}
\arguments{
\item{n}{the regressed variable}

\item{n.total}{integer Number of cells}

\item{x.lab}{character vector (default="Number of cells") x axis label}

\item{y.lab}{character vector (default="N") y axis label}

\item{legend.position}{character vector (default="right") legend position}

\item{label}{boolean (default=TRUE) Whether to show cell type labels}

\item{size}{integer (default=4) Label text size}

\item{palette}{color palette (default=NULL)}

\item{plot.line}{boolean (default=TRUE) Whether to plot the robust regression}

\item{line.width}{numeric (default=0.5) Regression line width}
}
\value{
ggplot2 plot of regression
}
\description{
Plot number of cells regression
}
\keyword{internal}
