% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOnthologyCorrelations}
\alias{plotOnthologyCorrelations}
\title{Plot onthology correlations}
\usage{
plotOnthologyCorrelations(type = NULL, ont.res)
}
\arguments{
\item{type}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Onthology resuls from estimateOnthology}
}
\value{
A ggplot2 object
}
\description{
Plot correlation matrix for onthologies between cell types
}
