% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOnthologyDistribution}
\alias{plotOnthologyDistribution}
\title{Plot onthology distribution}
\usage{
plotOnthologyDistribution(type = NULL, ont.res, cell.groups)
}
\arguments{
\item{type}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Onthology resuls from estimateOnthology}
}
\value{
A ggplot2 object
}
\description{
Bar plot of onthologies per cell type
}
