% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOnthologyDotplot}
\alias{plotOnthologyDotplot}
\title{Plot onthologies with dotplot}
\usage{
plotOnthologyDotplot(ont.res, genes, n = 20, p.adj = 0.05)
}
\arguments{
\item{ont.res}{Data frame with onthology resuls from estimateOnthology}

\item{genes}{Specify which genes are plotted, can either be 'down', 'up' or 'all' (default=NULL)}

\item{n}{Number of onthologies to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{p.adj}{Adjusted P cutoff (default=0.05)}
}
\value{
A ggplot2 object
}
\description{
Plot a dotplot of onthologies with adj. P values for a specific cell subgroup
}
