% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOnthologyHeatmap}
\alias{plotOnthologyHeatmap}
\title{Plot onthology heatmap}
\usage{
plotOnthologyHeatmap(
  type,
  ont.res,
  legend.position = "left",
  order = "all-max-row",
  n = 10
)
}
\arguments{
\item{type}{Onthology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default=NULL)}

\item{ont.res}{Onthology resuls from estimateOnthology}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{order}{Order of rows in heatmap. Can be 'unique' (only show pathways that are unique for any cell type); 'unique-max-row' (same as 'unique' but ordered by P value); 'all-max-rowsum' (all pathways ordered by cumulative P value for all cell types); 'all-max-row' (all pathways ordered by max P value) (default="all-max-row")}

\item{n}{Number of pathways to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}
}
\description{
Plot a heatmap of onthology P values per cell type
}
