% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologyDistribution}
\alias{plotOntologyDistribution}
\title{Plot ontology distribution}
\usage{
plotOntologyDistribution(type = NULL, ont.res, cell.groups)
}
\arguments{
\item{type}{Ontology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Ontology resuls from estimateOntology}
}
\value{
A ggplot2 object
}
\description{
Bar plot of ontologies per cell type
}
