% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologyFamily}
\alias{plotOntologyFamily}
\title{Plot ontology families}
\usage{
plotOntologyFamily(
  fam,
  data,
  plot.type = "complete",
  show.ids = FALSE,
  string.length = 18,
  legend.label.size = 1,
  legend.position = "topright",
  verbose = TRUE,
  n.cores = 1,
  reduce.edges = FALSE,
  font.size = 24
)
}
\arguments{
\item{fam}{List of ontology IDs for the chosen family}

\item{data}{Data frane if raw ontology data for the chosen cell type}

\item{plot.type}{How much of the family network should be plotted. Can be "complete" (entire network), "dense" (show 1 parent for each significant term), or "minimal" (only show significant terms) (default: complete)}

\item{show.ids}{Whether to show ontology IDs instead of names (default: F)}

\item{string.length}{Length of strings for wrapping in order to fit text within boxes (default: 18)}

\item{legend.label.size}{Size og legend labels (default: 1)}

\item{legend.position}{Position of legend (default: topright)}

\item{verbose}{Print messages (default: T)}

\item{n.cores}{Number of cores to use (default: 1)}
}
\value{
Rgraphviz object
}
\description{
Plot related ontologies in one hierarchical network plot
}
