% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologyHeatmap}
\alias{plotOntologyHeatmap}
\title{Plot ontology heatmap}
\usage{
plotOntologyHeatmap(
  type = "GO",
  ont.res,
  genes = NULL,
  legend.position = "left",
  selection = "all",
  n = 20,
  cell.subgroups = NULL
)
}
\arguments{
\item{type}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default=NULL)}

\item{ont.res}{Ontology resuls from estimateOntology}

\item{genes}{Specify which genes are plotted, can either be 'down', 'up' or 'all' (default=NULL)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{selection}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type); 'common' (only show terms that are not unique for any cell type); 'all' (all ontology terms) (default="all")}

\item{n}{Number of terms to show (default=10)}
}
\description{
Plot a heatmap of ontology P values per cell type
}
