% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologySimilarities}
\alias{plotOntologySimilarities}
\title{Plot ontology correlations}
\usage{
plotOntologySimilarities(type = NULL, ont.res)
}
\arguments{
\item{type}{Ontology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{List with ontology resuls from estimateOntology}
}
\value{
A ggplot2 object
}
\description{
Plot correlation matrix for ontologies between cell types
}
