% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologyTerms}
\alias{plotOntologyTerms}
\title{Plot ontology terms}
\usage{
plotOntologyTerms(
  type = NULL,
  ont.res,
  de.filter,
  cell.groups,
  legend.position = "bottom",
  label = T,
  label.x.pos = 0.01,
  label.y.pos = 1,
  rel_heights = c(2.5, 0.5),
  scale = 0.93
)
}
\arguments{
\item{type}{Ontology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Ontology results in a data frame}

\item{de.filter}{List of filtered differentially expressed genes, results from prepareOntologyData (default: stored list)}

\item{cell.groups}{Vector indicating cell group sizes with cell group names}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="bottom")}

\item{label.x.pos}{Plot label position on x axis (default=0.01)}

\item{label.y.pos}{Plot label position on y axis (default=1)}

\item{rel_heights}{Vector indicating relative heights for plots. Only relevant if show.legend=T. See cowplot::plot_grid for more info (default=c(2.5, 0.5))}

\item{scale}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (default=0.93)}
}
\value{
A ggplot2 object
}
\description{
Plot ontology terms as a function of both number of DE genes, and number of cells.
}
