% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPairwiseShiftsPerCellType}
\alias{plotPairwiseShiftsPerCellType}
\title{Plot pairwise expression shifts (used for expression shift distances - cluster-based)
Categorical and continuous covariates are supported.
Requires that covariate information is specified when running estimateExpressionShiftMagnitude()}
\usage{
plotPairwiseShiftsPerCellType(
  x,
  panel = "covariate",
  type = "box",
  show.jitter = TRUE,
  notch = TRUE,
  point.palette = NULL,
  jitter.alpha = 0.08,
  jitter.size = 1.1,
  trim = 0,
  palette = NULL,
  order.x = TRUE,
  yline = NULL,
  order.direction = "increasing",
  celltype.levels = NULL,
  pvalues = NULL,
  ylab = NULL,
  coord.flip = FALSE,
  cov.plot.keys = NULL,
  plot.theme = theme_bw(),
  line.size = 0.75,
  ns.symbol = "",
  pvalue.y = NULL,
  pvalue.size = 3.2,
  cont.palette = rev(RColorBrewer::brewer.pal(11, "Spectral"))
)
}
\arguments{
\item{x}{output of extractPairwiseShifts() after running estimateExpressionShiftMagnitude()}

\item{panel}{character c('covariate'(default), 'blocks', 'background') Type of a panel to plot}

\item{type}{character c('box' (default),'point') Type of plot to create}

\item{show.jitter}{boolean (default=TRUE) Whether to show individual data points}

\item{notch}{boolean Whether to show notches in the boxplot version (default=TRUE)}

\item{point.palette}{color palette for the points (default=NULL)}

\item{jitter.alpha}{transparency value for the data points (default: 0.08)}

\item{jitter.size}{marker size for the data points (default: 1.1)}

\item{trim}{numeric (default=0) Trim value for mean and CI calculation}

\item{palette}{\itemize{
\item cell type palette
}}

\item{order.x}{boolean (default=TRUE) Whether to order the x axis by mean/median values}

\item{yline}{numeric (default=0) y line to plot}

\item{order.direction}{character vector (default=c('increasing', 'decreasing')) Direction of ordering}

\item{celltype.levels}{character vector (default=NULL) Forced cell type order}

\item{pvalues}{named numeric vector (default=NULL) p-values per cell type to plot}

\item{ylab}{character vector (default=expression(hat(y)\link{AB} - mean(hat(y)\link{AA}))) y axis label}

\item{coord.flip}{flip coordinates of the covariate plot}

\item{plot.theme}{(default=ggplot2::theme_bw()) Plot theme}

\item{line.size}{numeric (default=0.75) Line size for error bars}

\item{ns.symbol}{character string (default="") Symbol to use for non-significant p-values}

\item{pvalue.y}{numeric (default=NULL) y position for the p-values}

\item{pvalue.size}{numeric (default=3.2) Font size for the p-values}
}
\value{
A ggplot2 object
}
\description{
Plot pairwise expression shifts (used for expression shift distances - cluster-based)
Categorical and continuous covariates are supported.
Requires that covariate information is specified when running estimateExpressionShiftMagnitude()
}
\keyword{internal}
