% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotProportionsSubset}
\alias{plotProportionsSubset}
\title{Plot proportions for a subset of cell types}
\usage{
plotProportionsSubset(
  legend.position = "right",
  cell.groups,
  sample.per.cell,
  sample.groups,
  cells.to.remove,
  cells.to.remain,
  notch = FALSE,
  alpha = 0.1,
  palette = NULL,
  show.significance = FALSE
)
}
\arguments{
\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{sample.per.cell}{Vector indicating sample name with cell names (default: stored vector)}

\item{sample.groups}{Vector indicating sample groups with sample names (default: stored vector)}

\item{cells.to.remove}{Vector of cell types to remove from the composition
#' @param alpha Transparency level on the data points (default: 0.1)}

\item{cells.to.remain}{Vector of cell types to remain in the composition}
}
\value{
A ggplot2 object
}
\description{
Plot the cell group proportions per sample
}
