% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotResidualsPerCelltype}
\alias{plotResidualsPerCelltype}
\title{Plot residuals per cell type
Categorical and continuous covariates are supported.}
\usage{
plotResidualsPerCelltype(
  res,
  design.mat,
  sample.ids = NULL,
  palette = NULL,
  font.size = 4,
  ylab = NULL,
  plot.theme = theme_bw(),
  jitter.size = 1,
  jitter.alpha = 0.8,
  yline = 0,
  cov.plot.keys = NULL,
  plot.per.celltype = FALSE,
  notch = TRUE,
  cont.palette = rev(RColorBrewer::brewer.pal(11, "Spectral")),
  residual.type = c("pearson", "raw")
)
}
\arguments{
\item{res}{output of performLMPermutations()}

\item{design.mat}{pairwise design matrices used for fitting the linear model}

\item{sample.ids}{character vector (default=NULL) Sample IDs corresponding to the rows of the design.mat$pairs}

\item{palette}{color palette for the points (default=NULL)}

\item{font.size}{integer (default=4) Font size for the plot}

\item{ylab}{character vector (default=NULL) y axis label}

\item{jitter.size}{numeric (default=1) Point size for the jittered points}

\item{jitter.alpha}{numeric (default=0.8) Transparency for the jittered points}

\item{cov.plot.keys}{character vector (default=NULL) Covariate name(s) to use for coloring
Requires that same covariate information is specified when running estimateExpressionShiftMagnitude()}

\item{plot.per.celltype}{boolean (default=FALSE) Whether to plot separate panels per cell type}

\item{cont.palette}{color palette for continuous covariates (default=rev(RColorBrewer::brewer.pal(11, "Spectral")))}

\item{cell.groups}{character vector defining cell type per sample}
}
\description{
Plot residuals per cell type
Categorical and continuous covariates are supported.
}
