% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_screen.R
\name{plotVariancePartition}
\alias{plotVariancePartition}
\title{Stacked bars of a variance partition per cell type}
\usage{
plotVariancePartition(parts, plot.theme = ggplot2::theme_bw())
}
\arguments{
\item{parts}{named list (per cell type) of \code{\link[=variancePartition]{variancePartition()}} results}

\item{plot.theme}{ggplot2 theme}
}
\value{
ggplot2 object
}
\description{
Stacked bars of a variance partition per cell type
}
