% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{prepareOntologyData}
\alias{prepareOntologyData}
\title{Prepare ontology data}
\usage{
prepareOntologyData(
  cms,
  de.raw,
  cell.groups,
  org.db,
  p.adj = 1,
  expr.cutoff = 0.05,
  universe = NULL,
  transposed = T,
  verbose = T,
  n.cores = 1
)
}
\arguments{
\item{cms}{list of counts matrices; column for gene and row for cell}

\item{de.raw}{List with differentially expressed genes per cell group}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{org.db}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{expr.cutoff}{Cutoff for cells per group expressing a DE gene, i.e., cutoff for highly-expressed genes (default=0.05)}

\item{universe}{Background gene list, NULL will take all input genes in de.raw (default=NULL)}

\item{transposed}{Indicate whether raw counts are transposed, i.e., cells as rows, genes as columns (default=T)}

\item{verbose}{Print progress (default=T)}

\item{n.cores}{Number of cores to use (default=1)}

\item{p.adj.cutoff}{Adj. P cutoff for filtering DE genes (default=0.05)}
}
\value{
A list containing DE ENSEMBL gene IDs, and filtered DE genes
}
\description{
Filter and prepare DE genes for ontology calculations
}
