% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{prepareOntologyPlotData}
\alias{prepareOntologyPlotData}
\title{Prepare plot data}
\usage{
prepareOntologyPlotData(ont.res, type, p.adj, min.genes)
}
\arguments{
\item{ont.res}{List of results from estimateOntology}

\item{type}{Type of ontology results, i.e., GO, GSEA, or DO}

\item{p.adj}{Cut-off for adj. P values}

\item{min.genes}{Min. number of significant genes per pathway}
}
\value{
List of ontology data for plotting
}
\description{
Prepare ontology results for plotting
}
