% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthology.R
\name{preparePathwayData}
\alias{preparePathwayData}
\title{Prepare pathway data}
\usage{
preparePathwayData(
  cms,
  de,
  cell.groups,
  transpose = T,
  OrgDB = org.Hs.eg.db,
  verbose = T,
  stat.cutoff = 3
)
}
\arguments{
\item{de}{List with differentially expressed genes per cell group}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{OrgDB}{Genome-wide annotation (default=org.Hs.eg.db)}

\item{verbose}{Print progress (default=T)}

\item{stat.cutoff}{Cutoff for filtering highly-expressed DE genes (default=3)}
}
\value{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
\description{
Filter and prepare DE genes for onthology calculations
}
