% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{produceResampling}
\alias{produceResampling}
\title{This function produces the resampling dataset (old two-group comparison)}
\usage{
produceResampling(cnts, groups, n.perm = 1000, seed = 239)
}
\arguments{
\item{cnts}{Counts of cell types in samples. Rows - samples, columns - cell types}

\item{groups}{Vector with boolean values. TRUE - sample in the case group, FALSE - sample in the control group}

\item{seed}{random seed}

\item{n.iter}{Number of permutations}

\item{remain.groups}{TRUE - if the labels of groups remain, FALSE - if to produce the null distribution}

\item{replace.samples}{TRUE - is bootstrap on samples, FALAE - if to remain samples}
}
\value{
Updated data frame with Z scores
}
\description{
This function produces the resampling dataset (old two-group comparison)
}
\keyword{internal}
