% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rblapply}
\alias{rblapply}
\title{Iterate over tree (list of lists of lists, etc) for \code{length(ids)} level,
interpret each element as a data.frame and bind them appending all levels as columns with
colnames correspoinding to \code{ids}}
\usage{
rblapply(list, ids, func)
}
\description{
Iterate over tree (list of lists of lists, etc) for \code{length(ids)} level,
interpret each element as a data.frame and bind them appending all levels as columns with
colnames correspoinding to \code{ids}
}
\keyword{internal}
