% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{referenceSet}
\alias{referenceSet}
\title{old two-group comparison: reference set selection}
\usage{
referenceSet(freqs, groups, p.thresh = 0.05)
}
\description{
old two-group comparison: reference set selection
}
\keyword{internal}
