% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{resampleContrast}
\alias{resampleContrast}
\title{Perform re-sampling in two ways: fixed total count and bootstrap}
\usage{
resampleContrast(
  d.counts,
  d.groups,
  n.cell.counts = 500,
  n.seed = 239,
  n.iter = 1000
)
}
\value{
Order of inner nodes
}
\description{
Perform re-sampling in two ways: fixed total count and bootstrap
}
