% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{resampleCounts}
\alias{resampleCounts}
\title{Resampling of cell type counts}
\usage{
resampleCounts(
  d.counts,
  n.tot.count = 500,
  d.groups = NULL,
  n.seed = 239,
  f.bootstrap = TRUE
)
}
\arguments{
\item{d.counts}{Cell count table}

\item{n.tot.count}{Total cell type counts}

\item{d.groups}{if provided then resampling controls presence of both groups in a new dataset}

\item{n.seed}{Random seed}

\item{f.bootstrap}{if to use the bootstrap, dafault (TRUE) means the use of bootstrap}
}
\value{
Resampled dataset
}
\description{
Resampling of cell type counts
}
