% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{runCoda}
\alias{runCoda}
\title{This function runs the coda analysis (old two-group comparison)}
\usage{
runCoda(
  cnts,
  groups,
  n.seed = 239,
  n.boot = 1000,
  ref.cell.type = NULL,
  null.distr = FALSE,
  method = "lda",
  n.cores = 1,
  verbose = TRUE
)
}
\description{
This function runs the coda analysis (old two-group comparison)
}
\keyword{internal}
