% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{saveDEasJSON}
\alias{saveDEasJSON}
\title{Save DE results as JSON files}
\usage{
saveDEasJSON(
  de.raw,
  saveprefix = NULL,
  dir.name = "JSON",
  gene.metadata = NULL,
  cluster.sep.chr = "<!!>"
)
}
\arguments{
\item{de.raw}{List of DE results}

\item{saveprefix}{Prefix for created files (default=NULL)}

\item{dir.name}{Name for directory with results. If it doesn't exist, it will be created. To disable, set as NULL (default="JSON")}

\item{gene.metadata}{(default=NULL)}

\item{cluster.sep.chr}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}
}
\description{
Save DE results as JSON files
}
