% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_partitions.R
\name{sbp2balances}
\alias{sbp2balances}
\title{Transform sequential binary partition (sbp) into balances}
\usage{
sbp2balances(d.counts, sbp.pop)
}
\arguments{
\item{d.counts}{Cell count table}

\item{sbp.pop}{matrix with binary partiotions in rows}
}
\value{
matrix of balances in columns; rows are samples
}
\description{
Transform sequential binary partition (sbp) into balances
}
\keyword{internal}
