% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_tree_new.R
\name{sbp2tree}
\alias{sbp2tree}
\title{Get tree from balances}
\usage{
sbp2tree(sbpart)
}
\arguments{
\item{sbpart}{A contrast matrix for balances, sequential binary partition}
}
\value{
phylo tree
}
\description{
Get tree from balances
}
\details{
We use ellipsis because a previous version of this function
contains additional and insignificant parameters.
}
\keyword{internal}
