% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sbpFromHclust}
\alias{sbpFromHclust}
\title{Build SBP (D x K) aligned to hclust internal nodes (1..K)}
\usage{
sbpFromHclust(h)
}
\description{
Build SBP (D x K) aligned to hclust internal nodes (1..K)
}
\keyword{internal}
