% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{standardizeDEColumns}
\alias{standardizeDEColumns}
\title{Standardize DE result column names across DE methods}
\usage{
standardizeDEColumns(res)
}
\description{
Standardize DE result column names across DE methods
}
\keyword{internal}
