% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{strpart}
\alias{strpart}
\title{Split strings and extract nth element}
\usage{
strpart(x, split, n, fixed = FALSE)
}
\arguments{
\item{x}{input character vector}

\item{split}{character vector containing a regular expression for splitting}

\item{n}{element to extract}

\item{fixed}{passed to strsplit}
}
\description{
Function building on base::strsplit to extract nth element of a character string after splitting
}
\keyword{internal}
