% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{subsampleCellsPerTypePerSample}
\alias{subsampleCellsPerTypePerSample}
\title{Subsample cells per cell type per sample}
\usage{
subsampleCellsPerTypePerSample(
  raw.mats.sub,
  cell.groups,
  sample.per.cell,
  core.samples,
  n.cells.subsample,
  min.cell.count = 1L,
  seed = NULL
)
}
\arguments{
\item{raw.mats.sub}{named list of raw count matrices (cells x genes)}
}
\description{
Subsample cells per cell type per sample
}
\keyword{internal}
