% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{subsetMatricesWithCommonGenes}
\alias{subsetMatricesWithCommonGenes}
\title{Subset matrices with common genes}
\usage{
subsetMatricesWithCommonGenes(cms, sample.groups = NULL)
}
\arguments{
\item{cms}{List with count matrices}

\item{sample.groups}{(default=NULL)}
}
\value{
list with count matrices with common genes
}
\description{
Subset matrices with common genes
}
\keyword{internal}
