% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{subsetMatricesWithCommonGenesSamples}
\alias{subsetMatricesWithCommonGenesSamples}
\title{Subset matrices with common genes (by sample vector)}
\usage{
subsetMatricesWithCommonGenesSamples(cms, samples = NULL)
}
\arguments{
\item{cms}{named list of count matrices (cells x genes)}

\item{samples}{character vector of sample IDs to keep (optional)}
}
\description{
Subset matrices with common genes (by sample vector)
}
\keyword{internal}
