% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_legend_position}
\alias{theme_legend_position}
\title{Helper function for creating color palettes
Syncs input to ggplot2::theme() params legend.position (i.e. the position of the legends)
and legend.justification (i.e. the anchor point for positioning legend inside the plot)}
\usage{
theme_legend_position(position)
}
\arguments{
\item{position}{character string to pass into ggplot2::theme(legend.position=position, legend.justification=position)}
}
\value{
returns ggplot2 theme() such that ggplot2::theme(legend.position=position, legend.justification=position)
}
\description{
Helper function for creating color palettes
Syncs input to ggplot2::theme() params legend.position (i.e. the position of the legends)
and legend.justification (i.e. the anchor point for positioning legend inside the plot)
}
