% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{twist}
\alias{twist}
\title{Swap character strings for their numeric names}
\usage{
twist(x)
}
\arguments{
\item{x}{Character string with numeric names}
}
\value{
Numeric vector with character names
}
\description{
Set numeric names as output and set names as input character strings
}
