% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{validateDEPerCellTypeParams}
\alias{validateDEPerCellTypeParams}
\title{Validate parameters per cell type}
\usage{
validateDEPerCellTypeParams(raw.mats, cell.groups, model)
}
\arguments{
\item{raw.mats}{List of raw count matrices}

\item{cell.groups}{Named clustering/annotation factor with cell names}

\item{model}{Model object returned by buildDesignMatrices}
}
\description{
Validate parameters per cell type
}
\keyword{internal}
