% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{validatePerCellTypeParams}
\alias{validatePerCellTypeParams}
\title{Validate parameters per cell type}
\usage{
validatePerCellTypeParams(
  raw.mats,
  cell.groups,
  sample.groups,
  ref.level,
  cluster.sep.chr
)
}
\arguments{
\item{raw.mats}{List of raw count matrices}

\item{cell.groups}{Named clustering/annotation factor with cell names}

\item{sample.groups}{Named factor with cell names indicating condition/sample, e.g., ctrl/disease}

\item{ref.level}{Reference cluster level in 'sample.groups', e.g., ctrl, healthy, wt}

\item{cluster.sep.chr}{Character string of length 1 specifying a delimiter to separate cluster and app names}
}
\description{
Validate parameters per cell type
}
