% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{wrap_strings}
\alias{wrap_strings}
\title{Wrap strings for readibility on plots}
\usage{
wrap_strings(strings, width)
}
\arguments{
\item{strings}{Text strings}

\item{width}{Max length before inserting line shift}
}
\value{
Text strings with inserted line shifts
}
\description{
Handy function for wrapping long text strings for usage in plots
}
\examples{
wrapped_string <- wrap_strings(c("This is a long text to be wrapped"), 10)

}
