% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
The class encompasses etc etc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{Number of cores (default=1)}

\item{\code{verbose}}{boolean Whether to provide verbose output with diagnostic messages (default=FALSE)}

\item{\code{test.results}}{list Results of the estimations, ready to use (default=list())}

\item{\code{cache}}{list Intermediate results of the estimations, which can be used during some other computations (default=list())}

\item{\code{data.object}}{list The main object storing data (Conos or Seurat) (default=list())}

\item{\code{sample.metada}}{Data frame with annotation of covariates per sample (default=NULL)}

\item{\code{full.meta}}{Full sample metadata (default=NULL)}

\item{\code{cell.groups}}{Named factor with cell names with cluster per cell (default=NULL)}

\item{\code{embedding}}{2D embedding to visualize the cells in (default=NULL)}

\item{\code{sample.per.cell}}{Named factor with cell names (default=NULL)}

\item{\code{formula}}{Design formula for the analysis (default=NULL)}

\item{\code{contrast}}{Character vector c(var, alt, ref) specifying contrasts for the analysis (default=NULL)}

\item{\code{model.matrices}}{Design model matrices for the analysis (default=NULL)}

\item{\code{ref.level}}{Reference level for sample.group vector (default=NULL)}

\item{\code{target.level}}{Target/disease level for sample.group vector}

\item{\code{sample.id}}{Character of column name containing sample IDs in sample.meta}

\item{\code{block.vars}}{Character of column name containing variable name to restrict permutations to}

\item{\code{numeric.ref}}{Reference points for numeric covariates in contrasts: "auto" or a named list}

\item{\code{method}}{for permutation testing (default=NULL)}

\item{\code{sample.groups.palette}}{Color palette for the sample.groups (default=NULL)}

\item{\code{cell.groups.palette}}{Color palette for the cell.groups (default=NULL)}

\item{\code{plot.theme}}{ggplot2 theme for all plots (default=NULL)}

\item{\code{plot.params}}{list with parameters, forwarded to all \code{plotEmbedding} calls (default=NULL)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sample.metada}}{Data frame with annotation of covariates per sample (default=NULL)}

\item{\code{sample.id}}{Character of column name containing sample IDs in sample.meta}

\item{\code{method}}{for permutation testing (default=NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cacoa-new}{\code{Cacoa$new()}}
\item \href{#method-Cacoa-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-Cacoa-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-Cacoa-plotExpressionShiftResiduals}{\code{Cacoa$plotExpressionShiftResiduals()}}
\item \href{#method-Cacoa-estimatePerCellTypeDE}{\code{Cacoa$estimatePerCellTypeDE()}}
\item \href{#method-Cacoa-estimateDEPerCellType}{\code{Cacoa$estimateDEPerCellType()}}
\item \href{#method-Cacoa-estimateDEStabilityPerCellType}{\code{Cacoa$estimateDEStabilityPerCellType()}}
\item \href{#method-Cacoa-estimateDEStabilityPerGene}{\code{Cacoa$estimateDEStabilityPerGene()}}
\item \href{#method-Cacoa-plotDEStabilityPerCellType}{\code{Cacoa$plotDEStabilityPerCellType()}}
\item \href{#method-Cacoa-plotDEStabilityPerGene}{\code{Cacoa$plotDEStabilityPerGene()}}
\item \href{#method-Cacoa-plotNumberOfDEGenes}{\code{Cacoa$plotNumberOfDEGenes()}}
\item \href{#method-Cacoa-plotVolcano}{\code{Cacoa$plotVolcano()}}
\item \href{#method-Cacoa-saveDeAsJson}{\code{Cacoa$saveDeAsJson()}}
\item \href{#method-Cacoa-plotEmbedding}{\code{Cacoa$plotEmbedding()}}
\item \href{#method-Cacoa-estimateOntology}{\code{Cacoa$estimateOntology()}}
\item \href{#method-Cacoa-estimateOntologyFamilies}{\code{Cacoa$estimateOntologyFamilies()}}
\item \href{#method-Cacoa-getFamiliesPerGO}{\code{Cacoa$getFamiliesPerGO()}}
\item \href{#method-Cacoa-plotNumOntologyTermsPerType}{\code{Cacoa$plotNumOntologyTermsPerType()}}
\item \href{#method-Cacoa-plotOntology}{\code{Cacoa$plotOntology()}}
\item \href{#method-Cacoa-plotOntologyHeatmap}{\code{Cacoa$plotOntologyHeatmap()}}
\item \href{#method-Cacoa-plotOntologyHeatmapCollapsed}{\code{Cacoa$plotOntologyHeatmapCollapsed()}}
\item \href{#method-Cacoa-plotOntologySimilarities}{\code{Cacoa$plotOntologySimilarities()}}
\item \href{#method-Cacoa-plotOntologyFamily}{\code{Cacoa$plotOntologyFamily()}}
\item \href{#method-Cacoa-saveOntologyAsTable}{\code{Cacoa$saveOntologyAsTable()}}
\item \href{#method-Cacoa-saveFamiliesAsTable}{\code{Cacoa$saveFamiliesAsTable()}}
\item \href{#method-Cacoa-plotCellGroupSizes}{\code{Cacoa$plotCellGroupSizes()}}
\item \href{#method-Cacoa-plotCellGroupAbundanceVariation}{\code{Cacoa$plotCellGroupAbundanceVariation()}}
\item \href{#method-Cacoa-plotCodaSpace}{\code{Cacoa$plotCodaSpace()}}
\item \href{#method-Cacoa-plotCodaContrastTree}{\code{Cacoa$plotCodaContrastTree()}}
\item \href{#method-Cacoa-plotContrastTree}{\code{Cacoa$plotContrastTree()}}
\item \href{#method-Cacoa-plotCompositionSimilarity}{\code{Cacoa$plotCompositionSimilarity()}}
\item \href{#method-Cacoa-estimateCellLoadings}{\code{Cacoa$estimateCellLoadings()}}
\item \href{#method-Cacoa-plotCodaGlobal}{\code{Cacoa$plotCodaGlobal()}}
\item \href{#method-Cacoa-plotCodaHeatmap}{\code{Cacoa$plotCodaHeatmap()}}
\item \href{#method-Cacoa-plotCodaCellTypes}{\code{Cacoa$plotCodaCellTypes()}}
\item \href{#method-Cacoa-plotCellLoadings}{\code{Cacoa$plotCellLoadings()}}
\item \href{#method-Cacoa-estimateCellDensity}{\code{Cacoa$estimateCellDensity()}}
\item \href{#method-Cacoa-plotCellDensity}{\code{Cacoa$plotCellDensity()}}
\item \href{#method-Cacoa-plotCellDensityVariation}{\code{Cacoa$plotCellDensityVariation()}}
\item \href{#method-Cacoa-estimateDiffCellDensity}{\code{Cacoa$estimateDiffCellDensity()}}
\item \href{#method-Cacoa-plotDiffCellDensity}{\code{Cacoa$plotDiffCellDensity()}}
\item \href{#method-Cacoa-plotDiffCellDensityResiduals}{\code{Cacoa$plotDiffCellDensityResiduals()}}
\item \href{#method-Cacoa-plotExpressionDistance}{\code{Cacoa$plotExpressionDistance()}}
\item \href{#method-Cacoa-getSampleDistanceMatrix}{\code{Cacoa$getSampleDistanceMatrix()}}
\item \href{#method-Cacoa-plotSampleDistances}{\code{Cacoa$plotSampleDistances()}}
\item \href{#method-Cacoa-estimateMetadataSeparation}{\code{Cacoa$estimateMetadataSeparation()}}
\item \href{#method-Cacoa-plotMetadataSeparation}{\code{Cacoa$plotMetadataSeparation()}}
\item \href{#method-Cacoa-estimateClusterFreeDE}{\code{Cacoa$estimateClusterFreeDE()}}
\item \href{#method-Cacoa-getMostChangedGenes}{\code{Cacoa$getMostChangedGenes()}}
\item \href{#method-Cacoa-estimateClusterFreeExpressionShifts}{\code{Cacoa$estimateClusterFreeExpressionShifts()}}
\item \href{#method-Cacoa-smoothClusterFreeZScores}{\code{Cacoa$smoothClusterFreeZScores()}}
\item \href{#method-Cacoa-estimateGenePrograms}{\code{Cacoa$estimateGenePrograms()}}
\item \href{#method-Cacoa-plotGeneProgramScores}{\code{Cacoa$plotGeneProgramScores()}}
\item \href{#method-Cacoa-plotGeneProgramGenes}{\code{Cacoa$plotGeneProgramGenes()}}
\item \href{#method-Cacoa-plotClusterFreeExpressionShifts}{\code{Cacoa$plotClusterFreeExpressionShifts()}}
\item \href{#method-Cacoa-plotClusterFreeResiduals}{\code{Cacoa$plotClusterFreeResiduals()}}
\item \href{#method-Cacoa-plotMostChangedGenes}{\code{Cacoa$plotMostChangedGenes()}}
\item \href{#method-Cacoa-plotGeneExpressionComparison}{\code{Cacoa$plotGeneExpressionComparison()}}
\item \href{#method-Cacoa-getConditionPerCell}{\code{Cacoa$getConditionPerCell()}}
\item \href{#method-Cacoa-getJointCountMatrix}{\code{Cacoa$getJointCountMatrix()}}
\item \href{#method-Cacoa-getGOEnvironment}{\code{Cacoa$getGOEnvironment()}}
\item \href{#method-Cacoa-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.metadata = NULL,
  sample.ids = NULL,
  formula = NULL,
  contrast = NULL,
  numeric.ref = "auto",
  block.vars = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  sample.groups.palette = NULL,
  cell.groups.palette = NULL,
  embedding = NULL,
  n.cores = 1,
  verbose = TRUE,
  graph.name = NULL,
  assay.name = "RNA",
  data.layer = "scale.data",
  plot.theme = ggplot2::theme_bw(),
  plot.params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  formula = NULL,
  contrast = NULL,
  pairContrast = NULL,
  pairFormula = NULL,
  block.vars = self$block.vars,
  sample.metadata = self$sample.meta,
  sample.ids = self$sample.ids,
  dist = NULL,
  dist.type = "shift",
  min.cells.per.sample = 10,
  min.samp.per.type = 2,
  min.gene.frac = 0.01,
  genes = NULL,
  perm.method = "freedman-lane",
  robust.method = "none",
  n.pcs = NULL,
  na.mode = "drop",
  alternative = "greater",
  return.residuals = TRUE,
  return.sampled.stats = TRUE,
  name = "expression.shifts",
  n.permutations = 1000,
  return.sampled.fits = FALSE,
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}

\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  type = "box",
  notch = TRUE,
  show.jitter = TRUE,
  cov.plot.keys = NULL,
  jitter.alpha = 0.05,
  jitter.size = 1,
  show.pvalues = c("adjusted", "raw", "none"),
  celltype.levels = NULL,
  panel = c("covariate", "block", "background"),
  order.direction = "increasing",
  ylab = "Centered Pairwise Shifts",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotExpressionShiftResiduals"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotExpressionShiftResiduals}{}}}
\subsection{Method \code{plotExpressionShiftResiduals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftResiduals(
  name = "expression.shifts",
  cov.plot.keys = NULL,
  residual.type = c("pearson", "raw"),
  jitter.alpha = 0.8,
  jitter.size = 1,
  yline = 0,
  plot.per.celltype = FALSE,
  palette = NULL,
  ylab = "Residual Variance",
  cont.palette = rev(RColorBrewer::brewer.pal(11, "Spectral")),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimatePerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimatePerCellTypeDE}{}}}
\subsection{Method \code{estimatePerCellTypeDE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDE(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateDEPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateDEPerCellType}{}}}
\subsection{Method \code{estimateDEPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEPerCellType(
  cell.groups = self$cell.groups,
  sample.meta = self$sample.meta,
  sample.per.cell = self$sample.per.cell,
  formula = NULL,
  contrast = NULL,
  name = "de",
  test = "DESeq2.Wald",
  resampling.method = NULL,
  n.resamplings = 30,
  seed.resampling = 239,
  min.cell.frac = 0.05,
  common.genes = FALSE,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  independent.filtering = FALSE,
  min.cell.count = 10,
  n.cells.subsample = NULL,
  verbose = self$verbose,
  fix.n.samples = NULL,
  genes.to.omit = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}

\item{\code{name}}{string Field name where the test results are stored}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateDEStabilityPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateDEStabilityPerCellType}{}}}
\subsection{Method \code{estimateDEStabilityPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityPerCellType(
  de.name = "de",
  name = "de.jaccards",
  top.n.genes = NULL,
  p.val.cutoff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateDEStabilityPerGene"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateDEStabilityPerGene}{}}}
\subsection{Method \code{estimateDEStabilityPerGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityPerGene(
  de.name = "de",
  top.n.genes = 500,
  p.adj.cutoff = NULL,
  visualize = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotDEStabilityPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotDEStabilityPerCellType}{}}}
\subsection{Method \code{plotDEStabilityPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityPerCellType(
  name = "de.jaccards",
  notch = FALSE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.pairs = FALSE,
  sort.order = TRUE,
  pallete = self$cell.groups.palette,
  set.fill = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotDEStabilityPerGene"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotDEStabilityPerGene}{}}}
\subsection{Method \code{plotDEStabilityPerGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityPerGene(
  name = "de",
  cell.type = NULL,
  stability.score = "stab.median.rank",
  top.n = 500
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotNumberOfDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotNumberOfDEGenes}{}}}
\subsection{Method \code{plotNumberOfDEGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumberOfDEGenes(
  name = "de",
  p.adjust = TRUE,
  pvalue.cutoff = 0.05,
  show.resampling.results = TRUE,
  show.jitter = FALSE,
  jitter.alpha = 0.05,
  type = "bar",
  notch = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotVolcano"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotVolcano}{}}}
\subsection{Method \code{plotVolcano()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotVolcano(
  name = "de",
  cell.types = NULL,
  palette = NULL,
  build.panel = TRUE,
  n.col = 3,
  color.var = "CellFrac",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-saveDeAsJson"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-saveDeAsJson}{}}}
\subsection{Method \code{saveDeAsJson()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveDeAsJson(
  saveprefix = NULL,
  dir.name = "JSON",
  de.raw = NULL,
  sample.groups = self$sample.groups,
  de.name = "de",
  ref.level = self$ref.level,
  gene.metadata = NULL,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{reference sample group level (default=self$ref.level)}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotEmbedding}{}}}
\subsection{Method \code{plotEmbedding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotEmbedding(
  embedding = self$embedding,
  color.by = NULL,
  plot.theme = self$plot.theme,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateOntology"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateOntology}{}}}
\subsection{Method \code{estimateOntology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntology(
  type = c("GO", "DO", "GSEA"),
  name = NULL,
  de.name = "de",
  org.db,
  n.top.genes = 500,
  p.adj = 1,
  p.adjust.method = "BH",
  readable = TRUE,
  min.gs.size = 10,
  max.gs.size = 500,
  keep.gene.sets = FALSE,
  ignore.cache = NULL,
  de.raw = NULL,
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values (default="BH")}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateOntologyFamilies"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateOntologyFamilies}{}}}
\subsection{Method \code{estimateOntologyFamilies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntologyFamilies(name = "GO", p.adj = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getFamiliesPerGO"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getFamiliesPerGO}{}}}
\subsection{Method \code{getFamiliesPerGO()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getFamiliesPerGO(
  name = "GO",
  go.term = NULL,
  go.id = NULL,
  common = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotNumOntologyTermsPerType"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotNumOntologyTermsPerType}{}}}
\subsection{Method \code{plotNumOntologyTermsPerType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumOntologyTermsPerType(
  name = "GO",
  genes = "up",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  families = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotOntology"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotOntology}{}}}
\subsection{Method \code{plotOntology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntology(
  cell.type,
  name = "GO",
  plot = "dot",
  genes = c("up", "down", "all"),
  subtype = c("BP", "CC", "MF"),
  n = 20,
  p.adj = 0.05,
  min.genes = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotOntologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotOntologyHeatmap}{}}}
\subsection{Method \code{plotOntologyHeatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmap(
  name = "GO",
  genes = "up",
  subtype = "BP",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  top.n = Inf,
  legend.position = "left",
  selection = "all",
  cluster = TRUE,
  cell.subgroups = NULL,
  row.order = TRUE,
  col.order = TRUE,
  max.log.p = 10,
  only.family.children = FALSE,
  description.regex = NULL,
  description.exclude.regex = NULL,
  clust.naming = "medoid",
  readjust.p = TRUE,
  p.adjust.method = "BH",
  palette = NULL,
  color.range = NULL,
  return.info = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values (default="BH")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotOntologyHeatmapCollapsed"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotOntologyHeatmapCollapsed}{}}}
\subsection{Method \code{plotOntologyHeatmapCollapsed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmapCollapsed(
  name = "GO",
  genes = "up",
  subtype = "BP",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  n = 20,
  legend.position = "left",
  selection = "all",
  max.log.p = 10,
  cluster = TRUE,
  cell.subgroups = NULL,
  palette = NULL,
  row.order = TRUE,
  col.order = TRUE,
  only.family.children = FALSE,
  readjust.p = TRUE,
  p.adjust.method = "BH",
  description.regex = NULL,
  description.exclude.regex = NULL,
  distance = "manhattan",
  clust.method = "complete",
  clust.naming = "consensus",
  n.words = 5,
  exclude.words = NULL,
  return.info = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values (default="BH")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotOntologySimilarities"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotOntologySimilarities}{}}}
\subsection{Method \code{plotOntologySimilarities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologySimilarities(
  name = "GO",
  subtype = c("BP", "MF", "CC"),
  genes = "up",
  p.adj = 0.05,
  only.family.children = FALSE,
  q.value = 0.2,
  min.genes = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotOntologyFamily"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotOntologyFamily}{}}}
\subsection{Method \code{plotOntologyFamily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyFamily(
  name = "GO",
  cell.type,
  family = NULL,
  genes = "up",
  subtype = "BP",
  plot.type = c("complete", "dense", "minimal"),
  show.ids = FALSE,
  string.length = 14,
  legend.label.size = 1,
  legend.position = "topright",
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-saveOntologyAsTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-saveOntologyAsTable}{}}}
\subsection{Method \code{saveOntologyAsTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveOntologyAsTable(
  file,
  name = "GO",
  subtype = NULL,
  genes = NULL,
  p.adj = 0.05,
  sep = "\\t",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-saveFamiliesAsTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-saveFamiliesAsTable}{}}}
\subsection{Method \code{saveFamiliesAsTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveFamiliesAsTable(
  file,
  name = "GO",
  subtype = NULL,
  genes = NULL,
  p.adj = 0.05,
  sep = "\\t",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCellGroupSizes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCellGroupSizes}{}}}
\subsection{Method \code{plotCellGroupSizes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellGroupSizes(
  cell.groups = self$cell.groups,
  condition = NULL,
  show.significance = FALSE,
  filter.empty.cell.types = TRUE,
  proportions = TRUE,
  palette = self$sample.groups.palette,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCellGroupAbundanceVariation"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCellGroupAbundanceVariation}{}}}
\subsection{Method \code{plotCellGroupAbundanceVariation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellGroupAbundanceVariation(
  cell.groups = self$cell.groups,
  type = "mad",
  rotate.xticks = TRUE,
  min.rel.abundance = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCodaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCodaSpace}{}}}
\subsection{Method \code{plotCodaSpace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaSpace(
  space = "CDA",
  cell.groups = self$cell.groups,
  font.size = 3,
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCodaContrastTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCodaContrastTree}{}}}
\subsection{Method \code{plotCodaContrastTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaContrastTree(
  name = "coda",
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  filter.empty.cell.types = TRUE,
  adjust.pvalues = TRUE,
  reorder.tree = TRUE,
  score.palette = c(low = "#2166AC", mid = "grey95", high = "#B2182B"),
  score.midpoint = 0,
  score.limits = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotContrastTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotContrastTree}{}}}
\subsection{Method \code{plotContrastTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotContrastTree(
  cell.groups = self$cell.groups,
  palette = self$sample.groups.palette,
  name = "coda",
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  filter.empty.cell.types = TRUE,
  adjust.pvalues = TRUE,
  h.method = c("both", "up", "down"),
  reorder.tree = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}

\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCompositionSimilarity"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCompositionSimilarity}{}}}
\subsection{Method \code{plotCompositionSimilarity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCompositionSimilarity(
  cell.groups = self$cell.groups,
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  palette = brewerPalette("YlOrRd", rev = FALSE),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateCellLoadings}{}}}
\subsection{Method \code{estimateCellLoadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellLoadings(
  n.permutations = 1000,
  name = "coda",
  n.seed = 239,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  filter.empty.cell.types = TRUE,
  n.cores = self$n.cores,
  verbose = self$verbose,
  method = "lda",
  formula = NULL,
  contrast = NULL,
  perm.method = c("freedman-lane", "block"),
  zero.pseudocount = 0.1,
  basis.type = c("default"),
  ref.p.thresh = 0.3,
  ref.min.size = 1,
  ref.max.size = 3,
  block.vars = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCodaGlobal"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCodaGlobal}{}}}
\subsection{Method \code{plotCodaGlobal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaGlobal(
  name = "coda",
  style = c("bar", "violin"),
  fdr.threshold = 0.05,
  drop.intercept = TRUE,
  use.contrast.title = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCodaHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCodaHeatmap}{}}}
\subsection{Method \code{plotCodaHeatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaHeatmap(
  name = "coda",
  coef.names = NULL,
  fdr.threshold = 0.05,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCodaCellTypes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCodaCellTypes}{}}}
\subsection{Method \code{plotCodaCellTypes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaCellTypes(
  name = "coda",
  coef = "contrast",
  style = c("bar", "violin"),
  value = c("predicted", "reference-centered", "volcano"),
  fdr.threshold = 0.05,
  sortCellTypes = FALSE,
  labelCellTypes = c("significant", "all", "none"),
  palette = self$cell.groups.palette,
  use.contrast.title = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCellLoadings}{}}}
\subsection{Method \code{plotCellLoadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellLoadings(
  alpha = 0.01,
  palette = self$cell.groups.palette,
  font.size = NULL,
  name = "coda",
  ordering = "pvalue",
  show.pvals = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateCellDensity}{}}}
\subsection{Method \code{estimateCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellDensity(
  bins = 400,
  method = "kde",
  name = "cell.density",
  beta = 30,
  estimate.variation = TRUE,
  contrast = NULL,
  formula = NULL,
  verbose = self$verbose,
  n.cores = self$n.cores,
  sample.metadata = self$sample.meta,
  block.vars = self$block.vars,
  bandwidth = 0.05,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCellDensity}{}}}
\subsection{Method \code{plotCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensity(
  show.grid = TRUE,
  add.points = TRUE,
  size = 0.1,
  show.legend = FALSE,
  palette = NULL,
  point.col = "#313695",
  contours = NULL,
  contour.color = "black",
  contour.conf = "10\%",
  name = "cell.density",
  show.cell.groups = TRUE,
  cell.groups = self$cell.groups,
  font.size = c(2, 4),
  color.range = c(0, "99\%"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotCellDensityVariation"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotCellDensityVariation}{}}}
\subsection{Method \code{plotCellDensityVariation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensityVariation(
  type = "mad",
  plot.type = "embedding",
  name = "cell.density",
  cutoff = NULL,
  condition = c("both", "ref", "target"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateDiffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateDiffCellDensity}{}}}
\subsection{Method \code{estimateDiffCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDiffCellDensity(
  type = "permutation",
  adjust = NULL,
  name = "cell.density",
  sample.metadata = self$sample.meta,
  formula = NULL,
  contrast = NULL,
  block.vars = self$block.vars,
  perm.method = "freedman-lane",
  robust.method = "none",
  na.mode = "drop",
  alternative = "two-sided",
  return.residuals = FALSE,
  return.sampled.stats = TRUE,
  n.permutations = 999,
  smooth = TRUE,
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotDiffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotDiffCellDensity}{}}}
\subsection{Method \code{plotDiffCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDiffCellDensity(
  name = "cell.density",
  size = 0.2,
  palette = NULL,
  adjust = NULL,
  contours = NULL,
  contour.color = "black",
  contour.conf = "10\%",
  plot.na = FALSE,
  color.range = NULL,
  mid.color = "gray95",
  scale.z.palette = adjust,
  min.z = qnorm(0.9),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotDiffCellDensityResiduals"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotDiffCellDensityResiduals}{}}}
\subsection{Method \code{plotDiffCellDensityResiduals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDiffCellDensityResiduals(
  name = "cell.density",
  type = "permutation",
  size = 0.2,
  palette = brewerPalette("GnBu"),
  alpha = 0.2,
  sample.metadata = self$sample.meta,
  cov.plot.keys = NULL,
  metric = c("mean_abs", "sd"),
  build.panel = TRUE,
  jitter.size = 1,
  jitter.alpha = 0.8,
  cont.palette = rev(RColorBrewer::brewer.pal(11, "Spectral")),
  residual.type = c("pearson", "raw"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotExpressionDistance"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotExpressionDistance}{}}}
\subsection{Method \code{plotExpressionDistance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistance(
  name = "expression.shifts",
  joint = FALSE,
  condition = NULL,
  values = c("unadjusted", "adjusted"),
  palette = NULL,
  show.significance = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getSampleDistanceMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getSampleDistanceMatrix}{}}}
\subsection{Method \code{getSampleDistanceMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getSampleDistanceMatrix(
  space = c("expression.shifts", "coda", "pseudo.bulk"),
  values = c("unadjusted", "adjusted"),
  cell.type = NULL,
  pair.set = c("all", "core"),
  dist = NULL,
  name = NULL,
  verbose = self$verbose,
  sample.subset = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{sample.subset}}{subset data for analysis only to the given samples}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotSampleDistances"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotSampleDistances}{}}}
\subsection{Method \code{plotSampleDistances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotSampleDistances(
  space = "expression.shifts",
  method = "MDS",
  values = "both",
  dist = NULL,
  name = NULL,
  cell.type = NULL,
  sample.meta = NULL,
  color.by = NULL,
  shape.by = NULL,
  palette = NULL,
  show.sample.size = FALSE,
  sample.colors = NULL,
  color.title = NULL,
  shape.title = NULL,
  title = NULL,
  both.ncol = 2,
  both.align = "hv",
  n.permutations = 2000,
  show.pvalues = FALSE,
  sample.subset = NULL,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}

\item{\code{sample.subset}}{subset data for analysis only to the given samples}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateMetadataSeparation"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateMetadataSeparation}{}}}
\subsection{Method \code{estimateMetadataSeparation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateMetadataSeparation(
  sample.meta,
  space = "expression.shifts",
  dist = NULL,
  space.name = NULL,
  sample.subset = NULL,
  name = "metadata.separation",
  n.permutations = 5000,
  trim = 0.05,
  k = 20,
  show.warning = TRUE,
  verbose = self$verbose,
  n.cores = self$n.cores,
  adjust.pvalues = TRUE,
  p.adjust.method = "BH",
  pvalue.cutoff = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.subset}}{subset data for analysis only to the given samples}

\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values (default="BH")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotMetadataSeparation"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotMetadataSeparation}{}}}
\subsection{Method \code{plotMetadataSeparation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotMetadataSeparation(
  name = "metadata.separation",
  pvalue.y = 0.93,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateClusterFreeDE"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateClusterFreeDE}{}}}
\subsection{Method \code{estimateClusterFreeDE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeDE(
  n.top.genes = Inf,
  genes = NULL,
  max.z = 20,
  min.expr.frac = 0.01,
  min.n.samp.per.cond = 2,
  sample.per.cell = self$sample.per.cell,
  sample.metadata = self$sample.meta,
  formula = NULL,
  contrast = NULL,
  perm.method = "freedman-lane",
  robust.method = "none",
  na.mode = "drop",
  alternative = "two-sided",
  min.n.obs.per.samp = 2,
  adjust.pvalues = FALSE,
  keep.means = FALSE,
  robust = FALSE,
  smooth = TRUE,
  wins = 0.01,
  n.permutations = 200,
  lfc.pseudocount = 1e-05,
  block.vars = NULL,
  min.edge.weight = 0.6,
  verbose = self$verbose,
  n.cores = self$n.cores,
  name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getMostChangedGenes}{}}}
\subsection{Method \code{getMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getMostChangedGenes(
  n,
  method = c("z", "z.adj", "lfc"),
  min.z = 0.5,
  min.lfc = 1,
  max.score = 20,
  cell.subset = NULL,
  excluded.genes = NULL,
  included.genes = NULL,
  name = "cluster.free.de"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{excluded.genes}}{List of genes to exclude during estimation. For example, a list of mitochondrial genes.}

\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateClusterFreeExpressionShifts}{}}}
\subsection{Method \code{estimateClusterFreeExpressionShifts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeExpressionShifts(
  n.top.genes = 3000,
  gene.selection = "expression",
  name = "cluster.free.expr.shifts",
  min.n.between = 2,
  min.n.within = max(min.n.between, 1),
  dist.type = "shift",
  contrast = NULL,
  formula = NULL,
  pairContrast = NULL,
  pairFormula = NULL,
  min.expr.frac = 0,
  min.n.obs.per.samp = 3,
  perm.method = "freedman-lane",
  alternative = "two-sided",
  adjust = TRUE,
  smooth = TRUE,
  robust.method = "none",
  na.mode = "drop",
  dist = "cor",
  log.vectors = (dist != "js"),
  wins = 0.025,
  n.permutations = 999,
  verbose = self$verbose,
  n.cores = self$n.cores,
  genes = NULL,
  min.edge.weight = 0,
  sample.metadata = self$sample.meta,
  block.vars = self$block.vars,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{gene.selection}}{character string Method to select top genes, "z" selects genes by cluster-free Z-score change, "lfc" uses log2(fold-change) instead,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "z".}

\item{\code{name}}{string Field name where the test results are stored}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-smoothClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-smoothClusterFreeZScores}{}}}
\subsection{Method \code{smoothClusterFreeZScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$smoothClusterFreeZScores(
  n.top.genes = 1000,
  smoothing = 20,
  filter = NULL,
  z.adj = FALSE,
  gene.selection = ifelse(z.adj, "z.adj", "z"),
  excluded.genes = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{gene.selection}}{character string Method to select top genes, "z" selects genes by cluster-free Z-score change, "lfc" uses log2(fold-change) instead,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "z".}

\item{\code{excluded.genes}}{List of genes to exclude during estimation. For example, a list of mitochondrial genes.}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-estimateGenePrograms"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-estimateGenePrograms}{}}}
\subsection{Method \code{estimateGenePrograms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGenePrograms(
  method = c("pam", "leiden", "fabia"),
  n.top.genes = Inf,
  genes = NULL,
  n.programs = 15,
  z.adj = FALSE,
  gene.selection = ifelse(z.adj, "z.adj", "z"),
  smooth = TRUE,
  abs.scores = FALSE,
  name = "gene.programs",
  cell.subset = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  max.z = 5,
  min.z = 0.5,
  min.change.frac = 0.01,
  de.name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{gene.selection}}{character string Method to select top genes, "z" selects genes by cluster-free Z-score change, "lfc" uses log2(fold-change) instead,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "z".}

\item{\code{name}}{string Field name where the test results are stored}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotGeneProgramScores"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotGeneProgramScores}{}}}
\subsection{Method \code{plotGeneProgramScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgramScores(
  name = "gene.programs",
  prog.ids = NULL,
  build.panel = TRUE,
  nrow = NULL,
  adj.list = NULL,
  legend.title = "Score",
  palette = NULL,
  min.genes.per.prog = 10,
  color.range = c("0.5\%", "99.5\%"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotGeneProgramGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotGeneProgramGenes}{}}}
\subsection{Method \code{plotGeneProgramGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgramGenes(
  program.id,
  name = "gene.programs",
  ordering = c("similarity", "loading"),
  max.genes = 9,
  build.panel = TRUE,
  ncol = 3,
  plots = "z.adj",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotClusterFreeExpressionShifts}{}}}
\subsection{Method \code{plotClusterFreeExpressionShifts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotClusterFreeExpressionShifts(
  cell.groups = self$cell.groups,
  smooth = TRUE,
  plot.na = FALSE,
  adjusted = TRUE,
  name = "cluster.free.expr.shifts",
  scale.z.palette = TRUE,
  min.z = qnorm(0.9),
  color.range = c("0", "97.5\%"),
  alpha = 0.2,
  font.size = c(3, 5),
  adj.list = NULL,
  pal.seq = brewerPalette("YlOrRd", rev = FALSE),
  pal.div = brewerPalette("RdBu", rev = TRUE),
  build.panel = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell IDs as names (default: stored vector)}

\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotClusterFreeResiduals"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotClusterFreeResiduals}{}}}
\subsection{Method \code{plotClusterFreeResiduals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotClusterFreeResiduals(
  name = "cluster.free.expr.shifts",
  palette = brewerPalette("GnBu"),
  residual.type = c("pearson", "raw"),
  alpha = 0.2,
  font.size = c(3, 5),
  build.panel = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotMostChangedGenes}{}}}
\subsection{Method \code{plotMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotMostChangedGenes(
  n.top.genes,
  method = "z",
  min.z = 0.5,
  min.lfc = 1,
  max.score = 20,
  cell.subset = NULL,
  excluded.genes = NULL,
  build.panel = TRUE,
  ncol = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric Number of top genes for estimation}

\item{\code{excluded.genes}}{List of genes to exclude during estimation. For example, a list of mitochondrial genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-plotGeneExpressionComparison"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-plotGeneExpressionComparison}{}}}
\subsection{Method \code{plotGeneExpressionComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneExpressionComparison(
  genes = NULL,
  scores = NULL,
  max.expr = "97.5\%",
  plots = c("z.adj", "z", "expression"),
  min.z = qnorm(0.9),
  max.z = 4,
  max.z.adj = NULL,
  max.lfc = 3,
  smoothed = FALSE,
  gene.palette = dark.red.palette,
  z.palette = NULL,
  z.adj.palette = z.palette,
  lfc.palette = NULL,
  scale.z.palette = TRUE,
  plot.na = -1,
  adj.list = NULL,
  build.panel = TRUE,
  nrow = 1,
  ncol = 1,
  cell.subset = NULL,
  groups = NULL,
  subgroups = NULL,
  keep.limits = NULL,
  name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getConditionPerCell"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getConditionPerCell}{}}}
\subsection{Method \code{getConditionPerCell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getConditionPerCell()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getJointCountMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getJointCountMatrix}{}}}
\subsection{Method \code{getJointCountMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getJointCountMatrix(force = FALSE, raw = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-getGOEnvironment"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-getGOEnvironment}{}}}
\subsection{Method \code{getGOEnvironment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getGOEnvironment(org.db, verbose = FALSE, ignore.cache = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cacoa-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cacoa-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
