% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
Cacoa R6 class

Cacoa R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{number of cores}

\item{\code{verbose}}{print diagnostic messages}

\item{\code{test.results}}{results of the estimations, ready to use}

\item{\code{cache}}{intermediate results of the estimations, which can be used during some other computations}

\item{\code{data.object}}{the main object storing data (Conos or Seurat)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups per sample}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-estimateExpressionShiftZScores}{\code{Cacoa$estimateExpressionShiftZScores()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftZScores}{\code{Cacoa$plotExpressionShiftZScores()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  n.cores = parallel::detectCores(logical = F),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
Calculate expression shift magnitudes of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  groups = NULL,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  sample.groups = self$sample.groups,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "expression.shifts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{cell cluster factor}

\item{\code{dist}}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance}

\item{\code{within.group.normalization}}{normalize the shift magnitude by the mean magnitude of within-group variation}

\item{\code{valid.comparisons}}{a logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons}

\item{\code{n.cells}}{number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{\code{n.top.genes}}{number of top highest-expressed genes to consider (default: all genes)}

\item{\code{n.subsamples}}{number of samples to draw (default:100)}

\item{\code{min.cells}}{minimum number of cells per cluster/per sample to be included in the analysis}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared}

\item{\code{n.cores}}{number of cores to use}

\item{\code{verbose}}{}

\item{\code{ref.level}}{reference sample group, e.g., ctrl, healthy, or untreated.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list include \itemize{
  \item{df: a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells}
  \item{ctdml: raw list of `n.subsamples` sampled distance matrices (cells were subsampled)}
  \item{sample.groups: same as the provided variable}
  \item{valid.comparisons: a matrix of valid comparisons (in this case all should be valid, since we're not restricting samples that should be compared)}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftZScores"></a>}}
\subsection{Method \code{estimateExpressionShiftZScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftZScores(
  groups,
  ref.level,
  sample.groups = self$sample.groups,
  n.od.genes = 1000,
  n.pcs = 100,
  pca.maxit = 1000,
  ignore.cache = F,
  name = "expression.z.scores"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  size.norm = F,
  cluster.per.cell = NULL,
  sample.per.cell = NULL,
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{size.norm}}{Plot size normalized results. Requires cluster.per.cell, and sample.per.cell (default=F)}

\item{\code{cluster.per.cell}}{Named cluster factor with cell names (default=NULL)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default=NULL)}

\item{\code{label}}{Plot labels on size normalized plots (default=T)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftZScores"></a>}}
\subsection{Method \code{plotExpressionShiftZScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftZScores(
  type.order = NULL,
  name = "expression.z.scores"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
