% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
Cacoa R6 class

Cacoa R6 class

Plot inter-sample expression distance

Plot sample-sample expression distance as a 2D embedding
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{number of cores}

\item{\code{verbose}}{print diagnostic messages}

\item{\code{test.results}}{results of the estimations, ready to use}

\item{\code{cache}}{intermediate results of the estimations, which can be used during some other computations}

\item{\code{data.object}}{the main object storing data (Conos or Seurat)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample}

\item{\code{cell.groups}}{named factor with cell names with cluster per cell}

\item{\code{embedding}}{a 2D embedding to visualize the cells in}

\item{\code{sample.per.cell}}{named factor with cell names}

\item{\code{ref.level}}{reference level for sample.group vector}

\item{\code{target.level}}{target/disease level for sample.group vector}

\item{\code{sample.groups.palette}}{a color palette for the sample.groups}

\item{\code{cell.groups.palette}}{a color palette for the cell.groups}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-estimateCommonExpressionShiftMagnitudes}{\code{Cacoa$estimateCommonExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-plotCommonExpressionShiftMagnitudes}{\code{Cacoa$plotCommonExpressionShiftMagnitudes()}}
\item \href{#method-estimatePerCellTypeDE}{\code{Cacoa$estimatePerCellTypeDE()}}
\item \href{#method-estimatePerCellTypeDEnew}{\code{Cacoa$estimatePerCellTypeDEnew()}}
\item \href{#method-plotDEGenes}{\code{Cacoa$plotDEGenes()}}
\item \href{#method-plotNumberOfDEGenes}{\code{Cacoa$plotNumberOfDEGenes()}}
\item \href{#method-saveDEasJSON}{\code{Cacoa$saveDEasJSON()}}
\item \href{#method-plotFilteredDEGenes}{\code{Cacoa$plotFilteredDEGenes()}}
\item \href{#method-prepareOntologyData}{\code{Cacoa$prepareOntologyData()}}
\item \href{#method-plotEmbedding}{\code{Cacoa$plotEmbedding()}}
\item \href{#method-estimateOntology}{\code{Cacoa$estimateOntology()}}
\item \href{#method-estimateOntologyFamilies}{\code{Cacoa$estimateOntologyFamilies()}}
\item \href{#method-estimateOntologyClusters}{\code{Cacoa$estimateOntologyClusters()}}
\item \href{#method-plotOntologyDistribution}{\code{Cacoa$plotOntologyDistribution()}}
\item \href{#method-plotOntologyTerms}{\code{Cacoa$plotOntologyTerms()}}
\item \href{#method-plotOntology}{\code{Cacoa$plotOntology()}}
\item \href{#method-plotOntologyHeatmap}{\code{Cacoa$plotOntologyHeatmap()}}
\item \href{#method-plotOntologySimilarities}{\code{Cacoa$plotOntologySimilarities()}}
\item \href{#method-plotOntologyFamily}{\code{Cacoa$plotOntologyFamily()}}
\item \href{#method-plotProportions}{\code{Cacoa$plotProportions()}}
\item \href{#method-plotCellNumbers}{\code{Cacoa$plotCellNumbers()}}
\item \href{#method-plotPcaSpace}{\code{Cacoa$plotPcaSpace()}}
\item \href{#method-plotCdaSpace}{\code{Cacoa$plotCdaSpace()}}
\item \href{#method-plotContrastTree}{\code{Cacoa$plotContrastTree()}}
\item \href{#method-estimateCellLoadings}{\code{Cacoa$estimateCellLoadings()}}
\item \href{#method-estimateGaPartiotion}{\code{Cacoa$estimateGaPartiotion()}}
\item \href{#method-plotCellLoadings}{\code{Cacoa$plotCellLoadings()}}
\item \href{#method-extimateWilcoxonTest}{\code{Cacoa$extimateWilcoxonTest()}}
\item \href{#method-estimateCellDensity}{\code{Cacoa$estimateCellDensity()}}
\item \href{#method-plotCellDensity}{\code{Cacoa$plotCellDensity()}}
\item \href{#method-diffCellDensity}{\code{Cacoa$diffCellDensity()}}
\item \href{#method-plotExpressionDistance}{\code{Cacoa$plotExpressionDistance()}}
\item \href{#method-plotExpressionDistanceEmbedding}{\code{Cacoa$plotExpressionDistanceEmbedding()}}
\item \href{#method-getContour}{\code{Cacoa$getContour()}}
\item \href{#method-estimateClusterFreeZScores}{\code{Cacoa$estimateClusterFreeZScores()}}
\item \href{#method-getMostChangedGenes}{\code{Cacoa$getMostChangedGenes()}}
\item \href{#method-estimateClusterFreeExpressionShifts}{\code{Cacoa$estimateClusterFreeExpressionShifts()}}
\item \href{#method-smoothClusterFreeZScores}{\code{Cacoa$smoothClusterFreeZScores()}}
\item \href{#method-estimateGeneProgrammes}{\code{Cacoa$estimateGeneProgrammes()}}
\item \href{#method-plotGeneProgrammeScores}{\code{Cacoa$plotGeneProgrammeScores()}}
\item \href{#method-plotGeneProgrammeGenes}{\code{Cacoa$plotGeneProgrammeGenes()}}
\item \href{#method-plotClusterFreeExpressionShifts}{\code{Cacoa$plotClusterFreeExpressionShifts()}}
\item \href{#method-plotMostChangedGenes}{\code{Cacoa$plotMostChangedGenes()}}
\item \href{#method-plotGeneExpressionComparison}{\code{Cacoa$plotGeneExpressionComparison()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  ref.level = NULL,
  target.level = NULL,
  sample.groups.palette = NULL,
  cell.groups.palette = NULL,
  embedding = extractEmbedding(data.object),
  n.cores = 1,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
Calculate expression shift magnitudes of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  sample.groups = self$sample.groups,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "expression.shifts",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{dist}}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance (default="JS")}

\item{\code{within.group.normalization}}{Normalize the shift magnitude by the mean magnitude of within-group variation (default=T)}

\item{\code{valid.comparisons}}{A logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons (default=NULL)}

\item{\code{n.cells}}{Number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{\code{n.top.genes}}{Number of top highest-expressed genes to consider (default: all genes)}

\item{\code{n.subsamples}}{Number of samples to draw (default=100)}

\item{\code{min.cells}}{Minimum number of cells per cluster/per sample to be included in the analysis (default=10)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{Number of cores (default: stored integer)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{Test name (default="expression.shifts")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list include \itemize{
\item{df: a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells}
\item{ctdml: raw list of \code{n.subsamples} sampled distance matrices (cells were subsampled)}
\item{sample.groups: same as the provided variable}
\item{valid.comparisons: a matrix of valid comparisons (in this case all should be valid, since we're not restricting samples that should be compared)}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCommonExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCommonExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateCommonExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCommonExpressionShiftMagnitudes(
  sample.groups = self$sample.groups,
  cell.groups = self$cell.groups,
  n.cells = NULL,
  n.randomizations = 50,
  n.subsamples = 30,
  min.cells = 10,
  n.cores = self$n.cores,
  verbose = self$verbose,
  mean.trim = 0.1,
  name = "common.expression.shifts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  type = "box",
  notch = TRUE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.size.dependency = FALSE,
  show.whiskers = TRUE,
  show.regression = TRUE,
  font.size = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\itemize{
\item results slot name (default: 'expression.shifts')
}}

\item{\code{type}}{\itemize{
\item type of a plot "bar" (default) or "box"
}}

\item{\code{notch}}{\itemize{
\item whether to show notches in the boxplot version (default=TRUE)
}}

\item{\code{show.jitter}}{whether to show indiivudal data points (default: FALSE)}

\item{\code{jitter.alpha}}{transparency value for the data points (default: 0.05)}

\item{\code{show.regression}}{whether to show a slope line in the size dependency plot}

\item{\code{show.regression}}{whether to show a whiskers in the size dependency plot}

\item{\code{font.size}}{font size for the cell type labels in the size dependency plot}

\item{\code{show.size.depenency}}{whether to show mean vs. number of cells in a cell type instead}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot common expression shift estimates across cell types
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCommonExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotCommonExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotCommonExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCommonExpressionShiftMagnitudes(
  name = "common.expression.shifts",
  show.subsampling.variability = FALSE,
  show.jitter = FALSE,
  jitter.alpha = 0.05,
  type = "bar",
  notch = TRUE,
  show.size.dependency = FALSE,
  show.whiskers = TRUE,
  show.regression = TRUE,
  font.size = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{result slot name (default: common.expression.shifts}

\item{\code{show.subsampling.variability}}{\itemize{
\item whether the spread should illustrate subsampling variability instead of the inter-sample variability (default: FALSE)
}}

\item{\code{show.jitter}}{whether to show indiivudal data points (default: FALSE)}

\item{\code{jitter.alpha}}{transparency value for the data points (default: 0.05)}

\item{\code{type}}{\itemize{
\item type of a plot "bar" (default) or "box"
}}

\item{\code{notch}}{\itemize{
\item whether to show notches in the boxplot version (default=TRUE)
}}

\item{\code{show.regression}}{whether to show a slope line in the size dependency plot}

\item{\code{show.regression}}{whether to show a whiskers in the size dependency plot}

\item{\code{font.size}}{font size for the cell type labels in the size dependency plot}

\item{\code{show.size.depenency}}{whether to show mean vs. number of cells in a cell type instead}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePerCellTypeDE}{}}}
\subsection{Method \code{estimatePerCellTypeDE()}}{
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDE(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  common.genes = FALSE,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  max.cell.count = Inf,
  test = "Wald",
  independent.filtering = FALSE,
  cluster.sep.chr = "<!!>",
  return.matrix = T,
  verbose = self$verbose,
  name = "de"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=NULL)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{common.genes}}{Only investigate common genes across cell groups (default=F)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=F)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=FALSE)}

\item{\code{min.cell.count}}{minimum number of cells that need to be present in a given cell type in a given sample in order to be taken into account (default=10)}

\item{\code{max.cell.count}}{maximal number of cells per cluster per sample to include in a comparison (useful for comparing the number of DE genes between cell types) (default: Inf)}

\item{\code{test}}{which DESeq2 test to use (options: "LRT" (default), "Wald")}

\item{\code{independent.filtering}}{independentFiltering parameter for DESeq2 (default=FALSE)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{return.matrix}}{Return merged matrix of results (default=TRUE)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{slot in which to save the results (default: 'de')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePerCellTypeDEnew"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePerCellTypeDEnew}{}}}
\subsection{Method \code{estimatePerCellTypeDEnew()}}{
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDEnew(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  target.level = self$target.level,
  common.genes = FALSE,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  max.cell.count = Inf,
  independent.filtering = FALSE,
  cluster.sep.chr = "<!!>",
  verbose = self$verbose,
  name = "de",
  test = "DESeq2.Wald",
  resampling.method = NULL,
  max.resamplings = 30,
  seed.resampling = 239,
  covariates = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=NULL)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{common.genes}}{Only investigate common genes across cell groups (default=F)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=F)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=FALSE)}

\item{\code{min.cell.count}}{minimum number of cells that need to be present in a given cell type in a given sample in order to be taken into account (default=10)}

\item{\code{max.cell.count}}{maximal number of cells per cluster per sample to include in a comparison (useful for comparing the number of DE genes between cell types) (default: Inf)}

\item{\code{independent.filtering}}{independentFiltering parameter for DESeq2 (default=FALSE)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{slot in which to save the results (default: 'de')}

\item{\code{test}}{which DESeq2 test to use (options: "LRT" (default), "Wald")}

\item{\code{resampling.method}}{which resampling method should be used "loo" for leave-one-out or "bootstrap", (default:NULL no resampling)}

\item{\code{return.matrix}}{Return merged matrix of results (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEGenes}{}}}
\subsection{Method \code{plotDEGenes()}}{
Plot number of significant DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEGenes(
  name = "de",
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = TRUE,
  p.adj = 0.05,
  size = 4,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{results slot in which the DE results should be stored (default: 'de')}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{palette}}{cell group palette (default: stored $cell.groups.palette)}

\item{\code{p.adjust.cutoff}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNumberOfDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotNumberOfDEGenes}{}}}
\subsection{Method \code{plotNumberOfDEGenes()}}{
Plot number of significant DE genes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumberOfDEGenes(
  name = "de",
  legend.position = "none",
  label = TRUE,
  p.adjust = TRUE,
  pvalue.cutoff = 0.05,
  show.resampling.results = TRUE,
  show.jitter = FALSE,
  jitter.alpha = 0.05,
  type = "bar",
  notch = TRUE,
  show.size.dependency = FALSE,
  show.whiskers = TRUE,
  show.regression = TRUE,
  font.size = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{results slot in which the DE results should be stored (default: 'de')}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{p.adjust}}{whether the cutoff should be based on the adjusted P value (default: TRUE)}

\item{\code{pvalue.cutoff}}{P value cutoff (default=0.05)}

\item{\code{show.resampling.results}}{whether to show uncertainty based on resampling results (default: TRUE)}

\item{\code{show.whiskers}}{whether to show a whiskers in the size dependency plot (default:TRUE)}

\item{\code{show.regression}}{whether to show a slope line in the size dependency plot (default:TRUE)}

\item{\code{font.size}}{font size for the cell type labels in the size dependency plot}

\item{\code{palette}}{cell group palette (default: stored $cell.groups.palette)}

\item{\code{show.size.depenency}}{whether to show mean vs. number of cells in a cell type instead (default: FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDEasJSON"></a>}}
\if{latex}{\out{\hypertarget{method-saveDEasJSON}{}}}
\subsection{Method \code{saveDEasJSON()}}{
Save DE results as JSON files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveDEasJSON(
  saveprefix = NULL,
  dir.name = "JSON",
  de.raw = NULL,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  gene.metadata = NULL,
  cluster.sep.chr = "<!!>",
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{saveprefix}}{Prefix for created files (default=NULL)}

\item{\code{dir.name}}{Name for directory with results (default="JSON")}

\item{\code{de.raw}}{List of DE results}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{gene.metadata}}{(default=NULL)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFilteredDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotFilteredDEGenes}{}}}
\subsection{Method \code{plotFilteredDEGenes()}}{
Plot number of highly-expressed DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotFilteredDEGenes(
  de.filter = self$test.results$ontology$de.filter,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepareOntologyData"></a>}}
\if{latex}{\out{\hypertarget{method-prepareOntologyData}{}}}
\subsection{Method \code{prepareOntologyData()}}{
Filter and prepare DE genes for ontology calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$prepareOntologyData(
  org.db,
  p.adj = 1,
  expr.cutoff = 0.05,
  de.raw = NULL,
  cell.groups = self$cell.groups,
  universe = NULL,
  transposed = TRUE,
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{p.adj}}{Cutoff for filtering highly-expressed DE genes (default=0.05)}

\item{\code{expr.cutoff}}{Cutoff for cells per group expressing a DE gene, i.e., cutoff for highly-expressed genes (default=0.05)}

\item{\code{de.raw}}{Differentially expressed genes per cell group, results from estimatePerCellTypeDE (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{universe}}{Only set this if a common background gene set is desired for all cell groups (default: NULL)}

\item{\code{transposed}}{Whether count matrices should be transposed (default=T)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{n.top.genes}}{Number of most different genes to take as input. If less are left after filtering for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-plotEmbedding}{}}}
\subsection{Method \code{plotEmbedding()}}{
Plot embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotEmbedding(
  embedding = self$embedding,
  plot.theme = ggplot2::theme_bw(),
  show.legend = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{embedding}}{A cell embedding to use (two-column data frame with rownames corresponding to cells) (default: stored embedding object)}

\item{\code{plot.theme}}{plot theme to use (default: ggplot2::theme_bw())}

\item{\code{...}}{other parameters are passed to \link[sccore:embeddingPlot]{embeddingPlot}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntology"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntology}{}}}
\subsection{Method \code{estimateOntology()}}{
Estimate ontology terms based on DEs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntology(
  type = "GO",
  org.db,
  n.top.genes = 500,
  de.gene.ids = NULL,
  go.environment = self$test.results$GO$go.environment,
  p.adjust.method = "BH",
  readable = TRUE,
  verbose = TRUE,
  qvalue.cutoff = 0.2,
  min.gs.size = 10,
  max.gs.size = 500,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information (default="GO")}

\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{de.gene.ids}}{List containing DE gene IDs, and filtered DE genes (default: stored list, results from prepareOntologyData)}

\item{\code{go.environment}}{Extracted GO environment. If set to NULL, the environment will be re-extracted (default: stored environment)}

\item{\code{p.adjust.method}}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{\code{readable}}{Mapping gene ID to gene name (default=T)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{qvalue.cutoff}}{Q value cutoff, please see clusterProfiler package for more information (default=0.2)}

\item{\code{min.gs.size}}{Minimal geneset size, please see clusterProfiler package for more information (default=5)}

\item{\code{max.gs.size}}{Minimal geneset size, please see clusterProfiler package for more information (default=5e2)}

\item{\code{min.genes}}{Minimum number of input genes overlapping with ontologies (default=0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing a list of terms per ontology, and a data frame with merged results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntologyFamilies"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntologyFamilies}{}}}
\subsection{Method \code{estimateOntologyFamilies()}}{
Estimate ontology families
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntologyFamilies(type = "GO", p.adj = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntologyClusters"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntologyClusters}{}}}
\subsection{Method \code{estimateOntologyClusters()}}{
Estimate Gene Ontology clusters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntologyClusters(
  type = "GO",
  subtype = NULL,
  genes = "all",
  ind.h = 0.66,
  total.h = 0.5,
  verbose = self$verbose,
  p.adj = p.adj,
  min.genes = min.genes,
  name = getOntClustField(type, subtype, genes)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types), "GO" or "DO". Default: "GO".}

\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all'. Default: "up".}

\item{\code{ind.h}}{Cut height for hierarchical clustering of terms per cell type.
Approximately equal to the fraction of genes, shared between the GOs. Default: 0.66.}

\item{\code{total.h}}{Cut height for hierarchical clustering of GOs across all subtypes.
Approximately equal to the fraction of subtypes for which two GOs should belong to the same cluster. Default: 0.5.}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{Name of the field to store the results. Default: \code{cacoa:::getOntClustField(type, subtype, genes)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing:
\itemize{
\item \code{df}: data.frame with information about individual gene ontolodies and columns \code{Cluster} and \code{ClusterName} for the clustering info
\item \code{hclust}: the object of class \link[stats:hclust]{hclust} with hierarchical clustering of GOs across all subtypes
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyDistribution}{}}}
\subsection{Method \code{plotOntologyDistribution()}}{
Bar plot of ontology terms per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyDistribution(
  genes = "all",
  type = "GO",
  p.adj = 0.05,
  min.genes = 1,
  cell.groups = self$cell.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="all")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyTerms"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyTerms}{}}}
\subsection{Method \code{plotOntologyTerms()}}{
Plot ontology terms as a function of both number of DE genes, and number of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyTerms(
  genes = "all",
  type = "GO",
  p.adj = 0.05,
  min.genes = 1,
  de.filter = self$test.results$gene.ids,
  cell.groups = self$cell.groups,
  label.x.pos = 0.01,
  label.y.pos = 1,
  scale = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default='all')}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{label.x.pos}}{Plot label position on x axis (default=0.01)}

\item{\code{label.y.pos}}{Plot label position on y axis (default=1)}

\item{\code{scale}}{Scaling of plots, adjust if e.g. label is misplaced. See \link[cowplot:plot_grid]{cowplot::plot_grid} for more info (default=1.0)}

\item{\code{...}}{parameters forwarded to \link{plotNCellRegression}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntology"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntology}{}}}
\subsection{Method \code{plotOntology()}}{
Plot a dotplot of ontology terms with adj. P values for a specific cell subgroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntology(
  cell.subgroups,
  plot = "dot",
  genes = "up",
  type = "GO",
  subtype = "BP",
  n = 20,
  p.adj = 0.05,
  min.genes = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types), "GO" or "DO" (default="GO")}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}

\item{\code{cell.subgroup}}{Cell group to plot (default=NULL)}

\item{\code{log.colors}}{Use log10 p-values for coloring (default=FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyHeatmap}{}}}
\subsection{Method \code{plotOntologyHeatmap()}}{
Plot a heatmap of ontology P values per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmap(
  genes = "up",
  type = "GO",
  subtype = "BP",
  min.genes = 1,
  p.adj = 0.05,
  legend.position = "left",
  selection = "all",
  n = 10,
  clusters = TRUE,
  cluster.name = NULL,
  cell.subgroups = NULL,
  color.range = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down' for downregulated genes, 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{selection}}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type); 'common' (only show terms that are present in at least two cell types); 'all' (all ontology terms) (default="all")}

\item{\code{n}}{Number of terms to show (default=10)}

\item{\code{clusters}}{Whether to show GO clusters or raw GOs (default=TRUE)}

\item{\code{cluster.name}}{Field with the results for GO clustering. Ignored if \code{clusters == FALSE}.}

\item{\code{cell.subgroups}}{Cell groups to plot (default=NULL)}

\item{\code{color.range}}{vector with two values for min/max values of p-values}

\item{\code{...}}{parameters forwarded to \link{plotHeatmap}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologySimilarities"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologySimilarities}{}}}
\subsection{Method \code{plotOntologySimilarities()}}{
Plot correlation matrix for ontology terms between cell types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologySimilarities(
  genes = "up",
  type = "GO",
  p.adj = 0.05,
  min.genes = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
#' @description Plot ontology families heatmap
#' @return A ggplot2 object
plotOntologyFamilyHeatmap=function(genes = "up", type = "GO", subtype = "BP", min.genes = 1, selection = "unique", n = 20, legend.position = "right") {Checks

if(!is.null(cell.subgroups) && (length(cell.subgroups) == 1))
stop("'cell.subgroups' must contain at least two groups. Please use plotOntology instead.")

if(is.null(selection) || (!selection \%in\% c("unique","common","all")))
stop("'selection' must be one of the following: 'unique', 'common', or 'all'.")Extract results

if (!clusters) {
ont.sum <- private$getOntologyPvalueResults(genes=genes, type=type, subtype=subtype, cell.subgroups=cell.subgroups,
p.adj=p.adj, min.genes=min.genes) \%>\%
groupOntologiesByCluster(field="Description")
} else {
name <- if (is.null(cluster.name)) getOntClustField(type, subtype, genes) else cluster.name
if (is.null(self$test.results[\link{name}])) {
if (!is.null(cluster.name))
stop("Can't find the results for ", cluster.name) # stop if user specified a wrong cluster.name\preformatted{  warning("Can't find the results for ", name, ". Running estimateOntologyClusters()...\\n")
  ont.sum <- self$estimateOntologyClusters(type=type, genes=genes, name=name, cell.subgroups=cell.subgroups,
                                           p.adj=p.adj, min.genes=min.genes)$df
\} else \{
  ont.sum <- self$test.results[[name]]$df
\}

ont.sum \%<>\% groupOntologiesByCluster(field="ClusterName")
}

}

if(selection=="unique") {
ont.sum \%<>\% .\link{rowSums(. > 0) == 1,}
} else if(selection=="common") {
ont.sum \%<>\% .\link{rowSums(. > 0) > 1,}
}
if(nrow(ont.sum) == 0) stop("Nothing to plot. Try another selection.")Plot

gg <- ont.sum \%>\%
.\link{, colSums(abs(.)) > 0} \%>\%
.[match(rowSums(.)\link{rowSums(abs(.)) > 0} \%>\% .\link{order(., decreasing=TRUE)} \%>\% names, rownames(.)),] \%>\%
tail(n) \%>\%
plotHeatmap(legend.position=legend.position, row.order=TRUE, color.range=color.range, ...) +
getGeneScale(genes=genes, type="fill", high="white", limits=color.range)

return(gg)
},
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyFamily"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyFamily}{}}}
\subsection{Method \code{plotOntologyFamily()}}{
Plot ontology family tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyFamily(
  type = "GO",
  cell.subgroups,
  family,
  genes = "up",
  subtype = "BP",
  plot.type = "complete",
  show.ids = FALSE,
  string.length = 18,
  legend.label.size = 1,
  legend.position = "topright",
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{n.cores}}{number of cores for parallelisation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Rgraphviz object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotProportions"></a>}}
\if{latex}{\out{\hypertarget{method-plotProportions}{}}}
\subsection{Method \code{plotProportions()}}{
Plot the cell group proportions per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotProportions(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  notch = FALSE,
  alpha = 0.2,
  palette = self$sample.groups.palette,
  show.significance = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{cells.to.remove}}{Vector of cell types to remove from the composition}

\item{\code{cells.to.remain}}{Vector of cell types to remain in the composition}

\item{\code{notch}}{Whether to show notch in the boxplots}

\item{\code{alpha}}{Transparency level on the data points (default: 0.2)}

\item{\code{palette}}{color palette to use for conditions (default: stored $sample.groups.palette)}

\item{\code{show.significance}}{whether to show statistical significance betwwen sample groups. wilcox.test was used; (* < 0.05; ** < 0.01; *** < 0.001)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellNumbers"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellNumbers}{}}}
\subsection{Method \code{plotCellNumbers()}}{
Plot the cell numbers per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellNumbers(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPcaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-plotPcaSpace}{}}}
\subsection{Method \code{plotPcaSpace()}}{
Plot compositions in CoDA-PCA space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotPcaSpace(
  cells.to.remove = NULL,
  font.size = NULL,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCdaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-plotCdaSpace}{}}}
\subsection{Method \code{plotCdaSpace()}}{
Plot compositions in CoDA-CDA space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCdaSpace(
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  font.size = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotContrastTree"></a>}}
\if{latex}{\out{\hypertarget{method-plotContrastTree}{}}}
\subsection{Method \code{plotContrastTree()}}{
Plot contrast tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotContrastTree(cells.to.remain = NULL, cells.to.remove = NULL)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellLoadings}{}}}
\subsection{Method \code{estimateCellLoadings()}}{
Plot Loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellLoadings(
  n.cell.counts = 1000,
  n.seed = 239,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  samples.to.remove = NULL,
  n.iter = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGaPartiotion"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGaPartiotion}{}}}
\subsection{Method \code{estimateGaPartiotion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGaPartiotion(
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellLoadings}{}}}
\subsection{Method \code{plotCellLoadings()}}{
Plot Loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellLoadings(
  alpha = 0.01,
  palette = self$cell.groups.palette,
  font.size = NULL,
  ordering = "by.pvalue",
  signif.threshold = 0.05,
  show.pvals = F
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{palette}}{palette specification for cell types (default: stored $cell.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extimateWilcoxonTest"></a>}}
\if{latex}{\out{\hypertarget{method-extimateWilcoxonTest}{}}}
\subsection{Method \code{extimateWilcoxonTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$extimateWilcoxonTest(
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups,
  cells.to.remove = NULL,
  cells.to.remain = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellDensity}{}}}
\subsection{Method \code{estimateCellDensity()}}{
Estimate cell density in giving embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellDensity(
  embedding = self$embedding,
  bins = 400,
  by.sample = TRUE,
  method = "kde",
  verbose = self$verbose,
  m = 50,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{number of bins for density estimation, default 400}

\item{\code{by.sample}}{if TRUE, density will estimated by sample and quantile normalization will applied to individual sample. If FALSE, cell condition.per.cell need to be provided and density will simply estimated by condition.per.cell.}

\item{\code{method}}{density estimation method, graphSmooth: graph smooth based density estimation. embGrid: embedding grid based density  estimation. (default: embGrid)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{m}}{numeric Maximum order of Chebyshev coeff to compute (default=50)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{emb}}{cell embedding matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellDensity}{}}}
\subsection{Method \code{plotCellDensity()}}{
Plot cell density
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensity(
  col = c("blue", "white", "red"),
  show.legend = FALSE,
  legend.position = NULL,
  show.grid = NULL,
  add.points = TRUE,
  point.col = "#FCFDBFFF",
  contours = NULL,
  contour.color = "white",
  contour.conf = "10\%"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add.points}}{default is TRUE, add points to cell density figure}

\item{\code{contours}}{specify cell types for contour, multiple cell types are also supported}

\item{\code{contour.color}}{color for contour line}

\item{\code{contour.conf}}{confidence interval of contour}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-diffCellDensity}{}}}
\subsection{Method \code{diffCellDensity()}}{
estimate differential cell density
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$diffCellDensity(
  condition.per.cell = NULL,
  method = "subtract",
  col = c("blue", "white", "red"),
  show.legend = FALSE,
  legend.position = NULL,
  title = NULL,
  show.grid = NULL,
  plot = TRUE,
  contours = NULL,
  contour.color = "white",
  contour.conf = "10\%",
  z.cutoff = NULL,
  size = 0.1,
  adjust.pvalues = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition.per.cell}}{A two-level factor on the cell names describing the conditions being compared (default: stored vector)}

\item{\code{method}}{method to calculate differential cell density of each bin; subtract: target density minus ref density; entropy: estimated kl divergence entropy between sample groups ;
t.test: z-score of t-test, global variance is setting for t.test;}

\item{\code{col}}{color palettes,  default is c('blue','white','red')}

\item{\code{contours}}{specify cell types for contour, multiple cell types are also supported}

\item{\code{contour.color}}{color for contour line}

\item{\code{contour.conf}}{confidence interval of contour}

\item{\code{z.cutoff}}{absolute z score cutoff}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionDistance"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionDistance}{}}}
\subsection{Method \code{plotExpressionDistance()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistance(
  name = "expression.shifts",
  notch = TRUE,
  sample.groups = self$sample.groups,
  weighted.distance = TRUE,
  min.cells = 10,
  palette = self$sample.groups.palette,
  show.significance = FALSE,
  alpha = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{notch}}{Show notches in plot, see ggplot2::geom_boxplot for more info (default=T)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{weighted.distance}}{whether to weigh the expression distance by the sizes of cell types (default: TRUE), or show distances for each individual cell type}

\item{\code{show.significance}}{whether to show statistical significance between sample groups. wilcox.test was used; (\* < 0.05; \\emph{\} < 0.01; \\emph{\}\* < 0.001)}

\item{\code{alpha}}{dot transparency}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored $cell.groups vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionDistanceEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionDistanceEmbedding}{}}}
\subsection{Method \code{plotExpressionDistanceEmbedding()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistanceEmbedding(
  name = "expression.shifts",
  sample.groups = self$sample.groups,
  cell.type = NULL,
  method = "tSNE",
  perplexity = 4,
  max_iter = 1000,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{cell.type}}{If a name of a cell type is specified, the sample distances will be assessed based on this cell type alone. Otherwise (cell.type=NULL, default), sample distances will be estimated as an average distance across all cell types (weighted by the minimum number of cells of that cell type between any two samples being compared)}

\item{\code{method}}{dimension reduction methods (MDS or tSNE ) , default is MDS}

\item{\code{perplexity}}{tSNE perpexity (default: 4)}

\item{\code{max_iter}}{tSNE max_iter (default: 1e3)}

\item{\code{palette}}{a set of colors to use for conditions (default: stored $sample.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getContour"></a>}}
\if{latex}{\out{\hypertarget{method-getContour}{}}}
\subsection{Method \code{getContour()}}{
Extract contour from embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getContour(cells, color = "white", linetype = 2, conf = "10\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf}}{confidence interval of contour}

\item{\code{cell}}{specify cell types for contour, mutiple cell types are also suported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeZScores}{}}}
\subsection{Method \code{estimateClusterFreeZScores()}}{
Estimate differential expression Z-scores between two conditions per individual cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeZScores(
  n.top.genes = NULL,
  max.z = 20,
  min.expr.frac = 0.001,
  normalize = c("ref", "both"),
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{max.z}}{z-score value to winsorize the estimates for reducing impact of outliers. Default: 20.}

\item{\code{min.expr.frac}}{minimal fraction of cell expressing a gene for estimating z-scores for it. Default: 0.001.}

\item{\code{normalize}}{whether to normalize z-scores over std for reference ("ref") or both ("both"). Default: "ref".}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{n.cores}}{number of cores for parallelisation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sparse matrix of z-scores with genes as columns and cells as rows.
Cells that have only one condition in their expression neighborhood have NA Z-scores for all genes.
Results are also stored in the \code{cluster.free.z} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMostChangedGenes}{}}}
\subsection{Method \code{getMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getMostChangedGenes(
  n,
  min.z = 0.5,
  max.z = 20,
  cell.subset = NULL,
  excluded.genes = NULL,
  included.genes = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{excluded.genes}}{list of genes to exclude during estimation. For example, a list of mitochondrial genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeExpressionShifts}{}}}
\subsection{Method \code{estimateClusterFreeExpressionShifts()}}{
Estimate Cluster-free Expression Shift
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeExpressionShifts(
  n.top.genes = 3000,
  gene.selection = "change",
  excluded.genes = NULL,
  verbose = self$verbose,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{gene.selection}}{a method to select top genes, "change" selects genes by cluster-free Z-score change,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "change".}

\item{\code{excluded.genes}}{list of genes to exclude during estimation. For example, a list of mitochondrial genes.}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{n.cores}}{number of cores for parallelisation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of cluster-free expression shifts per cell. Values above 1 correspond to difference between conditions.
Results are also stored in the \code{cluster.free.expr.shifts} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-smoothClusterFreeZScores}{}}}
\subsection{Method \code{smoothClusterFreeZScores()}}{
Performs graph smoothing of the cluster-free DE Z-scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$smoothClusterFreeZScores(
  n.top.genes = 1000,
  smoothing = 20,
  filter = NULL,
  gene.selection = "change",
  excluded.genes = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{smoothing}}{\code{beta} parameter of the \link[sccore:heatFilter]{heatFilter}. Default: 20.}

\item{\code{filter}}{graph filter function. Default: \link[sccore:heatFilter]{heatFilter}.}

\item{\code{gene.selection}}{a method to select top genes, "change" selects genes by cluster-free Z-score change,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "change".}

\item{\code{excluded.genes}}{list of genes to exclude during estimation. For example, a list of mitochondrial genes.}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{...}}{parameters forwarded to \link[sccore:smoothSignalOnGraph]{smoothSignalOnGraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sparse matrix of smoothed Z-scores. Results are also stored in the \code{cluster.free.z.smoothed} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGeneProgrammes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGeneProgrammes}{}}}
\subsection{Method \code{estimateGeneProgrammes()}}{
Estimate Gene Programmes based on cluster-free Z-scores on a subsample of
cells using \link[fabia:fabia]{fabia}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGeneProgrammes(
  n.top.genes = NULL,
  n.programmes = 15,
  gene.selection = "change",
  name = "gene.programmes",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{n.programmes}}{maximal number of gene programmes to find (parameter \code{p} for fabia). Default: 15.}

\item{\code{gene.selection}}{a method to select top genes, "change" selects genes by cluster-free Z-score change,
"expression" picks the most expressed genes and "od" picks overdispersed genes.  Default: "change".}

\item{\code{name}}{field name where the test results are stored}

\item{\code{...}}{keyword arguments forwarded to \link{estimateGeneProgrammes}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list includes:
\itemize{
\item \code{fabia}: \link[fabia:Factorization]{fabia::Factorization} object, result of the
\link[fabia:fabia]{fabia::fabia} call
\item \code{sample.ids}: ids of the subsampled cells used for fabia estimates
\item \code{scores.exact}: vector of fabia estimates of gene programme scores per cell. Estimated only for the
subsampled cells.
\item \code{scores.approx}: vector of approximate gene programme scores, estimated for all cells in the dataset
\item \code{loadings}: matrix with fabia gene loadings per programme
\item \code{gene.scores}: list of vectors of gene scores per programme. Contains only genes, selected for
the programme usin fabia biclustering.
\item \code{bi.clusts} fabia biclustering information, result of the \link[fabia:extractBic]{fabia::extractBic} call
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgrammeScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgrammeScores}{}}}
\subsection{Method \code{plotGeneProgrammeScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgrammeScores(
  name = "gene.programmes",
  approximate = FALSE,
  build.panel = TRUE,
  nrow = NULL,
  gradient.range.quantile = 0.975,
  plot.na = approximate,
  legend.title = "Score",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgrammeGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgrammeGenes}{}}}
\subsection{Method \code{plotGeneProgrammeGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgrammeGenes(
  programme.id,
  name = "gene.programmes",
  max.genes = 9,
  plot.expression = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-plotClusterFreeExpressionShifts}{}}}
\subsection{Method \code{plotClusterFreeExpressionShifts()}}{
Plot cluster-free expression shift z-scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotClusterFreeExpressionShifts(
  cell.groups = self$cell.groups,
  plot.both.conditions = FALSE,
  plot.na = FALSE,
  max.shift = NULL,
  alpha = 0.2,
  font.size = c(3, 5),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{cell type labels. Set to NULL if it shouldn't be shown}

\item{\code{plot.both.conditions}}{show both case and control cells. Normally, showing control cells doesn't
make sense, as control cells always have small distance from control.}

\item{\code{max.shift}}{all shift values above \code{max.shift} are set to this value when plotting. Default: 95\% of the shifts.}

\item{\code{font.size}}{size range for cell type labels}

\item{\code{...}}{parameters forwarded to \link[sccore:embeddingPlot]{embeddingPlot}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotMostChangedGenes}{}}}
\subsection{Method \code{plotMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotMostChangedGenes(
  n.top.genes,
  min.z = 0.5,
  max.z = 20,
  max.z.plot = max.z,
  cell.subset = NULL,
  excluded.genes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for estimation}

\item{\code{excluded.genes}}{list of genes to exclude during estimation. For example, a list of mitochondrial genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneExpressionComparison"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneExpressionComparison}{}}}
\subsection{Method \code{plotGeneExpressionComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneExpressionComparison(
  genes = NULL,
  scores = NULL,
  max.expr = NULL,
  plot.z = TRUE,
  plot.expression = TRUE,
  max.z = 5,
  smoothed = FALSE,
  plot.na = -1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
