% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
Cacoa R6 class

Cacoa R6 class

Plot inter-sample expression distance

Plot sample-sample expression distance as a 2D embedding
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{number of cores}

\item{\code{verbose}}{print diagnostic messages}

\item{\code{test.results}}{results of the estimations, ready to use}

\item{\code{cache}}{intermediate results of the estimations, which can be used during some other computations}

\item{\code{data.object}}{the main object storing data (Conos or Seurat)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample}

\item{\code{cell.groups}}{named factor with cell names with cluster per cell}

\item{\code{embedding}}{a 2D embedding to visualize the cells in}

\item{\code{sample.per.cell}}{named factor with cell names}

\item{\code{ref.level}}{reference level for sample.group vector}

\item{\code{target.level}}{target/disease level for sample.group vector}

\item{\code{sample.groups.palette}}{a color palette for the sample.groups}

\item{\code{cell.groups.palette}}{a color palette for the cell.groups}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-estimateCommonExpressionShiftMagnitudes}{\code{Cacoa$estimateCommonExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-plotCommonExpressionShiftMagnitudes}{\code{Cacoa$plotCommonExpressionShiftMagnitudes()}}
\item \href{#method-estimateExpressionShiftZScores}{\code{Cacoa$estimateExpressionShiftZScores()}}
\item \href{#method-plotExpressionShiftZScores}{\code{Cacoa$plotExpressionShiftZScores()}}
\item \href{#method-estimatePerCellTypeDE}{\code{Cacoa$estimatePerCellTypeDE()}}
\item \href{#method-plotDEGenes}{\code{Cacoa$plotDEGenes()}}
\item \href{#method-plotNumberOfDEGenes}{\code{Cacoa$plotNumberOfDEGenes()}}
\item \href{#method-saveDEasJSON}{\code{Cacoa$saveDEasJSON()}}
\item \href{#method-plotFilteredDEGenes}{\code{Cacoa$plotFilteredDEGenes()}}
\item \href{#method-prepareOntologyData}{\code{Cacoa$prepareOntologyData()}}
\item \href{#method-plotEmbedding}{\code{Cacoa$plotEmbedding()}}
\item \href{#method-estimateOntology}{\code{Cacoa$estimateOntology()}}
\item \href{#method-plotOntologyDistribution}{\code{Cacoa$plotOntologyDistribution()}}
\item \href{#method-plotOntologyTerms}{\code{Cacoa$plotOntologyTerms()}}
\item \href{#method-plotOntologyBarplot}{\code{Cacoa$plotOntologyBarplot()}}
\item \href{#method-plotOntologyDotplot}{\code{Cacoa$plotOntologyDotplot()}}
\item \href{#method-plotOntologyHeatmap}{\code{Cacoa$plotOntologyHeatmap()}}
\item \href{#method-plotOntologySimilarities}{\code{Cacoa$plotOntologySimilarities()}}
\item \href{#method-plotProportions}{\code{Cacoa$plotProportions()}}
\item \href{#method-plotCellNumbers}{\code{Cacoa$plotCellNumbers()}}
\item \href{#method-plotPcaSpace}{\code{Cacoa$plotPcaSpace()}}
\item \href{#method-plotCdaSpace}{\code{Cacoa$plotCdaSpace()}}
\item \href{#method-plotContrastTree}{\code{Cacoa$plotContrastTree()}}
\item \href{#method-estimateCellLoadings}{\code{Cacoa$estimateCellLoadings()}}
\item \href{#method-estimateGaPartiotion}{\code{Cacoa$estimateGaPartiotion()}}
\item \href{#method-plotCellLoadings}{\code{Cacoa$plotCellLoadings()}}
\item \href{#method-extimateWilcoxonTest}{\code{Cacoa$extimateWilcoxonTest()}}
\item \href{#method-estimateCellDensity}{\code{Cacoa$estimateCellDensity()}}
\item \href{#method-plotCellDensity}{\code{Cacoa$plotCellDensity()}}
\item \href{#method-diffCellDensity}{\code{Cacoa$diffCellDensity()}}
\item \href{#method-plotExpressionDistance}{\code{Cacoa$plotExpressionDistance()}}
\item \href{#method-plotExpressionDistanceEmbedding}{\code{Cacoa$plotExpressionDistanceEmbedding()}}
\item \href{#method-getContour}{\code{Cacoa$getContour()}}
\item \href{#method-estimateClusterFreeZScores}{\code{Cacoa$estimateClusterFreeZScores()}}
\item \href{#method-estimateClusterFreeExpressionShifts}{\code{Cacoa$estimateClusterFreeExpressionShifts()}}
\item \href{#method-smoothClusterFreeZScores}{\code{Cacoa$smoothClusterFreeZScores()}}
\item \href{#method-estimateGeneProgrammes}{\code{Cacoa$estimateGeneProgrammes()}}
\item \href{#method-plotGeneProgrammeScores}{\code{Cacoa$plotGeneProgrammeScores()}}
\item \href{#method-plotGeneProgrammeGenes}{\code{Cacoa$plotGeneProgrammeGenes()}}
\item \href{#method-plotClusterFreeExpressionShifts}{\code{Cacoa$plotClusterFreeExpressionShifts()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  ref.level = NULL,
  target.level = NULL,
  sample.groups.palette = NULL,
  cell.groups.palette = NULL,
  embedding = extractEmbedding(data.object),
  n.cores = 1,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
Calculate expression shift magnitudes of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  sample.groups = self$sample.groups,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "expression.shifts",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{dist}}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance (default="JS")}

\item{\code{within.group.normalization}}{Normalize the shift magnitude by the mean magnitude of within-group variation (default=T)}

\item{\code{valid.comparisons}}{A logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons (default=NULL)}

\item{\code{n.cells}}{Number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{\code{n.top.genes}}{Number of top highest-expressed genes to consider (default: all genes)}

\item{\code{n.subsamples}}{Number of samples to draw (default=100)}

\item{\code{min.cells}}{Minimum number of cells per cluster/per sample to be included in the analysis (default=10)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{Number of cores (default: stored integer)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{Test name (default="expression.shifts")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list include \itemize{
  \item{df: a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells}
  \item{ctdml: raw list of `n.subsamples` sampled distance matrices (cells were subsampled)}
  \item{sample.groups: same as the provided variable}
  \item{valid.comparisons: a matrix of valid comparisons (in this case all should be valid, since we're not restricting samples that should be compared)}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCommonExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCommonExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateCommonExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCommonExpressionShiftMagnitudes(
  sample.groups = self$sample.groups,
  cell.groups = self$cell.groups,
  n.cells = NULL,
  n.randomizations = 50,
  n.subsamples = 30,
  min.cells = 10,
  n.cores = 1,
  verbose = self$verbose,
  mean.trim = 0.1,
  name = "common.expression.shifts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  size.norm = F,
  notch = T,
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{size.norm}}{Plot size normalized results. Requires cell.groups, and sample.per.cell (default=F)}

\item{\code{notch}}{Show notches in plot, see ggplot2::geom_boxplot for more info (default=T)}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCommonExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotCommonExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotCommonExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCommonExpressionShiftMagnitudes(
  name = "common.expression.shifts",
  show.subsampling.variability = FALSE,
  show.jitter = FALSE,
  jitter.alpha = 0.05,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftZScores}{}}}
\subsection{Method \code{estimateExpressionShiftZScores()}}{
Calculate expression shift Z scores of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftZScores(
  cell.groups = self$cell.groups,
  ref.level = self$ref.level,
  sample.groups = self$sample.groups,
  sample.per.cell = self$sample.per.cell,
  n.od.genes = 1000,
  n.pcs = 100,
  pca.maxit = 1000,
  ignore.cache = F,
  name = "expression.z.scores",
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{ref.level}}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{sample.per.cell}}{(default: stored vector)}

\item{\code{n.od.genes}}{(default=1000)}

\item{\code{n.pcs}}{(default=100)}

\item{\code{pca.maxit}}{(default=1000)}

\item{\code{ignore.cache}}{(default=F)}

\item{\code{name}}{(default="expression.z.scores")}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftZScores}{}}}
\subsection{Method \code{plotExpressionShiftZScores()}}{
Plot results from cao$estimateExpressionShiftZScores()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftZScores(
  type.order = NULL,
  name = "expression.z.scores",
  size.norm = F,
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type.order}}{(default=NULL)}

\item{\code{name}}{Test results to plot (default=expression.z.shifts)}

\item{\code{size.norm}}{Show cluster size-normalized results}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePerCellTypeDE}{}}}
\subsection{Method \code{estimatePerCellTypeDE()}}{
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDE(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  common.genes = F,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  max.cell.count = Inf,
  test = "Wald",
  independent.filtering = FALSE,
  cluster.sep.chr = "<!!>",
  return.matrix = T,
  verbose = self$verbose,
  name = "de"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=NULL)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{common.genes}}{Only investigate common genes across cell groups (default=F)}

\item{\code{n.cores}}{Number of cores (default=1)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=F)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=FALSE)}

\item{\code{min.cell.count}}{minimum number of cells that need to be present in a given cell type in a given sample in order to be taken into account (default=10)}

\item{\code{max.cell.count}}{maximal number of cells per cluster per sample to include in a comparison (useful for comparing the number of DE genes between cell types) (default: Inf)}

\item{\code{test}}{which DESeq2 test to use (options: "LRT" (default), "Wald")}

\item{\code{independent.filtering}}{independentFiltering parameter for DESeq2 (default=FALSE)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{return.matrix}}{Return merged matrix of results (default=TRUE)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{name}}{slot in which to save the results (default: 'de')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEGenes}{}}}
\subsection{Method \code{plotDEGenes()}}{
Plot number of significant DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEGenes(
  name = "de",
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T,
  p.adjust.cutoff = 0.05,
  size = 4,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{results slot in which the DE results should be stored (default: 'de')}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{p.adjust.cutoff}}{Adjusted P cutoff (default=0.05)}

\item{\code{palette}}{cell group palette (default: stored $cell.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNumberOfDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotNumberOfDEGenes}{}}}
\subsection{Method \code{plotNumberOfDEGenes()}}{
Plot number of significant DE genes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumberOfDEGenes(
  name = "de",
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T,
  p.adjust = TRUE,
  pvalue.cutoff = 0.05,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{results slot in which the DE results should be stored (default: 'de')}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{p.adjust}}{whether the cutoff should be based on the adjusted P value (default: TRUE)}

\item{\code{pvalue.cutoff}}{P value cutoff (default=0.05)}

\item{\code{palette}}{cell group palette (default: stored $cell.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDEasJSON"></a>}}
\if{latex}{\out{\hypertarget{method-saveDEasJSON}{}}}
\subsection{Method \code{saveDEasJSON()}}{
Save DE results as JSON files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveDEasJSON(
  saveprefix = NULL,
  dir.name = "JSON",
  de.raw = self$test.results$de,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  gene.metadata = NULL,
  cluster.sep.chr = "<!!>",
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{saveprefix}}{Prefix for created files (default=NULL)}

\item{\code{dir.name}}{Name for directory with results (default="JSON")}

\item{\code{de.raw}}{List of DE results}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{gene.metadata}}{(default=NULL)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFilteredDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotFilteredDEGenes}{}}}
\subsection{Method \code{plotFilteredDEGenes()}}{
Plot number of highly-expressed DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotFilteredDEGenes(
  de.filter = self$test.results$ontology$de.filter,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepareOntologyData"></a>}}
\if{latex}{\out{\hypertarget{method-prepareOntologyData}{}}}
\subsection{Method \code{prepareOntologyData()}}{
Filter and prepare DE genes for ontology calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$prepareOntologyData(
  org.db,
  n.top.genes = 100,
  p.adj.cutoff = 0.05,
  expr.cutoff = 0.05,
  de.raw = self$test.results$de,
  cell.groups = self$cell.groups,
  universe = NULL,
  transposed = T,
  verbose = T,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{n.top.genes}}{Number of most different genes to take as input. If less are left after filtering for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}

\item{\code{p.adj.cutoff}}{Cutoff for filtering highly-expressed DE genes (default=0.05)}

\item{\code{expr.cutoff}}{Cutoff for cells per group expressing a DE gene, i.e., cutoff for highly-expressed genes (default=0.05)}

\item{\code{de.raw}}{Differentially expressed genes per cell group, results from estimatePerCellTypeDE (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{universe}}{Only set this if a common background gene set is desired for all cell groups (default: NULL)}

\item{\code{transposed}}{Whether count matrices should be transposed (default=T)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{n.cores}}{Number of cores to use (default: stored integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-plotEmbedding}{}}}
\subsection{Method \code{plotEmbedding()}}{
Plot embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotEmbedding(
  embedding = self$embedding,
  plot.theme = ggplot2::theme_bw(),
  show.legend = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{embedding}}{A cell embedding to use (two-column data frame with rownames corresponding to cells) (default: stored embedding object)}

\item{\code{plot.theme}}{plot theme to use (default: ggplot2::theme_bw())}

\item{\code{...}}{other parameters are passed to \link[sccore:embeddingPlot]{embeddingPlot}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntology"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntology}{}}}
\subsection{Method \code{estimateOntology()}}{
Estimate ontology terms based on DEs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntology(
  type = "GO",
  org.db,
  de.gene.ids = self$test.results$ontology$de.gene.ids,
  go.environment = self$test.results$GO$go.environment,
  p.adj = 0.05,
  p.adjust.method = "BH",
  readable = T,
  verbose = T,
  min.genes = 0,
  qvalueCutoff = 0.2,
  minGSSize = 5,
  maxGSSize = 500,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information (default="GO")}

\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{de.gene.ids}}{List containing DE gene IDs, and filtered DE genes (default: stored list, results from prepareOntologyData)}

\item{\code{go.environment}}{Extracted GO environment. If set to NULL, the environment will be re-extracted (default: stored environment)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}

\item{\code{p.adjust.method}}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{\code{readable}}{Mapping gene ID to gene name (default=T)}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{min.genes}}{Minimum number of input genes overlapping with ontologies (default=0)}

\item{\code{qvalueCutoff}}{Q value cutoff, please see clusterProfiler package for more information (default=0.2)}

\item{\code{minGSSize}}{Minimal geneset size, please see clusterProfiler package for more information (default=5)}

\item{\code{maxGSSize}}{Minimal geneset size, please see clusterProfiler package for more information (default=5e2)}

\item{\code{...}}{Additional parameters for sccore:::plapply function}

\item{\code{universe}}{All measured genes (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing a list of terms per ontology, and a data frame with merged results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyDistribution}{}}}
\subsection{Method \code{plotOntologyDistribution()}}{
Bar plot of ontology terms per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyDistribution(
  genes = c("up", "down"),
  type = "GO",
  cell.groups = self$cell.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyTerms"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyTerms}{}}}
\subsection{Method \code{plotOntologyTerms()}}{
Plot ontology terms as a function of both number of DE genes, and number of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyTerms(
  genes = c("up", "down"),
  type = "GO",
  de.filter = self$test.results$ontology$de.filter,
  cell.groups = self$cell.groups,
  label.x.pos = 0.01,
  label.y.pos = 1,
  scale = 0.93
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up', 'all' or a combination of these (default=c("up","down"))}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{label.x.pos}}{Plot label position on x axis (default=0.01)}

\item{\code{label.y.pos}}{Plot label position on y axis (default=1)}

\item{\code{scale}}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (default=0.93)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyBarplot"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyBarplot}{}}}
\subsection{Method \code{plotOntologyBarplot()}}{
Plot a barplot of ontology terms with adj. P values for a specific cell subgroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyBarplot(
  genes = "up",
  type = "BP",
  cell.subgroups = NULL,
  n = 20,
  p.adj = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{cell.subgroups}}{Cell group to plot (default=NULL)}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyDotplot"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyDotplot}{}}}
\subsection{Method \code{plotOntologyDotplot()}}{
Plot a dotplot of ontology terms with adj. P values for a specific cell subgroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyDotplot(
  genes = "up",
  type = "BP",
  cell.subgroups = NULL,
  n = 20,
  p.adj = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{cell.subgroups}}{Cell group to plot (default=NULL)}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyHeatmap}{}}}
\subsection{Method \code{plotOntologyHeatmap()}}{
Plot a heatmap of ontology P values per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmap(
  genes = "up",
  type = "GO",
  legend.position = "left",
  selection = "all",
  n = 10,
  cell.subgroups = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down' for downregulated genes, 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{selection}}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type); 'common' (only show terms that are present in at least two cell types); 'all' (all ontology terms) (default="all")}

\item{\code{n}}{Number of terms to show (default=10)}

\item{\code{cell.subgroups}}{Cell groups to plot (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologySimilarities"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologySimilarities}{}}}
\subsection{Method \code{plotOntologySimilarities()}}{
Plot correlation matrix for ontology terms between cell types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologySimilarities(genes = "up", type = "GO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotProportions"></a>}}
\if{latex}{\out{\hypertarget{method-plotProportions}{}}}
\subsection{Method \code{plotProportions()}}{
Plot the cell group proportions per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotProportions(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  notch = FALSE,
  alpha = 0.2,
  palette = self$sample.groups.palette,
  show.significance = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{cells.to.remove}}{Vector of cell types to remove from the composition}

\item{\code{cells.to.remain}}{Vector of cell types to remain in the composition}

\item{\code{notch}}{Whether to show notch in the boxplots}

\item{\code{alpha}}{Transparency level on the data points (default: 0.2)}

\item{\code{palette}}{color palette to use for conditions (default: stored $sample.groups.palette)}

\item{\code{show.significance}}{whether to show statistical significance betwwen sample groups. wilcox.test was used; (* < 0.05; ** < 0.01; *** < 0.001)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellNumbers"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellNumbers}{}}}
\subsection{Method \code{plotCellNumbers()}}{
Plot the cell numbers per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellNumbers(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPcaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-plotPcaSpace}{}}}
\subsection{Method \code{plotPcaSpace()}}{
Plot compositions in CoDA-PCA space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotPcaSpace(
  cells.to.remove = NULL,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCdaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-plotCdaSpace}{}}}
\subsection{Method \code{plotCdaSpace()}}{
Plot compositions in CoDA-CDA space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCdaSpace(
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotContrastTree"></a>}}
\if{latex}{\out{\hypertarget{method-plotContrastTree}{}}}
\subsection{Method \code{plotContrastTree()}}{
Plot contrast tree
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotContrastTree(cells.to.remain = NULL, cells.to.remove = NULL)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellLoadings}{}}}
\subsection{Method \code{estimateCellLoadings()}}{
Plot Loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellLoadings(
  n.cell.counts = 1000,
  n.seed = 239,
  aplha = 0.01,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  samples.to.remove = NULL,
  signif.threshold = 0.2
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGaPartiotion"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGaPartiotion}{}}}
\subsection{Method \code{estimateGaPartiotion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGaPartiotion(
  n.cell.counts = 1000,
  n.seed = 239,
  aplha = 0.01,
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellLoadings}{}}}
\subsection{Method \code{plotCellLoadings()}}{
Plot Loadings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellLoadings(
  n.cell.counts = 1000,
  n.seed = 239,
  aplha = 0.01,
  font.size = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  palette = self$cell.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{palette}}{palette specification for cell types (default: stored $cell.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extimateWilcoxonTest"></a>}}
\if{latex}{\out{\hypertarget{method-extimateWilcoxonTest}{}}}
\subsection{Method \code{extimateWilcoxonTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$extimateWilcoxonTest(
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups,
  cells.to.remove = NULL,
  cells.to.remain = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellDensity}{}}}
\subsection{Method \code{estimateCellDensity()}}{
Estimate cell density in giving embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellDensity(
  embedding = self$embedding,
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  sample.per.cell = self$sample.per.cell,
  ref.level = self$ref.level,
  target.level = self$target.level,
  bins = 400,
  by.sample = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default: stored vector)}

\item{\code{ref.level}}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{\code{target.level}}{target/disease level for sample.group vector}

\item{\code{bins}}{number of bins for density esitmation, default 400}

\item{\code{by.sample}}{if TRUE, density will esitmated by sample and quantiles normlization will applied to indivisual sample. If FALSE, cell condition.per.cell need to be provided and density will simply esitmated by condition.per.cell.}

\item{\code{emb}}{cell embedding matrix}

\item{\code{condition.per.cell}}{Named group factor with cell names. Must have exactly two levels.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellDensity}{}}}
\subsection{Method \code{plotCellDensity()}}{
Plot cell density
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensity(
  col = c("blue", "white", "red"),
  show.legend = NULL,
  legend.position = NULL,
  title = NULL,
  show.grid = NULL,
  add.ponits = TRUE,
  condition.per.cell = NULL,
  color = "B",
  point.col = "#FCFDBFFF",
  contours = NULL,
  contour.color = "white",
  contour.conf = "10\%"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add.ponits}}{default is TRUE, add points to cell density figure}

\item{\code{condition.per.cell}}{Named group factor with cell names. Must have exactly two levels. condition.per.cell must be provided when add.ponits is TRUE}

\item{\code{contours}}{specify cell types for contour, mutiple cell types are also suported}

\item{\code{contour.color}}{color for contour line}

\item{\code{contour.conf}}{confidence interval of contour}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-diffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-diffCellDensity}{}}}
\subsection{Method \code{diffCellDensity()}}{
esitmate differential cell density
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$diffCellDensity(
  condition.per.cell = NULL,
  method = "substract",
  col = c("blue", "white", "red"),
  show.legend = NULL,
  legend.position = NULL,
  title = NULL,
  show.grid = NULL,
  plot = TRUE,
  contours = NULL,
  contour.color = "white",
  contour.conf = "10\%",
  z.cutoff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{condition.per.cell}}{A two-level factor on the cell names describing the conditions being compared (default: stored vector)}

\item{\code{method}}{method to cacuated differential cell density of each bin; substract: target density minus ref density; entropy: estimated kl divergence entropy betwwen sample grapups ; t.test: zscore of t-test,global variacen is setting for t.test;}

\item{\code{col}}{color palettes,  default is c('blue','white','red')}

\item{\code{contours}}{specify cell types for contour, mutiple cell types are also suported}

\item{\code{contour.color}}{color for contour line}

\item{\code{contour.conf}}{confidence interval of contour}

\item{\code{z.cutoff}}{absolute z score cutoff}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionDistance"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionDistance}{}}}
\subsection{Method \code{plotExpressionDistance()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistance(
  name = "expression.shifts",
  notch = TRUE,
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  weighted.distance = TRUE,
  min.cells = 10,
  palette = self$sample.groups.palette,
  show.significance = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{notch}}{Show notches in plot, see ggplot2::geom_boxplot for more info (default=T)}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored $cell.groups vector)}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{weighted.distance}}{whether to weigh the expression distance by the sizes of cell types (default: TRUE), or show distances for each individual cell type}

\item{\code{show.significance}}{whether to show statistical significance betwwen sample groups. wilcox.test was used; (* < 0.05; ** < 0.01; *** < 0.001)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionDistanceEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionDistanceEmbedding}{}}}
\subsection{Method \code{plotExpressionDistanceEmbedding()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistanceEmbedding(
  name = "expression.shifts",
  sample.groups = self$sample.groups,
  cell.type = NULL,
  method = "tSNE",
  perplexity = 4,
  max_iter = 1000,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{cell.type}}{If a name of a cell type is specified, the sample distances will be assessed based on this cell type alone. Otherwise (cell.type=NULL, default), sample distances will be estimated as an average distance across all cell types (weighted by the minimum number of cells of that cell type between any two samples being compared)}

\item{\code{method}}{dimension reduction methods (MDS or tSNE ) , default is MDS}

\item{\code{perplexity}}{tSNE perpexity (default: 4)}

\item{\code{max_iter}}{tSNE max_iter (default: 1e3)}

\item{\code{palette}}{a set of colors to use for conditions (default: stored $sample.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getContour"></a>}}
\if{latex}{\out{\hypertarget{method-getContour}{}}}
\subsection{Method \code{getContour()}}{
Extract contour from embedding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getContour(cells, color = "white", linetype = 2, conf = "10\%")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf}}{confidence interval of contour}

\item{\code{cell}}{specify cell types for contour, mutiple cell types are also suported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeZScores}{}}}
\subsection{Method \code{estimateClusterFreeZScores()}}{
Estimate differencial expression Z-scores between two conditions per individual cell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeZScores(
  n.od.genes = NULL,
  verbose = self$verbose,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.od.genes}}{number of overdispersed genes for estimating Z-scores}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{...}}{parameters forwarded to \link{estimateClusterFreeZScores}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sparse matrix of z-scores with genes as columns and cells as rows.
Cells that have only one condition in their expression neighborhood have NA Z-scores for all genes.
Results are also stored in the `cluster.free.z` field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeExpressionShifts}{}}}
\subsection{Method \code{estimateClusterFreeExpressionShifts()}}{
Estimate Cluster-free Expression Shift
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeExpressionShifts(
  n.od.genes = NULL,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.od.genes}}{number of overdispersed genes for estimating expression shifts}

\item{\code{verbose}}{show progress (default: stored value)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of cluster-free expression shifts per cell. Values above 1 correspond to difference between conditions.
Results are also stored in the `cluster.free.expr.shifts` field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-smoothClusterFreeZScores}{}}}
\subsection{Method \code{smoothClusterFreeZScores()}}{
Performs graph smoothing of the cluster-free DE Z-scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$smoothClusterFreeZScores(
  smoothing = 20,
  filter = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{smoothing}}{`beta` parameter of the \link[=sccore:heatFilter]{heatFilter}. Default: 20.}

\item{\code{filter}}{graph filter function. Default: \link[=sccore:heatFilter]{heatFilter}.}

\item{\code{n.cores}}{number of cores for parallelisation}

\item{\code{verbose}}{show progress (default: stored value)}

\item{\code{...}}{parameters forwarded to \link[=sccore:smoothSignalOnGraph]{smoothSignalOnGraph}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sparse matrix of smoothed Z-scores. Results are also stored in the `cluster.free.z.smoothed` field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGeneProgrammes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGeneProgrammes}{}}}
\subsection{Method \code{estimateGeneProgrammes()}}{
Estimate Gene Programmes based on cluster-free Z-scores on a subsample of
cells using \link[=fabia:fabia]{fabia}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGeneProgrammes(
  smoothed = TRUE,
  n.programmes = 15,
  name = "gene.programmes",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.programmes}}{maximal number of gene programmes to find (parameter `p` for fabia). Default: 15.}

\item{\code{name}}{field name where the test results are stored}

\item{\code{...}}{keyword arguments forwarded to \link{estimateGeneProgrammes}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgrammeScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgrammeScores}{}}}
\subsection{Method \code{plotGeneProgrammeScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgrammeScores(
  name = "gene.programmes",
  approximate = TRUE,
  build.panel = TRUE,
  nrow = NULL,
  gradient.range.quantile = 0.975,
  plot.na = approximate,
  legend.title = "Score",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgrammeGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgrammeGenes}{}}}
\subsection{Method \code{plotGeneProgrammeGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgrammeGenes(
  programme.id,
  name = "gene.programmes",
  max.genes = 9,
  build.panel = TRUE,
  nrow = NULL,
  color.range = c(-3, 3),
  smoothed = FALSE,
  size = 0.1,
  alpha = 0.1,
  plot.na = -1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-plotClusterFreeExpressionShifts}{}}}
\subsection{Method \code{plotClusterFreeExpressionShifts()}}{
Plot cluster-free expression shift z-scores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotClusterFreeExpressionShifts(
  cell.groups = self$cell.groups,
  plot.both.conditions = FALSE,
  plot.na = FALSE,
  max.shift = 3,
  alpha = 0.2,
  font.size = c(2, 3),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{cell type labels. Set to NULL if it shouldn't be shown}

\item{\code{plot.both.conditions}}{show both case and control cells. Normally, showing control cells doesn't
make sense, as control cells always have small distance from control.}

\item{\code{max.shift}}{all shift values above `max.shift` are set to this value when plotting.}

\item{\code{font.size}}{size range for cell type labels}

\item{\code{...}}{parameters forwarded to \link[sccore:embeddingPlot]{embeddingPlot}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
