% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
The class encompasses etc etc
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{Number of cores (default=1)}

\item{\code{verbose}}{boolean Whether to provide verbose output with diagnostic messages (default=FALSE)}

\item{\code{test.results}}{list Results of the estimations, ready to use (default=list())}

\item{\code{cache}}{list Intermediate results of the estimations, which can be used during some other computations (default=list())}

\item{\code{data.object}}{list The main object storing data (Conos or Seurat) (default=list())}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample (default=NULL)}

\item{\code{cell.groups}}{Named factor with cell names with cluster per cell (default=NULL)}

\item{\code{embedding}}{2D embedding to visualize the cells in (default=NULL)}

\item{\code{sample.per.cell}}{Named factor with cell names (default=NULL)}

\item{\code{ref.level}}{Reference level for sample.group vector (default=NULL)}

\item{\code{target.level}}{Target/disease level for sample.group vector (default=NULL)}

\item{\code{sample.groups.palette}}{Color palette for the sample.groups (default=NULL)}

\item{\code{cell.groups.palette}}{Color palette for the cell.groups (default=NULL)}

\item{\code{plot.theme}}{ggplot2 theme for all plots (default=NULL)}

\item{\code{plot.params}}{parameters, forwarded to all \code{plotEmbedding} calls  (default=NULL)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-estimateCommonExpressionShiftMagnitudes}{\code{Cacoa$estimateCommonExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-estimatePerCellTypeDE}{\code{Cacoa$estimatePerCellTypeDE()}}
\item \href{#method-estimateDEPerCellType}{\code{Cacoa$estimateDEPerCellType()}}
\item \href{#method-estimateDEStabilityPerCellType}{\code{Cacoa$estimateDEStabilityPerCellType()}}
\item \href{#method-estimateDEStabilityPerTest}{\code{Cacoa$estimateDEStabilityPerTest()}}
\item \href{#method-estimateDEStabilityPerGene}{\code{Cacoa$estimateDEStabilityPerGene()}}
\item \href{#method-estimateDEStabilityBetweenTests}{\code{Cacoa$estimateDEStabilityBetweenTests()}}
\item \href{#method-estimateDEStabilityTrend}{\code{Cacoa$estimateDEStabilityTrend()}}
\item \href{#method-estimateGOStabilityPerCellTypePermDir}{\code{Cacoa$estimateGOStabilityPerCellTypePermDir()}}
\item \href{#method-estimateGOStabilityPerCellTypePerm}{\code{Cacoa$estimateGOStabilityPerCellTypePerm()}}
\item \href{#method-estimateGOStabilityPerCellTypePermGseaDir}{\code{Cacoa$estimateGOStabilityPerCellTypePermGseaDir()}}
\item \href{#method-estimateGOStabilityPerCellTypePermGsea}{\code{Cacoa$estimateGOStabilityPerCellTypePermGsea()}}
\item \href{#method-estimateGOStabilityTrend}{\code{Cacoa$estimateGOStabilityTrend()}}
\item \href{#method-plotGOStabilityPerCellType}{\code{Cacoa$plotGOStabilityPerCellType()}}
\item \href{#method-plotDEStabilityTrend}{\code{Cacoa$plotDEStabilityTrend()}}
\item \href{#method-plotDEStabilityFDR}{\code{Cacoa$plotDEStabilityFDR()}}
\item \href{#method-plotDEStabilityFDR1loo}{\code{Cacoa$plotDEStabilityFDR1loo()}}
\item \href{#method-plotDEStabilityBetweenTests}{\code{Cacoa$plotDEStabilityBetweenTests()}}
\item \href{#method-plotDEStabilityPerCellType}{\code{Cacoa$plotDEStabilityPerCellType()}}
\item \href{#method-plotDEStabilityPerTest}{\code{Cacoa$plotDEStabilityPerTest()}}
\item \href{#method-plotDEStabilityPerGene}{\code{Cacoa$plotDEStabilityPerGene()}}
\item \href{#method-plotNumberOfDEGenes}{\code{Cacoa$plotNumberOfDEGenes()}}
\item \href{#method-plotVolcano}{\code{Cacoa$plotVolcano()}}
\item \href{#method-saveDEasJSON}{\code{Cacoa$saveDEasJSON()}}
\item \href{#method-plotEmbedding}{\code{Cacoa$plotEmbedding()}}
\item \href{#method-estimateOntology}{\code{Cacoa$estimateOntology()}}
\item \href{#method-estimateOntologyFamilies}{\code{Cacoa$estimateOntologyFamilies()}}
\item \href{#method-getFamiliesPerGO}{\code{Cacoa$getFamiliesPerGO()}}
\item \href{#method-plotNumOntologyTermsPerType}{\code{Cacoa$plotNumOntologyTermsPerType()}}
\item \href{#method-plotOntology}{\code{Cacoa$plotOntology()}}
\item \href{#method-plotOntologyHeatmap}{\code{Cacoa$plotOntologyHeatmap()}}
\item \href{#method-plotOntologyHeatmapCollapsed}{\code{Cacoa$plotOntologyHeatmapCollapsed()}}
\item \href{#method-plotOntologySimilarities}{\code{Cacoa$plotOntologySimilarities()}}
\item \href{#method-plotOntologyFamily}{\code{Cacoa$plotOntologyFamily()}}
\item \href{#method-saveOntologyAsTable}{\code{Cacoa$saveOntologyAsTable()}}
\item \href{#method-saveFamiliesAsTable}{\code{Cacoa$saveFamiliesAsTable()}}
\item \href{#method-plotCellGroupSizes}{\code{Cacoa$plotCellGroupSizes()}}
\item \href{#method-plotCellGroupAbundanceVariation}{\code{Cacoa$plotCellGroupAbundanceVariation()}}
\item \href{#method-plotCodaSpace}{\code{Cacoa$plotCodaSpace()}}
\item \href{#method-plotContrastTree}{\code{Cacoa$plotContrastTree()}}
\item \href{#method-plotCompositionSimilarity}{\code{Cacoa$plotCompositionSimilarity()}}
\item \href{#method-estimateCellLoadings}{\code{Cacoa$estimateCellLoadings()}}
\item \href{#method-estimateGaPartition}{\code{Cacoa$estimateGaPartition()}}
\item \href{#method-plotCellLoadings}{\code{Cacoa$plotCellLoadings()}}
\item \href{#method-estimateWilcoxonTest}{\code{Cacoa$estimateWilcoxonTest()}}
\item \href{#method-estimateCellDensity}{\code{Cacoa$estimateCellDensity()}}
\item \href{#method-plotCellDensity}{\code{Cacoa$plotCellDensity()}}
\item \href{#method-plotCellDensityVariation}{\code{Cacoa$plotCellDensityVariation()}}
\item \href{#method-estimateDiffCellDensity}{\code{Cacoa$estimateDiffCellDensity()}}
\item \href{#method-plotDiffCellDensity}{\code{Cacoa$plotDiffCellDensity()}}
\item \href{#method-plotExpressionDistance}{\code{Cacoa$plotExpressionDistance()}}
\item \href{#method-getSampleDistanceMatrix}{\code{Cacoa$getSampleDistanceMatrix()}}
\item \href{#method-plotSampleDistances}{\code{Cacoa$plotSampleDistances()}}
\item \href{#method-estimateMetadataSeparation}{\code{Cacoa$estimateMetadataSeparation()}}
\item \href{#method-estimateClusterFreeDE}{\code{Cacoa$estimateClusterFreeDE()}}
\item \href{#method-getMostChangedGenes}{\code{Cacoa$getMostChangedGenes()}}
\item \href{#method-estimateClusterFreeExpressionShifts}{\code{Cacoa$estimateClusterFreeExpressionShifts()}}
\item \href{#method-smoothClusterFreeZScores}{\code{Cacoa$smoothClusterFreeZScores()}}
\item \href{#method-estimateGenePrograms}{\code{Cacoa$estimateGenePrograms()}}
\item \href{#method-plotGeneProgramScores}{\code{Cacoa$plotGeneProgramScores()}}
\item \href{#method-plotGeneProgramGenes}{\code{Cacoa$plotGeneProgramGenes()}}
\item \href{#method-plotClusterFreeExpressionShifts}{\code{Cacoa$plotClusterFreeExpressionShifts()}}
\item \href{#method-plotMostChangedGenes}{\code{Cacoa$plotMostChangedGenes()}}
\item \href{#method-plotGeneExpressionComparison}{\code{Cacoa$plotGeneExpressionComparison()}}
\item \href{#method-getConditionPerCell}{\code{Cacoa$getConditionPerCell()}}
\item \href{#method-getJointCountMatrix}{\code{Cacoa$getJointCountMatrix()}}
\item \href{#method-getGOEnvironment}{\code{Cacoa$getGOEnvironment()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Cacoa class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  ref.level = NULL,
  target.level = NULL,
  sample.groups.palette = NULL,
  cell.groups.palette = NULL,
  embedding = extractEmbedding(data.object),
  graph.name = NULL,
  n.cores = 1,
  verbose = TRUE,
  plot.theme = ggplot2::theme_bw(),
  plot.params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data.object}}{Object used to initialize the Cacoa class. Either a raw or normalized count matrix, Conos object, or Seurat object.}

\item{\code{sample.groups}}{(default=NULL)}

\item{\code{cell.groups}}{(default=NULL)}

\item{\code{sample.per.cell}}{(default=NULL)}

\item{\code{ref.level}}{(default=NULL)}

\item{\code{target.level}}{(default=NULL)}

\item{\code{sample.groups.palette}}{(default=NULL)}

\item{\code{cell.groups.palette}}{(default=NULL)}

\item{\code{embedding}}{(default=extractEmbedding(data.object))}

\item{\code{graph.name}}{(default=NULL)}

\item{\code{n.cores}}{numeric (default=1)}

\item{\code{verbose}}{boolean (default=TRUE)}

\item{\code{plot.theme}}{(default=ggplot2::theme_bw())}

\item{\code{plot.params}}{(default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'Cacoa' object
Calculate expression shift magnitudes of different clusters between conditions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  dist = NULL,
  dist.type = "cross.both",
  min.cells.per.sample = 10,
  min.samp.per.type = 2,
  min.gene.frac = 0.01,
  ref.level = self$ref.level,
  sample.groups = self$sample.groups,
  verbose = self$verbose,
  n.cores = self$n.cores,
  name = "expression.shifts",
  n.permutations = 1000,
  genes = NULL,
  n.pcs = NULL,
  top.n.genes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default=stored vector, self$cell.groups)}

\item{\code{sample.per.cell}}{(default=self$sample.per.cell)}

\item{\code{dist}}{'cor' - correlation distance, 'l1' - manhattan distance or 'l2' - euclidean (default=NULL, depends on dimensionality)}

\item{\code{dist.type}}{character string (default="cross.both")}

\item{\code{min.cells.per.sample}}{numeric (default=10)}

\item{\code{min.samp.per.type}}{numeric (default=2)}

\item{\code{min.gene.frac}}{numeric (default=0.01)}

\item{\code{ref.level}}{(default=self$ref.level)}

\item{\code{sample.groups}}{(default=self$sample.groups)}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{Number of cores (default: stored integer)}

\item{\code{name}}{Test name (default="expression.shifts")}

\item{\code{n.permutations}}{numeric (default=1000)}

\item{\code{genes}}{(default=NULL)}

\item{\code{n.pcs}}{numeric Number of principal components (default=NULL)}

\item{\code{...}}{extra parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List including:
\itemize{
\item \code{dist.df}: a table with cluster distances (normalized if within.gorup.normalization=TRUE), cell type and the number of cells # TODO: update
\item \code{p.dist.info}: list of distance matrices per cell type
\item \code{sample.groups}: filtered sample groups
\item \code{cell.groups}: filtered cell groups
}

Estimate common expression shift magnitudes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCommonExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCommonExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateCommonExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCommonExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  name = "common.expression.shifts",
  min.cells.per.sample = 10,
  min.samp.per.type = 2,
  min.gene.frac = 0.01,
  genes = NULL,
  n.permutations = 1000,
  trim = 0.2,
  p.adjust.method = "BH",
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell names (default: stored vector)}

\item{\code{name}}{character string Field name where the test results are stored (default="common.expression.shifts")}

\item{\code{min.cells.per.sample}}{numeric (default=10)}

\item{\code{min.samp.per.type}}{numeric (default=2)}

\item{\code{min.gene.frac}}{numeric (default=0.01)}

\item{\code{genes}}{(default=NULL)}

\item{\code{n.permutations}}{numeric (default=1000)}

\item{\code{n.permutations}}{numeric (default=1000)}

\item{\code{trim}}{numeric (deafult=0.2)}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values. Please see DOSE package for more information (default="BH")}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{...}}{additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of expression shift magnitudes. Distributions per type, p-values, and adjusted p-values.
Plot results from cao$estimateExpressionShiftMagnitudes() (shift.type="normal") or
cao$estimateCommonExpressionShiftMagnitudes() (shift.type="common")
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = NULL,
  type = "box",
  notch = TRUE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.pvalues = c("adjusted", "raw", "none"),
  shift.type = c("normal", "common"),
  ylab = "normalized expression distance",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{\itemize{
\item results slot name (default=NULL)
}}

\item{\code{type}}{\itemize{
\item type of a plot "bar" (default) or "box"
}}

\item{\code{show.jitter}}{boolean Whether to show indivudal data points (default=FALSE)}

\item{\code{jitter.alpha}}{numeric Transparency value for the data points (default=0.05)}

\item{\code{show.pvalues}}{character string Which p-values to plot. Accepted values are "none", "raw", or "adjusted". (default=c("adjusted", "raw", "none"))}

\item{\code{ylab}}{character string Label of the y-axis (default="normalized expression distance")}

\item{\code{...}}{additional arguments}

\item{\code{shift.types}}{character string Either "normal" or "common". (default=c("normal", "common"))}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePerCellTypeDE}{}}}
\subsection{Method \code{estimatePerCellTypeDE()}}{
Alias for estimateDEPerCellType
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDE(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEPerCellType}{}}}
\subsection{Method \code{estimateDEPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEPerCellType(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  target.level = self$target.level,
  name = "de",
  test = "DESeq2.Wald",
  resampling.method = NULL,
  n.resamplings = 30,
  seed.resampling = 239,
  min.cell.frac = 0.05,
  covariates = NULL,
  common.genes = FALSE,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  independent.filtering = FALSE,
  min.cell.count = 10,
  n.cells.subsample = NULL,
  verbose = self$verbose,
  fix.n.samples = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=self$cell.groups)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample (default=self$sample.groups)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy (default=self$ref.level)}

\item{\code{target.level}}{Reference level in 'sample.groups', e.g., case, diseased (default=self$target.level)}

\item{\code{name}}{character string Slot in which to save the results (default='de')}

\item{\code{test}}{character string Which DESeq2 test to use. The available options are "LRT", "Wald". (default="DESeq2.Wald")}

\item{\code{resampling.method}}{which resampling method should be used "loo" for leave-one-out or "bootstrap", (default=NULL, i.e. no resampling)}

\item{\code{n.resamplings}}{numeric Number of resamplings to perform (default=30)}

\item{\code{covariates}}{numeric (default=NULL)}

\item{\code{common.genes}}{boolean Whether to investigate common genes across cell groups (default=FALSE)}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{cooks.cutoff}}{boolean cooksCutoff for DESeq2 (default=FALSE)}

\item{\code{independent.filtering}}{boolean independentFiltering parameter for DESeq2 (default=FALSE)}

\item{\code{min.cell.count}}{minimum number of cells that need to be present in a given cell type in a given sample in order to be taken into account (default=10)}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{...}}{additional parameters}

\item{\code{seed.resamplings}}{numeric Seed to use for resamplings, input to set.seed() (default=239)}

\item{\code{min.cell.frace}}{numeric Minimum fraction of cells to use to perform DE (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
Estimate DE stability per cell type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEStabilityPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEStabilityPerCellType}{}}}
\subsection{Method \code{estimateDEStabilityPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityPerCellType(
  de.name = "de",
  name = "de.jaccards",
  top.n.genes = NULL,
  p.val.cutoff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.name}}{character string Results slot name (default='de')}

\item{\code{name}}{character string (default='de.jaccards')}

\item{\code{top.n.genes}}{numeric Number of top DE genes to return (default=NULL)}

\item{\code{p.val.cutoff}}{numeric The p-value cutoff to apply for returned DE values (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate DE stability per test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEStabilityPerTest"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEStabilityPerTest}{}}}
\subsection{Method \code{estimateDEStabilityPerTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityPerTest(
  de.names,
  name = "jacc.per.test",
  top.n.genes = NULL,
  p.val.cutoff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='de.jaccards')}

\item{\code{top.n.genes}}{numeric Number of top DE genes to return (default=NULL)}

\item{\code{p.val.cutoff}}{numeric The p-value cutoff to apply for returned DE values (default=NULL)}

\item{\code{de.name}}{character string Results slot name (default='de')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate DE stability per gene
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEStabilityPerGene"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEStabilityPerGene}{}}}
\subsection{Method \code{estimateDEStabilityPerGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityPerGene(
  de.name,
  top.n.genes = 500,
  p.adj.cutoff = NULL,
  visualize = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.name}}{character string Results slot name (default='de')}

\item{\code{top.n.genes}}{numeric Number of top DE genes to return (default=500)}

\item{\code{visualize}}{boolean Whether to visualize results (default=FALSE)}

\item{\code{p.val.cutoff}}{numeric The p-value cutoff to apply for returned DE values (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate DE stability between tests
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEStabilityBetweenTests"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEStabilityBetweenTests}{}}}
\subsection{Method \code{estimateDEStabilityBetweenTests()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityBetweenTests(
  de.names,
  name = "jacc.bw.tests",
  top.n.genes = NULL,
  p.val.cutoff = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='jacc.bw.tests')}

\item{\code{top.n.genes}}{numeric Number of top DE genes to return (default=NULL)}

\item{\code{p.val.cutoff}}{numeric The p-value cutoff to apply for returned DE values (default=NULL)}

\item{\code{de.name}}{character string Results slot name (default='de')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate DE stability trend
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDEStabilityTrend"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDEStabilityTrend}{}}}
\subsection{Method \code{estimateDEStabilityTrend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDEStabilityTrend(
  de.name = "de",
  name = "de.trend",
  top.n.genes = c(100, 200, 300),
  p.val.cutoffs = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.name}}{character string Results slot name (default='de')}

\item{\code{name}}{character string (default='de.trend')}

\item{\code{top.n.genes}}{numeric Number of top DE genes to return (default=c(100, 200, 300))}

\item{\code{p.val.cutoff}}{numeric The p-value cutoff to apply for returned DE values (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate GO stability per cell type permutation directory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGOStabilityPerCellTypePermDir"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGOStabilityPerCellTypePermDir}{}}}
\subsection{Method \code{estimateGOStabilityPerCellTypePermDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGOStabilityPerCellTypePermDir(
  de.test = "de",
  name = "go.jaccards",
  top.n.de = 500,
  padj.go = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.test}}{character string (default='de')}

\item{\code{name}}{character string (default='de.trend')}

\item{\code{top.n.de}}{numeric Number of top DE genes to return (default=500)}

\item{\code{padj.go}}{numeric (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate GO stability per cell type permutation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGOStabilityPerCellTypePerm"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGOStabilityPerCellTypePerm}{}}}
\subsection{Method \code{estimateGOStabilityPerCellTypePerm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGOStabilityPerCellTypePerm(
  de.test = "de",
  top.n.de = 500,
  padj.go = 0.05,
  name = "go.jaccards"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.test}}{character string (default='de')}

\item{\code{top.n.de}}{numeric Number of top DE genes to return (default=500)}

\item{\code{padj.go}}{numeric (default=0.05)}

\item{\code{name}}{character string (default='de.trend')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate GO stability per cell GSEA directory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGOStabilityPerCellTypePermGseaDir"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGOStabilityPerCellTypePermGseaDir}{}}}
\subsection{Method \code{estimateGOStabilityPerCellTypePermGseaDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGOStabilityPerCellTypePermGseaDir(
  de.test = "de",
  padj.go = 0.05,
  name = "go.jaccards"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.test}}{character string (default='de')}

\item{\code{padj.go}}{numeric (default=0.05)}

\item{\code{name}}{character string (default='de.trend')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate GO stability per cell GSEA directory
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGOStabilityPerCellTypePermGsea"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGOStabilityPerCellTypePermGsea}{}}}
\subsection{Method \code{estimateGOStabilityPerCellTypePermGsea()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGOStabilityPerCellTypePermGsea(
  de.test = "de",
  padj.go = 0.05,
  name = "go.jaccards"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.test}}{character string (default='de')}

\item{\code{padj.go}}{numeric (default=0.05)}

\item{\code{name}}{character string (default='de.trend')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate GO stability trend
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGOStabilityTrend"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGOStabilityTrend}{}}}
\subsection{Method \code{estimateGOStabilityTrend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGOStabilityTrend(
  de.name = "go",
  name = "go.trend",
  top.n.genes = c(100, 200, 300),
  p.val.cutoffs = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='de.trend')}

\item{\code{top.n.genes}}{numeric vector (default=c(100, 200, 300))}

\item{\code{p.val.cutoffs}}{numeric (default=NULL)}

\item{\code{de.test}}{character string (default='de')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
plot GO stability per cell type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGOStabilityPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-plotGOStabilityPerCellType}{}}}
\subsection{Method \code{plotGOStabilityPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGOStabilityPerCellType(
  name = "go.jaccards",
  notch = FALSE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.pairs = FALSE,
  sort.order = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='go.jaccards')}

\item{\code{notch}}{boolean (default=FALSE)}

\item{\code{show.jitter}}{boolean (default=TRUE)}

\item{\code{jitter.alpha}}{numeric (default=0.05)}

\item{\code{show.pairs}}{boolean (default=FALSE)}

\item{\code{sort.order}}{boolean (default=FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
plot DE stability trends
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityTrend"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityTrend}{}}}
\subsection{Method \code{plotDEStabilityTrend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityTrend(
  name = "de.trend",
  notch = FALSE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.pairs = FALSE,
  sort.order = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='de.trend')}

\item{\code{notch}}{boolean (default=FALSE)}

\item{\code{show.jitter}}{boolean (default=TRUE)}

\item{\code{jitter.alpha}}{numeric (default=0.05)}

\item{\code{show.pairs}}{boolean (default=FALSE)}

\item{\code{sort.order}}{boolean (default=FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
plot DE stability FDR
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityFDR"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityFDR}{}}}
\subsection{Method \code{plotDEStabilityFDR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityFDR(
  de.name = "de",
  p.adj.cutoffs = c(0.001, 0.005, 0.01, 0.05, 0.1, 0.2),
  cell.types = NULL,
  type = "relative"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.name}}{character string (default='de')}

\item{\code{p.adj.cutoffs}}{numeric vector (default=cc(0.001, 0.005, 0.01, 0.05, 0.1, 0.2))}

\item{\code{cell.types}}{(default=NULL)}

\item{\code{type}}{character string (default='relative')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
plot DE stability FDR
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityFDR1loo"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityFDR1loo}{}}}
\subsection{Method \code{plotDEStabilityFDR1loo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityFDR1loo(
  cell.type,
  de.name = "de",
  p.adj.cutoffs = c(0.001, 0.005, 0.01, 0.05, 0.1, 0.2),
  cell.types = NULL,
  type = "relative"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.name}}{character string (default='de')}

\item{\code{p.adj.cutoffs}}{numeric vector (default=cc(0.001, 0.005, 0.01, 0.05, 0.1, 0.2))}

\item{\code{cell.types}}{(default=NULL)}

\item{\code{type}}{character string (default='relative')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
plot DE stability between tests
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityBetweenTests"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityBetweenTests}{}}}
\subsection{Method \code{plotDEStabilityBetweenTests()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityBetweenTests(name = "jacc.bw.tests", cell.types = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='jacc.bw.tests')}

\item{\code{cell.type}}{(default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot DE stability per cell type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityPerCellType"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityPerCellType}{}}}
\subsection{Method \code{plotDEStabilityPerCellType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityPerCellType(
  name = "de.jaccards",
  notch = FALSE,
  show.jitter = TRUE,
  jitter.alpha = 0.05,
  show.pairs = FALSE,
  sort.order = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Results slot name (default='de.jaccards')}

\item{\code{notch}}{boolean Whether (default=FALSE)}

\item{\code{show.jitter}}{boolean Whether to show jitter on plots (default=TRUE)}

\item{\code{jitter.alpha}}{numeric Parameter for jitter (default=0.05)}

\item{\code{show.pairs}}{boolwan Whether to show pairs (default=FALSE)}

\item{\code{sort.order}}{whether to show notches in the boxplot version (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot DE stability per test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityPerTest"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityPerTest}{}}}
\subsection{Method \code{plotDEStabilityPerTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityPerTest(
  name = "jacc.per.test",
  notch = FALSE,
  show.jitter = TRUE,
  jitter.alpha = 0.5,
  show.pairs = FALSE,
  sort.order = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Results slot name (default='jacc.per.test')}

\item{\code{notch}}{boolean Whether (default=FALSE)}

\item{\code{show.jitter}}{boolean Whether to show jitter on plots (default=TRUE)}

\item{\code{jitter.alpha}}{numeric Parameter for jitter (default=0.05)}

\item{\code{show.pairs}}{boolwan Whether to show pairs (default=FALSE)}

\item{\code{sort.order}}{whether to show notches in the boxplot version (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot DE stability per gene
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEStabilityPerGene"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEStabilityPerGene}{}}}
\subsection{Method \code{plotDEStabilityPerGene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEStabilityPerGene(
  name = "de",
  cell.type = NULL,
  stability.score = "stab.median.rank"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Results slot name (default='de')}

\item{\code{cell.type}}{(default=NULL)}

\item{\code{stability.score}}{character string (default='stab.median.rank')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot number of significant DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNumberOfDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotNumberOfDEGenes}{}}}
\subsection{Method \code{plotNumberOfDEGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumberOfDEGenes(
  name = "de",
  p.adjust = TRUE,
  pvalue.cutoff = 0.05,
  show.resampling.results = TRUE,
  show.jitter = FALSE,
  jitter.alpha = 0.05,
  type = "bar",
  notch = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Results slot in which the DE results should be stored (default='de')}

\item{\code{p.adjust}}{boolean Whether the cutoff should be based on the adjusted P value (default=TRUE)}

\item{\code{pvalue.cutoff}}{numeric P-value cutoff (default=0.05)}

\item{\code{show.resampling.results}}{boolean Whether to show uncertainty based on resampling results (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Make volcano plots
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotVolcano"></a>}}
\if{latex}{\out{\hypertarget{method-plotVolcano}{}}}
\subsection{Method \code{plotVolcano()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotVolcano(
  name = "de",
  cell.types = NULL,
  palette = NULL,
  build.panel = TRUE,
  n.col = 3,
  color.var = "CellFrac",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='de')}

\item{\code{cell.types}}{(default=NULL)}

\item{\code{palette}}{(default=NULL)}

\item{\code{build.panel}}{boolean (default=TRUE)}

\item{\code{n.col}}{numeric Number of columns (default=3)}

\item{\code{color.var}}{character strign (default='CellFrac')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Save DE results as JSON files
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDEasJSON"></a>}}
\if{latex}{\out{\hypertarget{method-saveDEasJSON}{}}}
\subsection{Method \code{saveDEasJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveDEasJSON(
  saveprefix = NULL,
  dir.name = "JSON",
  de.raw = NULL,
  sample.groups = self$sample.groups,
  de.name = "de",
  ref.level = self$ref.level,
  gene.metadata = NULL,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{saveprefix}}{Prefix for created files (default=NULL)}

\item{\code{dir.name}}{Name for directory with results (default="JSON")}

\item{\code{de.raw}}{List of DE results}

\item{\code{sample.groups}}{a two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{gene.metadata}}{(default=NULL)}

\item{\code{verbose}}{boolean Whether to show progress}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
saved JSON object
Plot embedding
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotEmbedding"></a>}}
\if{latex}{\out{\hypertarget{method-plotEmbedding}{}}}
\subsection{Method \code{plotEmbedding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotEmbedding(
  embedding = self$embedding,
  color.by = NULL,
  plot.theme = self$plot.theme,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{embedding}}{A cell embedding to use (two-column data frame with rownames corresponding to cells) (default: stored embedding object)}

\item{\code{color.by}}{color cells by 'cell.groups', 'condition' or 'sample'. Overrides \code{groups} and \code{palette}. (default: NULL)}

\item{\code{plot.theme}}{plot theme to use (default: \code{self$plot.theme})}

\item{\code{...}}{other parameters passed to \link[sccore:embeddingPlot]{embeddingPlot}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
embedding plot as output by sccore::embeddingPlot
Estimate ontology terms based on DEs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntology"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntology}{}}}
\subsection{Method \code{estimateOntology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntology(
  type = c("GO", "DO", "GSEA"),
  name = NULL,
  de.name = "de",
  org.db,
  n.top.genes = 500,
  p.adj = 1,
  p.adjust.method = "BH",
  readable = TRUE,
  min.gs.size = 10,
  max.gs.size = 500,
  keep.gene.sets = FALSE,
  ignore.cache = NULL,
  de.raw = NULL,
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information (default="GO")}

\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{n.top.genes}}{numeric Number of top highest-expressed genes to consider (default=500)}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values. Please see DOSE package for more information (default="BH")}

\item{\code{readable}}{boolean Mapping gene ID to gene name (default=TRUE)}

\item{\code{min.gs.size}}{numeric Minimal geneset size, please see clusterProfiler package for more information (default=5)}

\item{\code{max.gs.size}}{numeric Minimal geneset size, please see clusterProfiler package for more information (default=5e2)}

\item{\code{verbose}}{boolean Whether to show progress}

\item{\code{n.cores}}{numeric Number of cores for parallelization}

\item{\code{...}}{further argument for ontology estimation. Pass \code{nPerm} with \code{type='GSEA'} to use fgseaSimple method}

\item{\code{min.genes}}{numeric Minimum number of input genes overlapping with ontologies (default=0)}

\item{\code{qvalue.cutoff}}{numeric Q value cutoff, please see clusterProfiler package for more information (default=0.2)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing a list of terms per ontology, and a data frame with merged results
Estimate ontology families based on ontology results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntologyFamilies"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntologyFamilies}{}}}
\subsection{Method \code{estimateOntologyFamilies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntologyFamilies(name = "GO", p.adj = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Type of ontology result: "GO", "GSEA", or "DO" (default="GO")}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of families and ontology data per cell type
Identify families containing a specific ontology term or ID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getFamiliesPerGO"></a>}}
\if{latex}{\out{\hypertarget{method-getFamiliesPerGO}{}}}
\subsection{Method \code{getFamiliesPerGO()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getFamiliesPerGO(
  name = "GO",
  go.term = NULL,
  go.id = NULL,
  common = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default="GO")}

\item{\code{go.term}}{Character vector with term description(s) (default=NULL)}

\item{\code{go.id}}{Character vector with ID(s) (default=NULL)}

\item{\code{common}}{boolean Only identify families with all the supplied terms or IDs (default = FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame
Bar plot of ontology terms per cell type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNumOntologyTermsPerType"></a>}}
\if{latex}{\out{\hypertarget{method-plotNumOntologyTermsPerType}{}}}
\subsection{Method \code{plotNumOntologyTermsPerType()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotNumOntologyTermsPerType(
  name = "GO",
  genes = "up",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default="GO")}

\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{q.value}}{numeric (default=0.2)}

\item{\code{min.genes}}{numeric (default=1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot a dotplot of ontology terms with adj. P values for a specific cell subgroup
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntology"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntology}{}}}
\subsection{Method \code{plotOntology()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntology(
  cell.type,
  name = "GO",
  plot = "dot",
  genes = c("up", "down", "all"),
  subtype = c("BP", "CC", "MF"),
  n = 20,
  p.adj = 0.05,
  min.genes = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types), "GO" or "DO" (default="GO")}

\item{\code{cell.subgroup}}{Cell group to plot}

\item{\code{log.colors}}{Use log10 p-values for coloring (default=FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot a heatmap of ontology P values per cell type
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyHeatmap}{}}}
\subsection{Method \code{plotOntologyHeatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmap(
  name = "GO",
  genes = "up",
  subtype = "BP",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  top.n = Inf,
  legend.position = "left",
  selection = "all",
  cluster = TRUE,
  cell.subgroups = NULL,
  row.order = TRUE,
  col.order = TRUE,
  max.log.p = 10,
  only.family.children = FALSE,
  description.regex = NULL,
  description.exclude.regex = NULL,
  clust.naming = "medoid",
  readjust.p = TRUE,
  p.adjust.method = "BH",
  palette = NULL,
  color.range = NULL,
  return.info = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{genes}}{Specify which genes to plot, can either be 'down' for downregulated genes, 'up' or 'all'
(default="up")}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{top.n}}{Number of terms to show (default=Inf)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{selection}}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type);
'common' (only show terms that are present in at least two cell types); 'all' (all ontology terms)
(default="all")}

\item{\code{cell.subgroups}}{Cell groups to plot (default=NULL). This affects only visualization, but not clustering.}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values. Please see DOSE package for more information (default="BH")}

\item{\code{color.range}}{vector with two values for min/max values of p-values}

\item{\code{...}}{parameters forwarded to \link{plotHeatmap}}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{clusters}}{Whether to show GO clusters or raw GOs (default=TRUE)}

\item{\code{cluster.name}}{Field with the results for GO clustering. Ignored if \code{clusters == FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot a heatmap of ontology p-values per cell type heatmap, collapsed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyHeatmapCollapsed"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyHeatmapCollapsed}{}}}
\subsection{Method \code{plotOntologyHeatmapCollapsed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmapCollapsed(
  name = "GO",
  genes = "up",
  subtype = "BP",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1,
  n = 20,
  legend.position = "left",
  selection = "all",
  max.log.p = 10,
  cluster = TRUE,
  cell.subgroups = NULL,
  palette = NULL,
  row.order = TRUE,
  col.order = TRUE,
  only.family.children = FALSE,
  readjust.p = TRUE,
  p.adjust.method = "BH",
  description.regex = NULL,
  description.exclude.regex = NULL,
  distance = "manhattan",
  clust.method = "complete",
  clust.naming = "consensus",
  n.words = 5,
  exclude.words = NULL,
  return.info = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{genes}}{Specify which genes to plot, can either be 'down' for downregulated genes, 'up' or 'all'
(default="up")}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{selection}}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type);
'common' (only show terms that are present in at least two cell types); 'all' (all ontology terms)
(default="all")}

\item{\code{cell.subgroups}}{Cell groups to plot (default=NULL). This affects only visualization, but not clustering.}

\item{\code{p.adjust.method}}{character string Method for calculating adjusted p-values. Please see DOSE package for more information (default="BH")}

\item{\code{...}}{parameters forwarded to \link{plotHeatmap}}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{top.n}}{Number of terms to show (default=Inf)}

\item{\code{clusters}}{Whether to show GO clusters or raw GOs (default=TRUE)}

\item{\code{cluster.name}}{Field with the results for GO clustering. Ignored if \code{clusters == FALSE}.}

\item{\code{color.range}}{vector with two values for min/max values of p-values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot correlation matrix for ontology terms between cell types
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologySimilarities"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologySimilarities}{}}}
\subsection{Method \code{plotOntologySimilarities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologySimilarities(
  name = "GO",
  genes = "up",
  p.adj = 0.05,
  q.value = 0.2,
  min.genes = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default="GO")}

\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{q.value}}{numeric (default=0.2)}

\item{\code{min.genes}}{numeric (default=1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot related ontologies in one hierarchical network plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyFamily"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyFamily}{}}}
\subsection{Method \code{plotOntologyFamily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyFamily(
  name = "GO",
  cell.type,
  family = NULL,
  genes = "up",
  subtype = "BP",
  plot.type = "complete",
  show.ids = FALSE,
  string.length = 14,
  legend.label.size = 1,
  legend.position = "topright",
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{family}}{Family within cell subtype to plot (numeric value)}

\item{\code{genes}}{Only for GO results: Direction of genes, must be "up", "down", or "all" (default: up)}

\item{\code{subtype}}{Only for GO results: Type of result, must be "BP", "MF", or "CC" (default: BP)}

\item{\code{plot.type}}{How much of the family network should be plotted. Can be "complete" (entire network), "dense" (show 1 parent for each significant term), or "minimal" (only show significant terms) (default: complete)}

\item{\code{show.ids}}{Whether to show ontology IDs instead of names (default: FALSE)}

\item{\code{string.length}}{Length of strings for wrapping in order to fit text within boxes (default: 14)}

\item{\code{legend.label.size}}{Size og legend labels (default: 1)}

\item{\code{legend.position}}{Position of legend (default: topright)}

\item{\code{verbose}}{Print messages (default: stored value)}

\item{\code{n.cores}}{Number of cores to use (default: stored value)}

\item{\code{type}}{Type of ontology result, i.e., GO, GSEA, or DO (default: GO)}

\item{\code{cell.subgroups}}{Cell subtype to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Rgraphviz object
Save ontology results as a table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveOntologyAsTable"></a>}}
\if{latex}{\out{\hypertarget{method-saveOntologyAsTable}{}}}
\subsection{Method \code{saveOntologyAsTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveOntologyAsTable(
  file,
  name = "GO",
  subtype = NULL,
  genes = NULL,
  p.adj = 0.05,
  sep = "\\t",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{character string File name passed to write.table(). Set to NULL to return the table instead of saving.}

\item{\code{name}}{chacter string Name of ontology result (default='GO')}

\item{\code{subtype}}{character string Only for GO results: Type of result to filter by, must be "BP", "MF", or "CC" (default: NULL)}

\item{\code{genes}}{Only for GO results: Direction of genes to filter by, must be "up", "down", or "all" (default: NULL)}

\item{\code{p.adj}}{numeric Cut-off for adjusted p-values (default=0.05)}

\item{\code{sep}}{Separator (default: tab)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table for import into text editor
Save family results as a table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveFamiliesAsTable"></a>}}
\if{latex}{\out{\hypertarget{method-saveFamiliesAsTable}{}}}
\subsection{Method \code{saveFamiliesAsTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveFamiliesAsTable(
  file,
  name = "GO",
  subtype = NULL,
  genes = NULL,
  p.adj = 0.05,
  sep = "\\t",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{character string File name passed to write.table(). Set to NULL to return the table instead of saving.}

\item{\code{name}}{chacter string Name of ontology result (default='GO')}

\item{\code{subtype}}{Only for GO results: Type of result to filter by, must be "BP", "MF", or "CC" (default=NULL)}

\item{\code{genes}}{Only for GO results: Direction of genes to filter by, must be "up", "down", or "all" (default=NULL)}

\item{\code{p.adj}}{Adjusted P to filter by (default=0.05)}

\item{\code{sep}}{Separator (default=tab)}

\item{\code{type}}{character string Type of ontology result, i.e., GO, GSEA, or DO (default='GO')}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
table for import into text editor
Plot the cell group sizes or proportions per sample
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellGroupSizes"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellGroupSizes}{}}}
\subsection{Method \code{plotCellGroupSizes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellGroupSizes(
  cell.groups = self$cell.groups,
  show.significance = FALSE,
  filter.empty.cell.types = TRUE,
  proportions = TRUE,
  palette = self$sample.groups.palette,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell names (default: stored vector)}

\item{\code{show.significance}}{whether to show statistical significance betwwen sample groups. wilcox.test was used; (\code{*} < 0.05; \verb{**} < 0.01; \verb{***} < 0.001)}

\item{\code{filter.empty.cell.types}}{boolean (default=TRUE)}

\item{\code{proportions}}{boolean Whether to plot proportions or absolute numbers (default=TRUE)}

\item{\code{palette}}{color palette to use for conditions (default: stored $sample.groups.palette)}

\item{\code{palette}}{(default=self$sample.group.palette)}

\item{\code{...}}{additional plot parameters, forwarded to \link{plotCountBoxplotsPerType}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot the cell group sizes or proportions per sample
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellGroupAbundanceVariation"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellGroupAbundanceVariation}{}}}
\subsection{Method \code{plotCellGroupAbundanceVariation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellGroupAbundanceVariation(
  cell.groups = self$cell.groups,
  type = "mad",
  rotate.xticks = TRUE,
  min.rel.abundance = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{(default=self$cell.groups)}

\item{\code{type}}{character string (default='mad')}

\item{\code{rotate.xticks}}{boolean (deafult=TRUE)}

\item{\code{min.rel.abundance}}{numeric (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
Plot compositions in CoDA space (PCA or CDA)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCodaSpace"></a>}}
\if{latex}{\out{\hypertarget{method-plotCodaSpace}{}}}
\subsection{Method \code{plotCodaSpace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCodaSpace(
  space = "CDA",
  cell.groups = self$cell.groups,
  font.size = 3,
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  palette = self$sample.groups.palette
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{space}}{either 'PCA' or 'CDA'}

\item{\code{cell.groups}}{vector Indicates cell groups with cell names (default: stored vector)}

\item{\code{font.size}}{numeric (default=3)}

\item{\code{cells.to.remain}}{(default=NULL)}

\item{\code{cells.to.remove}}{(default=NULL)}

\item{\code{samples.to.remove}}{(default=NULL)}

\item{\code{palette}}{(self$sample.groups.palette)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot contrast tree
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotContrastTree"></a>}}
\if{latex}{\out{\hypertarget{method-plotContrastTree}{}}}
\subsection{Method \code{plotContrastTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotContrastTree(
  cell.groups = self$cell.groups,
  palette = self$sample.groups.palette,
  name = "coda",
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  filter.empty.cell.types = TRUE,
  adjust.pvalues = TRUE,
  h.method = c("both", "up", "down"),
  reorder.tree = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{(default=self$cell.groups)}

\item{\code{palette}}{(default=self$sample.groups.palette)}

\item{\code{name}}{(default='coda')}

\item{\code{cells.to.remain}}{(default=NULL)}

\item{\code{cells.to.remove}}{(default=NULL)}

\item{\code{filter.empty.cell.types}}{boolean (default=TRUE)}

\item{\code{adjust.pvalues}}{boolean (default=TRUE)}

\item{\code{h.method}}{(default=c('both', 'up', 'down'))}

\item{\code{reorder.tree}}{boolean (default=TRUE)}

\item{\code{...}}{additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot composition similarity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCompositionSimilarity"></a>}}
\if{latex}{\out{\hypertarget{method-plotCompositionSimilarity}{}}}
\subsection{Method \code{plotCompositionSimilarity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCompositionSimilarity(
  cell.groups = self$cell.groups,
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  palette = brewerPalette("YlOrRd", rev = FALSE),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector Indicates cell groups with cell names (default: stored vector)}

\item{\code{palette}}{(default=brewerPalette("YlOrRd", rev=FALSE))}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate cell loadings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellLoadings}{}}}
\subsection{Method \code{estimateCellLoadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellLoadings(
  n.boot = 1000,
  ref.cell.type = NULL,
  name = "coda",
  n.seed = 239,
  cells.to.remove = NULL,
  cells.to.remain = NULL,
  samples.to.remove = NULL,
  filter.empty.cell.types = TRUE,
  n.cores = self$n.cores,
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.boot}}{numeric (default=1000)}

\item{\code{ref.cell.type}}{(default=NULL)}

\item{\code{name}}{character string (default='coda')}

\item{\code{n.seed}}{numeric (default=239)}

\item{\code{cells.to.remove}}{vector (default=NULL)}

\item{\code{cells.to.remain}}{vector (default=NULL)}

\item{\code{samples.to.remove}}{vector (default=NULL)}

\item{\code{filter.empty.cell.types}}{vector (default=NULL)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{verbose}}{boolean (default=self$verbose)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
resulting cell loadings
Estimate Ga partitions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGaPartition"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGaPartition}{}}}
\subsection{Method \code{estimateGaPartition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGaPartition(
  cells.to.remain = NULL,
  cells.to.remove = NULL,
  samples.to.remove = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells.to.remain}}{vector (default=NULL)}

\item{\code{cells.to.remove}}{vector (default=NULL)}

\item{\code{samples.to.remove}}{vector (default=NULL)}

\item{\code{...}}{additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
estimated Ga partitions
Plot Loadings
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellLoadings"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellLoadings}{}}}
\subsection{Method \code{plotCellLoadings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellLoadings(
  alpha = 0.01,
  palette = self$cell.groups.palette,
  font.size = NULL,
  name = "coda",
  ordering = "pvalue",
  show.pvals = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{numeric (default=0.01)}

\item{\code{palette}}{palette specification for cell types (default: stored $cell.groups.palette)}

\item{\code{font.size}}{(default=NULL)}

\item{\code{name}}{character string (default='pvalue')}

\item{\code{ordering}}{character string (default='pvalue')}

\item{\code{show.pvals}}{boolean (default=TRUE)}

\item{\code{...}}{additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate Wilcoxon signed-rank test
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateWilcoxonTest"></a>}}
\if{latex}{\out{\hypertarget{method-estimateWilcoxonTest}{}}}
\subsection{Method \code{estimateWilcoxonTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateWilcoxonTest(
  cell.groups = self$cell.groups,
  cells.to.remove = NULL,
  cells.to.remain = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{vector (default=NULL)}

\item{\code{cells.to.remove}}{vector (default=NULL)}

\item{\code{cells.to.remain}}{vector (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
p-values
Estimate cell density in giving embedding
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-estimateCellDensity}{}}}
\subsection{Method \code{estimateCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateCellDensity(
  bins = 400,
  method = "kde",
  name = "cell.density",
  beta = 30,
  estimate.variation = TRUE,
  sample.groups = self$sample.groups,
  verbose = self$verbose,
  n.cores = self$n.cores,
  bandwidth = 0.05,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{numeric Number of bins for density estimation (default=400)}

\item{\code{method}}{character string Density estimation method, graph: graph smooth based density estimation. kde: embedding grid based density  estimation. (default: 'kde')}

\item{\code{name}}{slot in which to save the results (default='cell.density')}

\item{\code{beta}}{numeric Smoothing strength parameter of the \link[sccore:heatFilter]{heatFilter} for graph based cell density (default=30)}

\item{\code{estimate.variation}}{boolean (default=TRUE)}

\item{\code{sample.groups}}{(default=self$sample.groups)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{bandwidth}}{numeric KDE bandwidth multiplier (default=0.5). The full bandwidth is estimated by multiplying this value on the difference between 90\% and 10\%
of the corresponding embedding dimension. Set it to NULL to use \link[MASS:bandwidth.nrd]{bandwidth.nrd} estimator. (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
estimated cell densities
Plot cell density depending on the method that was used for estimating \code{cao$test.resulst[[name]]}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellDensity}{}}}
\subsection{Method \code{plotCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensity(
  show.grid = TRUE,
  add.points = TRUE,
  size = 0.1,
  show.legend = FALSE,
  palette = NULL,
  point.col = "#313695",
  contours = NULL,
  contour.color = "black",
  contour.conf = "10\%",
  name = "cell.density",
  show.cell.groups = TRUE,
  cell.groups = self$cell.groups,
  font.size = c(2, 4),
  color.range = c(0, "99\%"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add.points}}{boolean Add points to cell density figure (default=TRUE)}

\item{\code{size}}{numeric (default=0.1)}

\item{\code{show.legend}}{boolean (default=FALSE)}

\item{\code{palette}}{(default=NULL)}

\item{\code{point.col}}{character string (default='#313695')}

\item{\code{contours}}{specify cell types for contour, multiple cell types are also supported (default=NULL)}

\item{\code{contour.color}}{color for contour line (default='black')}

\item{\code{contour.conf}}{confidence interval of contour  (default='10\%')}

\item{\code{name}}{slot in which to saved results from estimateCellDensity (default='cell.density')}

\item{\code{show.cell.groups}}{boolean (default=TRUE)}

\item{\code{cell.groups}}{(default=self$cell.groups)}

\item{\code{font.size}}{(default=c(2, 4))}

\item{\code{color.range}}{(default=c(0, "99\%"))}

\item{\code{...}}{plot style parameters forwarded to \link[sccore:styleEmbeddingPlot]{sccore::styleEmbeddingPlot}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot cell density variation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellDensityVariation"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellDensityVariation}{}}}
\subsection{Method \code{plotCellDensityVariation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellDensityVariation(
  type = "mad",
  plot.type = "embedding",
  name = "cell.density",
  cutoff = NULL,
  condition = c("both", "ref", "target"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{character string (default='mad')}

\item{\code{plot.type}}{character string (default='embedding')}

\item{\code{name}}{character string (default='cell.density')}

\item{\code{cutoff}}{(default=NULL)}

\item{\code{condition}}{character vector (default=c('both', 'ref', 'target'))}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
Estimate differential cell density
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateDiffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-estimateDiffCellDensity}{}}}
\subsection{Method \code{estimateDiffCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateDiffCellDensity(
  type = "permutation",
  adjust.pvalues = NULL,
  name = "cell.density",
  n.permutations = 400,
  smooth = TRUE,
  verbose = self$verbose,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{method to calculate differential cell density; permutation, t.test, wilcox or subtract (target subtract ref density);}

\item{\code{adjust.pvalues}}{whether to adjust Z-scores for multiple comparison using BH method (default: FALSE for type='sutract', TRUE for everything else)}

\item{\code{name}}{slot with results from estimateCellDensity. New results will be appended there. (Default: 'cell.density')}

\item{\code{n.permutations}}{numeric (default=400)}

\item{\code{smooth}}{boolean (default=TRUE)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{...}}{additional arguments to the function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
estimated differential cell densities
Estimate differential cell density
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDiffCellDensity"></a>}}
\if{latex}{\out{\hypertarget{method-plotDiffCellDensity}{}}}
\subsection{Method \code{plotDiffCellDensity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDiffCellDensity(
  type = NULL,
  name = "cell.density",
  size = 0.2,
  palette = NULL,
  adjust.pvalues = NULL,
  contours = NULL,
  contour.color = "black",
  contour.conf = "10\%",
  plot.na = FALSE,
  color.range = NULL,
  mid.color = "gray95",
  scale.z.palette = adjust.pvalues,
  min.z = qnorm(0.9),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{method to calculate differential cell density; t.test, wilcox or subtract (target subtract ref density);}

\item{\code{name}}{slot with results from estimateCellDensity. New results will be appended there. (Default: 'cell.density')}

\item{\code{size}}{numeric (default=0.2)}

\item{\code{palette}}{color palette, default is c('blue','white','red')}

\item{\code{adjust.pvalues}}{(default=NULL)}

\item{\code{contours}}{specify cell types for contour, multiple cell types are also supported (default: NULL)}

\item{\code{contour.color}}{color for contour line (default: 'black')}

\item{\code{contour.conf}}{confidence interval of contour (default: '10\%')}

\item{\code{plot.na}}{boolean (default=FALSE)}

\item{\code{color.range}}{(default=NULL)}

\item{\code{mid.color}}{character string (default='gray95')}

\item{\code{scale.z.palette}}{(default=adjust.pvalues)}

\item{\code{min.z}}{(default=qnorm(0.9))}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
Plot inter-sample expression distance. The inputs to this function are the results from cao$estimateExpressionShiftMagnitudes()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionDistance"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionDistance}{}}}
\subsection{Method \code{plotExpressionDistance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionDistance(
  name = "expression.shifts",
  joint = FALSE,
  palette = self$sample.groups.palette,
  show.significance = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Test results to plot (default=expression.shifts)}

\item{\code{joint}}{boolean Whether to show joint boxplot with the expression distance weighed by the sizes of cell types (default: TRUE), or show distances for each individual cell type}

\item{\code{palette}}{(default=self$sample.groups.palette)}

\item{\code{show.significance}}{whether to show statistical significance between sample groups. wilcox.test was used; (\code{*} < 0.05; \verb{**} < 0.01; \verb{***} < 0.001)}

\item{\code{...}}{other plot parameters, forwarded to \link{plotCountBoxplotsPerType}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Plot inter-sample expression distance. The inputs to this function are the results from cao$estimateExpressionShiftMagnitudes()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSampleDistanceMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-getSampleDistanceMatrix}{}}}
\subsection{Method \code{getSampleDistanceMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getSampleDistanceMatrix(
  space = c("expression.shifts", "coda", "pseudo.bulk"),
  cell.type = NULL,
  dist = NULL,
  name = NULL,
  verbose = self$verbose,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{space}}{(default=c('expression.shifts', 'coda', 'pseudo.bulk'))}

\item{\code{cell.type}}{(default=NULL)}

\item{\code{dist}}{(default=NULL)}

\item{\code{name}}{(default=NULL)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{...}}{additional arguments}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Project samples to 2D space with MDS. Plots results from cao$estimateExpressionShiftMagnitudes() or cao$estimateCellLoadings()
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSampleDistances"></a>}}
\if{latex}{\out{\hypertarget{method-plotSampleDistances}{}}}
\subsection{Method \code{plotSampleDistances()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotSampleDistances(
  space = "expression.shifts",
  method = "MDS",
  dist = NULL,
  name = NULL,
  cell.type = NULL,
  palette = NULL,
  show.sample.size = FALSE,
  sample.colors = NULL,
  color.title = NULL,
  title = NULL,
  n.permutations = 2000,
  show.pvalues = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{space}}{character string "expression.shifts" Results from cao$estimateExpressionShiftMagnitudes(); CDA- cell composition shifts result from cao$estimateCellLoadings(); sudo.bulk- expression distance of sudo bulk}

\item{\code{method}}{character string "MDS"}

\item{\code{dist}}{'cor' - correlation distance, 'l1' - manhattan distance or 'l2' - euclidean (default correlation distance)}

\item{\code{name}}{(default=NULL)}

\item{\code{cell.type}}{If a name of a cell type is specified, the sample distances will be assessed based on this cell type alone. Otherwise (cell.type=NULL, default), sample distances will be estimated as an average distance across all cell types (weighted by the minimum number of cells of that cell type between any two samples being compared)}

\item{\code{palette}}{a set of colors to use for conditions (default: stored $sample.groups.palette)}

\item{\code{show.sample.size}}{make point size proportional to the log10 of the number of cells per sample (default: FALSE)}

\item{\code{sample.colors}}{(default=NULL)}

\item{\code{color.title}}{(default=NULL)}

\item{\code{title}}{(default=NULL)}

\item{\code{n.permutations}}{numeric (default=2000)}

\item{\code{show.pvalues}}{boolean (default=FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
Estimate metadata separation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateMetadataSeparation"></a>}}
\if{latex}{\out{\hypertarget{method-estimateMetadataSeparation}{}}}
\subsection{Method \code{estimateMetadataSeparation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateMetadataSeparation(
  sample.meta,
  space = "expression.shifts",
  dist = NULL,
  space.name = NULL,
  name = "metadata.separation",
  n.permutations = 5000,
  trim = 0.05,
  show.warning = TRUE,
  verbose = self$verbose,
  n.cores = self$n.cores,
  adjust.pvalues = TRUE,
  p.adjust.method = "BH",
  pvalue.cutoff = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample.meta}}{sample metadata}

\item{\code{space}}{(default="expression shifts")}

\item{\code{dist}}{(default=NULL)}

\item{\code{space.name}}{(default=NULL)}

\item{\code{name}}{(default="metadata.separation")}

\item{\code{n.permutations}}{(default=5000)}

\item{\code{trim}}{(default=0.05)}

\item{\code{show.warning}}{boolean (default=TRUE)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{adjust.pvalues}}{boolean (default=TRUE)}

\item{\code{p.adjust.method}}{character string (default="BH")}

\item{\code{pvalue.cutoff}}{numeric (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
results
Estimate differential expression Z-scores between two conditions per individual cell
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeDE"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeDE}{}}}
\subsection{Method \code{estimateClusterFreeDE()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeDE(
  n.top.genes = Inf,
  genes = NULL,
  max.z = 20,
  min.expr.frac = 0.01,
  min.n.samp.per.cond = 2,
  min.n.obs.per.samp = 2,
  robust = FALSE,
  norm.both = TRUE,
  adjust.pvalues = FALSE,
  smooth = TRUE,
  wins = 0.01,
  n.permutations = 200,
  lfc.pseudocount = 1e-05,
  min.edge.weight = 0.6,
  verbose = self$verbose,
  n.cores = self$n.cores,
  name = "cluster.free.de"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{(default=Inf)}

\item{\code{genes}}{(default=NULL)}

\item{\code{max.z}}{z-score value to winsorize the estimates for reducing impact of outliers. Default: 20.}

\item{\code{min.expr.frac}}{minimal fraction of cell expressing a gene for estimating z-scores for it. Default: 0.001.}

\item{\code{min.n.samp.per.cond}}{minimul number of samples per condition for estimating z-scores (default: 2)}

\item{\code{min.n.obs.per.samp}}{minimul number of cells per samples for estimating z-scores (default: 2)}

\item{\code{robust}}{whether to use median estimates instead of mean. Using median is more robust,
but greatly increase the number of zeros in the data, leading to bias towards highly-express genes. (Default: FALSE)}

\item{\code{norm.both}}{boolean (default=TRUE)}

\item{\code{adjust.pvalues}}{boolean (default=FALSE)}

\item{\code{wins}}{numeric (default=0.01)}

\item{\code{n.permutations}}{numeric (default=200)}

\item{\code{lfc.pseudocount}}{pseudocount value for estimation of log2(fold-change)}

\item{\code{min.edge.weight}}{numeric (default=0.6)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{name}}{character string (default='cluster.free.de')}

\item{\code{smoooth}}{boolean (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list with sparce matrices containing various DE metrics with genes as columns and cells as rows:
\itemize{
\item \code{z}: DE Z-scores
\item \code{reference} mean or median expression in reference samples
\item \code{target} mean or median expression in target samples
\item \code{lfc}: log2(fold-change) of expression
Cells that have only one condition in their expression neighborhood have NA Z-scores for all genes.
Results are also stored in the \code{cluster.free.de} field.
Get most changed genes
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-getMostChangedGenes}{}}}
\subsection{Method \code{getMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getMostChangedGenes(
  n,
  method = c("z", "z.adj", "lfc"),
  min.z = 0.5,
  min.lfc = 1,
  max.score = 20,
  cell.subset = NULL,
  excluded.genes = NULL,
  included.genes = NULL,
  name = "cluster.free.de"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{(default=c("z", "z.adj", "lfc"))}

\item{\code{min.z}}{numeric (default=0.5)}

\item{\code{min.lfc}}{numeric (default=1)}

\item{\code{max.score}}{numeric (default=20)}

\item{\code{cell.subset}}{(default=NULL)}

\item{\code{excluded.genes}}{(default=NULL)}

\item{\code{included.genes}}{(default=NULL)}

\item{\code{name}}{(default="cluster.free.de")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
results
Estimate Cluster-free Expression Shift
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-estimateClusterFreeExpressionShifts}{}}}
\subsection{Method \code{estimateClusterFreeExpressionShifts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateClusterFreeExpressionShifts(
  n.top.genes = 3000,
  gene.selection = "z",
  name = "cluster.free.expr.shifts",
  min.n.between = 2,
  min.n.within = max(min.n.between, 1),
  min.expr.frac = 0,
  min.n.obs.per.samp = 3,
  normalize.both = FALSE,
  dist = "cor",
  log.vectors = (dist != "js"),
  wins = 0.025,
  genes = NULL,
  n.permutations = 500,
  verbose = self$verbose,
  n.cores = self$n.cores,
  min.edge.weight = 0,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{number of top genes for the distance estimation (default: 3000)}

\item{\code{gene.selection}}{character string (default="z")}

\item{\code{name}}{character string (default="cluster.free.expr.shifts")}

\item{\code{min.n.between}}{minimal number of pairs between condition for distance estimation (default: 2)}

\item{\code{min.n.within}}{minimal number of pairs within one condition for distance estimation (default: \code{min.n.between})}

\item{\code{min.n.obs.per.samp}}{minimal number of cells per sample for using it in distance estimation (default: 3)}

\item{\code{normalize.both}}{whether to normalize results relative to distances within both conditions (TRUE) or only to the control (FALSE)}

\item{\code{dist}}{distance measure. Options: "cor" (correlation), "cosine" or "js" (JensenShannon)}

\item{\code{log.vectors}}{whether to use log10 on the normalized expression before estimating the distance.
In most cases, must be TRUE for "cosine" and "cor" distances and always must be FALSE for "js". (default: \code{dist != 'js'})}

\item{\code{wins}}{numeric (default=0.025)}

\item{\code{genes}}{(default=NULL)}

\item{\code{n.permutations}}{numeric (default=500)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{min.edge.weight}}{numeric (default=0.0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of cluster-free expression shifts per cell. Values above 1 correspond to difference between conditions.
Results are also stored in the \code{cluster.free.expr.shifts} field.
Performs graph smoothing of the cluster-free DE Z-scores
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothClusterFreeZScores"></a>}}
\if{latex}{\out{\hypertarget{method-smoothClusterFreeZScores}{}}}
\subsection{Method \code{smoothClusterFreeZScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$smoothClusterFreeZScores(
  n.top.genes = 1000,
  smoothing = 20,
  filter = NULL,
  z.adj = FALSE,
  gene.selection = ifelse(z.adj, "z.adj", "z"),
  excluded.genes = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric (default=1000)}

\item{\code{smoothing}}{\code{beta} parameter of the \link[sccore:heatFilter]{heatFilter}. (default=20)}

\item{\code{filter}}{graph filter function. (default=NULL)}

\item{\code{z.adj}}{boolean (default=FALSE)}

\item{\code{gene.selection}}{(default=ifelse(z.adj, "z.adj", "z"))}

\item{\code{excluded.genes}}{List of genes to exclude during estimation. For example, a list of mitochondrial genes.}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{name}}{character string (default='cluster.free.de')}

\item{\code{...}}{parameters forwarded to \link[sccore:smoothSignalOnGraph]{smoothSignalOnGraph}}

\item{\code{exluded.genes}}{(default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sparse matrix of smoothed Z-scores. Results are also stored in the \code{cluster.free.de$z.smoothed} field.
Estimate Gene Programs based on cluster-free Z-scores on a subsample of
cells using \link[fabia:fabia]{fabia}. # TODO: update it
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateGenePrograms"></a>}}
\if{latex}{\out{\hypertarget{method-estimateGenePrograms}{}}}
\subsection{Method \code{estimateGenePrograms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateGenePrograms(
  method = c("pam", "leiden", "fabia"),
  n.top.genes = Inf,
  genes = NULL,
  n.programs = 15,
  z.adj = FALSE,
  gene.selection = ifelse(z.adj, "z.adj", "z"),
  smooth = TRUE,
  abs.scores = FALSE,
  name = "gene.programs",
  cell.subset = NULL,
  n.cores = self$n.cores,
  verbose = self$verbose,
  max.z = 5,
  min.z = 0.5,
  min.change.frac = 0.01,
  de.name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{(default=Inf)}

\item{\code{genes}}{(default=NULL)}

\item{\code{n.programs}}{maximal number of gene programs to find (parameter \code{p} for fabia). (default=15)}

\item{\code{z.adj}}{boolean (default=FALSE)}

\item{\code{gene.selection}}{(default=ifelse(z.adj, "z.adj", "z"))}

\item{\code{smooth}}{boolean (default=TRUE)}

\item{\code{abs.scores}}{boolean (default=FALSE)}

\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{cell.subset}}{(default=NULL)}

\item{\code{n.cores}}{numeric (default=self$n.cores)}

\item{\code{verbose}}{boolean (default=self$verbose)}

\item{\code{max.z}}{numeric (default=5)}

\item{\code{min.z}}{numeric (default=0.5)}

\item{\code{min.change.frac}}{numeric (default=0.01)}

\item{\code{de.name}}{character string (default="cluster.free.de")}

\item{\code{...}}{keyword arguments forwarded to estimateGenePrograms}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list includes:
\itemize{
\item \code{fabia}: \link[fabia:Factorization]{fabia::Factorization} object, result of the
\link[fabia:fabia]{fabia::fabia} call
\item \code{sample.ids}: ids of the subsampled cells used for fabia estimates
\item \code{scores.exact}: vector of fabia estimates of gene program scores per cell. Estimated only for the
subsampled cells.
\item \code{scores.approx}: vector of approximate gene program scores, estimated for all cells in the dataset
\item \code{loadings}: matrix with fabia gene loadings per program
\item \code{gene.scores}: list of vectors of gene scores per program. Contains only genes, selected for
the program usin fabia biclustering.
\item \code{bi.clusts} fabia biclustering information, result of the \link[fabia:extractBic]{fabia::extractBic} call
Plot gene program scores
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgramScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgramScores}{}}}
\subsection{Method \code{plotGeneProgramScores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgramScores(
  name = "gene.programs",
  prog.ids = NULL,
  build.panel = TRUE,
  nrow = NULL,
  adj.list = NULL,
  legend.title = "Score",
  palette = NULL,
  min.genes.per.prog = 10,
  color.range = c("0.5\%", "99.5\%"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string (default='gene.programs')}

\item{\code{prog.ids}}{(default=NULL)}

\item{\code{build.panel}}{boolean (default=TRUE)}

\item{\code{nrow}}{(default=NULL)}

\item{\code{adj.list}}{(default=NULL)}

\item{\code{legend.title}}{character string (default="Score")}

\item{\code{palette}}{(default=NULL)}

\item{\code{min.genes.per.prog}}{numeric (default=10)}

\item{\code{color.range}}{(default=c("0.5\%", "99.5\%"))}

\item{\code{...}}{additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Plot gene program genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneProgramGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneProgramGenes}{}}}
\subsection{Method \code{plotGeneProgramGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneProgramGenes(
  program.id,
  name = "gene.programs",
  ordering = c("similarity", "loading"),
  max.genes = 9,
  plots = "z.adj",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string Field name where the test results are stored}

\item{\code{Plot}}{cluster-free expression shift z-scores}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotClusterFreeExpressionShifts"></a>}}
\if{latex}{\out{\hypertarget{method-plotClusterFreeExpressionShifts}{}}}
\subsection{Method \code{plotClusterFreeExpressionShifts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotClusterFreeExpressionShifts(
  cell.groups = self$cell.groups,
  smooth = TRUE,
  plot.na = FALSE,
  name = "cluster.free.expr.shifts",
  scale.z.palette = TRUE,
  min.z = qnorm(0.9),
  color.range = c("0", "97.5\%"),
  alpha = 0.2,
  font.size = c(3, 5),
  adj.list = NULL,
  palette = brewerPalette("YlOrRd", rev = FALSE),
  build.panel = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{cell type labels. Set to NULL if it shouldn't be shown}

\item{\code{smooth}}{boolean (default=TRUE)}

\item{\code{plot.na}}{boolean (default=FALSE)}

\item{\code{name}}{(default='cluster.free.expr.shifts')}

\item{\code{scale.z.palette}}{boolean (default=TRUE)}

\item{\code{min.z}}{(default=qnorm(0.9))}

\item{\code{color.range}}{(default=c("0", "97.5\%"))}

\item{\code{alpha}}{numeric (default=0.2)}

\item{\code{font.size}}{size range for cell type labels}

\item{\code{palette}}{(default=brewerPalette("YlOrRd", rev=FALSE))}

\item{\code{build.panel}}{boolean (default=TRUE)}

\item{\code{...}}{parameters forwarded to \link[sccore:embeddingPlot]{embeddingPlot}
Plot most changed genes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMostChangedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotMostChangedGenes}{}}}
\subsection{Method \code{plotMostChangedGenes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotMostChangedGenes(
  n.top.genes,
  method = "z",
  min.z = 0.5,
  min.lfc = 1,
  max.score = 20,
  cell.subset = NULL,
  excluded.genes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n.top.genes}}{numeric}

\item{\code{method}}{character string (default='z')}

\item{\code{min.lfc}}{numeric (default=1)}

\item{\code{max.score}}{numeric (default=20)}

\item{\code{cell.subset}}{(default=NULL)}

\item{\code{excluded.genes}}{(default=NULL)}

\item{\code{...}}{additional parameters input to self$plotGeneExpressionComparison()
Plot gene expression comparison}

\item{\code{min}}{numeric (default=0.5)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotGeneExpressionComparison"></a>}}
\if{latex}{\out{\hypertarget{method-plotGeneExpressionComparison}{}}}
\subsection{Method \code{plotGeneExpressionComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotGeneExpressionComparison(
  genes = NULL,
  scores = NULL,
  max.expr = "97.5\%",
  plots = c("z.adj", "z", "expression"),
  min.z = qnorm(0.9),
  max.z = 4,
  max.z.adj = NULL,
  max.lfc = 3,
  smoothed = FALSE,
  gene.palette = dark.red.palette,
  z.palette = NULL,
  z.adj.palette = z.palette,
  lfc.palette = NULL,
  scale.z.palette = TRUE,
  plot.na = -1,
  adj.list = NULL,
  build.panel = TRUE,
  nrow = 1,
  cell.subset = NULL,
  groups = NULL,
  subgroups = NULL,
  keep.limits = NULL,
  name = "cluster.free.de",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{(default=NULL)}

\item{\code{scores}}{(default=NULL)}

\item{\code{max.expr}}{character string (default="97.5\%")}

\item{\code{plots}}{(default=c("z.adj", "z", "expression"))}

\item{\code{min.z}}{(default=qname(0.9))}

\item{\code{max.z}}{numeric (default=4)}

\item{\code{max.z.adj}}{(default=NULL)}

\item{\code{max.lfc}}{numeric (default=3)}

\item{\code{smoothed}}{boolean (default=FALSE)}

\item{\code{gene.palette}}{(default=dark.red.palette)}

\item{\code{z.palette}}{(default=NULL)}

\item{\code{z.adj.palette}}{(default=z.palette)}

\item{\code{lfc.palette}}{(default=NULL)}

\item{\code{scale.z.palette}}{boolean (default=TRUE)}

\item{\code{plot.na}}{(default=-1)}

\item{\code{adj.list}}{(default=NULL)}

\item{\code{build.panel}}{boolean (default=TRUE)}

\item{\code{nrow}}{(default=1)}

\item{\code{cell.subset}}{(default=NULL)}

\item{\code{groups}}{(default=NULL)}

\item{\code{subgroups}}{(default=NULL)}

\item{\code{keep.limits}}{(default=NULL)}

\item{\code{name}}{character string Field name where the test results are stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list
Get condition per cell
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConditionPerCell"></a>}}
\if{latex}{\out{\hypertarget{method-getConditionPerCell}{}}}
\subsection{Method \code{getConditionPerCell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getConditionPerCell()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
conditions per cell
Get joint count matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getJointCountMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-getJointCountMatrix}{}}}
\subsection{Method \code{getJointCountMatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getJointCountMatrix(force = FALSE, raw = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{boolean (default=FALSE)}

\item{\code{raw}}{boolean (default=TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
joint count matrix
Get GO environment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getGOEnvironment"></a>}}
\if{latex}{\out{\hypertarget{method-getGOEnvironment}{}}}
\subsection{Method \code{getGOEnvironment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getGOEnvironment(org.db, verbose = FALSE, ignore.cache = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org.db}}{object of class OrgDB from Bioconductor (e.g. org.Hs.eg.db)}

\item{\code{verbose}}{boolean (default=FALSE)}

\item{\code{ignore.cache}}{(default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GO environment
Extract coda data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
