% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
Cacoa R6 class

Cacoa R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{number of cores}

\item{\code{verbose}}{print diagnostic messages}

\item{\code{test.results}}{results of the estimations, ready to use}

\item{\code{cache}}{intermediate results of the estimations, which can be used during some other computations}

\item{\code{data.object}}{the main object storing data (Conos or Seurat)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample}

\item{\code{cell.groups}}{named factor with cell names with cluster per cell}

\item{\code{sample.per.cell}}{named factor with cell names}

\item{\code{ref.level}}{reference level for sample.group vector}

\item{\code{target.level}}{target/disease level for sample.group vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-estimateExpressionShiftZScores}{\code{Cacoa$estimateExpressionShiftZScores()}}
\item \href{#method-plotExpressionShiftZScores}{\code{Cacoa$plotExpressionShiftZScores()}}
\item \href{#method-getPerCellTypeDE}{\code{Cacoa$getPerCellTypeDE()}}
\item \href{#method-plotDEGenes}{\code{Cacoa$plotDEGenes()}}
\item \href{#method-prepareOnthologyData}{\code{Cacoa$prepareOnthologyData()}}
\item \href{#method-plotFilteredDEGenes}{\code{Cacoa$plotFilteredDEGenes()}}
\item \href{#method-estimateOnthology}{\code{Cacoa$estimateOnthology()}}
\item \href{#method-plotOnthologyDistribution}{\code{Cacoa$plotOnthologyDistribution()}}
\item \href{#method-plotOnthologyTerms}{\code{Cacoa$plotOnthologyTerms()}}
\item \href{#method-plotOnthologyHeatmap}{\code{Cacoa$plotOnthologyHeatmap()}}
\item \href{#method-plotOnthologyCorrelations}{\code{Cacoa$plotOnthologyCorrelations()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  ref.level = NULL,
  target.level = NULL,
  n.cores = 1,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
Calculate expression shift magnitudes of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  sample.groups = self$sample.groups,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "expression.shifts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{dist}}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance (default="JS")}

\item{\code{within.group.normalization}}{Normalize the shift magnitude by the mean magnitude of within-group variation (default=T)}

\item{\code{valid.comparisons}}{A logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons (default=NULL)}

\item{\code{n.cells}}{Number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{\code{n.top.genes}}{Number of top highest-expressed genes to consider (default: all genes)}

\item{\code{n.subsamples}}{Number of samples to draw (default=100)}

\item{\code{min.cells}}{Minimum number of cells per cluster/per sample to be included in the analysis (default=10)}

\item{\code{sample.groups}}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{Number of cores (default: stored integer)}

\item{\code{verbose}}{Print progress}

\item{\code{name}}{Test name (default="expression.shifts")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list include \itemize{
  \item{df: a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells}
  \item{ctdml: raw list of `n.subsamples` sampled distance matrices (cells were subsampled)}
  \item{sample.groups: same as the provided variable}
  \item{valid.comparisons: a matrix of valid comparisons (in this case all should be valid, since we're not restricting samples that should be compared)}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  size.norm = F,
  normalized.distance = F,
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{size.norm}}{Plot size normalized results. Requires cell.groups, and sample.per.cell (default=F)}

\item{\code{normalized.distance}}{Plot the absolute median distance (default=F)}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default: stored vector)}

\item{\code{label}}{Plot labels on size normalized plots (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftZScores}{}}}
\subsection{Method \code{estimateExpressionShiftZScores()}}{
Calculate expression shift Z scores of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftZScores(
  cell.groups = self$cell.groups,
  ref.level = self$ref.level,
  sample.groups = self$sample.groups,
  sample.per.cell = self$sample.per.cell,
  n.od.genes = 1000,
  n.pcs = 100,
  pca.maxit = 1000,
  ignore.cache = F,
  name = "expression.z.scores"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{ref.level}}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{\code{sample.groups}}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{sample.per.cell}}{(default: stored vector)}

\item{\code{n.od.genes}}{(default=1000)}

\item{\code{n.pcs}}{(default=100)}

\item{\code{pca.maxit}}{(default=1000)}

\item{\code{ignore.cache}}{(default=F)}

\item{\code{name}}{(default="expression.z.scores")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftZScores}{}}}
\subsection{Method \code{plotExpressionShiftZScores()}}{
Plot results from cao$estimateExpressionShiftZScores()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftZScores(
  type.order = NULL,
  name = "expression.z.scores"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type.order}}{(default=NULL)}

\item{\code{name}}{Test results to plot (default=expression.z.shifts)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-getPerCellTypeDE}{}}}
\subsection{Method \code{getPerCellTypeDE()}}{
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$getPerCellTypeDE(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  independent.filtering = FALSE,
  cluster.sep.chr = "<!!>",
  return.matrix = F,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=NULL)}

\item{\code{sample.groups}}{a list of two character vector specifying the app groups to compare (default=NULL)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{n.cores}}{Number of cores (default=1)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=F)}

\item{\code{min.cell.count}}{(default=10)}

\item{\code{independent.filtering}}{independentFiltering for DESeq2 (default=F)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{return.matrix}}{Return merged matrix of results (default=F)}

\item{\code{verbose}}{Show progress (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEGenes}{}}}
\subsection{Method \code{plotDEGenes()}}{
Plot number of significant DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEGenes(
  de.raw = self$test.results$de$results,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T,
  p.adjust.cutoff = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.raw}}{List with differentially expressed genes per cell group (default: stored list, results from getPerCellTypeDE)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{p.adjust.cutoff}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepareOnthologyData"></a>}}
\if{latex}{\out{\hypertarget{method-prepareOnthologyData}{}}}
\subsection{Method \code{prepareOnthologyData()}}{
Filter and prepare DE genes for onthology calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$prepareOnthologyData(
  de.raw = self$test.results$de$results,
  cell.groups = self$cell.groups,
  universe = self$test.results$de$universe,
  org = "human",
  n.top.genes = Inf,
  stat.cutoff = 3,
  transposed = T,
  verbose = T,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.raw}}{Differentially expressed genes per cell group, results from getPerCellTypeDE (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{universe}}{All measured genes (default: stored vector)}

\item{\code{org}}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{\code{n.top.genes}}{Number of genes to use for analysis (default=Inf)}

\item{\code{stat.cutoff}}{Cutoff for filtering highly-expressed DE genes (default=3)}

\item{\code{transposed}}{Whether count matrices should be transposed (default=T)}

\item{\code{verbose}}{Print progress (default=T)}

\item{\code{n.cores}}{Number of cores to use (default: stored integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFilteredDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotFilteredDEGenes}{}}}
\subsection{Method \code{plotFilteredDEGenes()}}{
Plot number of highly-expressed DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotFilteredDEGenes(
  de.filter = self$test.results$onthology$de.filter,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.filter}}{Filtered DE genes, results from prepareOnthologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOnthology"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOnthology}{}}}
\subsection{Method \code{estimateOnthology()}}{
Calculate onthologies based on DEs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOnthology(
  type = NULL,
  org = "human",
  de.gene.ids = self$test.results$onthology$de.gene.ids,
  universe = self$test.results$onthology$universe,
  go.environment = self$test.results$go$go.environment,
  p.adj = 0.05,
  p.adjust.method = "BH",
  readable = T,
  verbose = T,
  n.cores = self$n.cores,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Onthology type, either GO (gene onthology) or DO (disease onthology). Please see DOSE package for more information.}

\item{\code{org}}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{\code{de.gene.ids}}{List containing DE gene IDs, and filtered DE genes (default: stored list, results from prepareOnthologyData)}

\item{\code{universe}}{All measured genes (default: stored vector)}

\item{\code{go.environment}}{Extracted GO environment. If set to NULL, the environment will be re-extracted (default: stored environment)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}

\item{\code{p.adjust.method}}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{\code{readable}}{Mapping gene ID to gene name (default=T)}

\item{\code{verbose}}{Print progress (default=T)}

\item{\code{n.cores}}{Number of cores used (default: stored vector)}

\item{\code{...}}{Additional parameters for sccore:::plapply function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing a list of onthologies per type of onthology, and a data frame with merged results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOnthologyDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-plotOnthologyDistribution}{}}}
\subsection{Method \code{plotOnthologyDistribution()}}{
Bar plot of onthologies per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOnthologyDistribution(type = NULL, cell.groups = self$cell.groups)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOnthologyTerms"></a>}}
\if{latex}{\out{\hypertarget{method-plotOnthologyTerms}{}}}
\subsection{Method \code{plotOnthologyTerms()}}{
Plot onthology terms as a function of both number of DE genes, and number of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOnthologyTerms(
  type = NULL,
  de.filter = self$test.results$onthology$de.filter,
  cell.groups = self$cell.groups,
  show.legend = T,
  legend.position = "bottom",
  label.x.pos = 0.01,
  label.y.pos = 1,
  rel_heights = c(2.5, 0.5),
  scale = 0.93
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{\code{de.filter}}{Filtered DE genes, results from prepareOnthologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{show.legend}}{Include legend in plot (default=T)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="bottom")}

\item{\code{label.x.pos}}{Plot label position on x axis (default=0.01)}

\item{\code{label.y.pos}}{Plot label position on y axis (default=1)}

\item{\code{rel_heights}}{Relative heights for plots. Only relevant if show.legend=T. See cowplot::plot_grid for more info (default=c(2.5, 0.5))}

\item{\code{scale}}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (default=0.93)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOnthologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotOnthologyHeatmap}{}}}
\subsection{Method \code{plotOnthologyHeatmap()}}{
Plot a heatmap of onthology P values per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOnthologyHeatmap(
  type = NULL,
  legend.position = "left",
  order = "all-max-row",
  n = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Onthology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default=NULL)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{order}}{Order of rows in heatmap. Can be 'unique' (only show onthologies that are unique for any cell type); 'unique-max-row' (same as 'unique' but ordered by P value); 'all-max-rowsum' (all onthologies ordered by cumulative P value for all cell types); 'all-max-row' (all onthologies ordered by max P value) (default="all-max-row")}

\item{\code{n}}{Number of onthologies to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{ont.data}}{List containing a list of results from estimateOnthology}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOnthologyCorrelations"></a>}}
\if{latex}{\out{\hypertarget{method-plotOnthologyCorrelations}{}}}
\subsection{Method \code{plotOnthologyCorrelations()}}{
Plot correlation matrix for onthologies between cell types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOnthologyCorrelations(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{\code{ont.data}}{List containing a list of results from estimateOnthology}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
