% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacoa.R
\name{Cacoa}
\alias{Cacoa}
\title{Cacoa R6 class}
\description{
Cacoa R6 class

Cacoa R6 class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n.cores}}{number of cores}

\item{\code{verbose}}{print diagnostic messages}

\item{\code{test.results}}{results of the estimations, ready to use}

\item{\code{cache}}{intermediate results of the estimations, which can be used during some other computations}

\item{\code{data.object}}{the main object storing data (Conos or Seurat)}

\item{\code{sample.groups}}{2-factor vector with annotation of groups/condition per sample}

\item{\code{cell.groups}}{named factor with cell names with cluster per cell}

\item{\code{sample.per.cell}}{named factor with cell names}

\item{\code{ref.level}}{reference level for sample.group vector}

\item{\code{target.level}}{target/disease level for sample.group vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cacoa$new()}}
\item \href{#method-estimateExpressionShiftMagnitudes}{\code{Cacoa$estimateExpressionShiftMagnitudes()}}
\item \href{#method-plotExpressionShiftMagnitudes}{\code{Cacoa$plotExpressionShiftMagnitudes()}}
\item \href{#method-estimateExpressionShiftZScores}{\code{Cacoa$estimateExpressionShiftZScores()}}
\item \href{#method-plotExpressionShiftZScores}{\code{Cacoa$plotExpressionShiftZScores()}}
\item \href{#method-estimatePerCellTypeDE}{\code{Cacoa$estimatePerCellTypeDE()}}
\item \href{#method-plotDEGenes}{\code{Cacoa$plotDEGenes()}}
\item \href{#method-saveDEasJSON}{\code{Cacoa$saveDEasJSON()}}
\item \href{#method-prepareOntologyData}{\code{Cacoa$prepareOntologyData()}}
\item \href{#method-plotFilteredDEGenes}{\code{Cacoa$plotFilteredDEGenes()}}
\item \href{#method-estimateOntology}{\code{Cacoa$estimateOntology()}}
\item \href{#method-plotOntologyDistribution}{\code{Cacoa$plotOntologyDistribution()}}
\item \href{#method-plotOntologyTerms}{\code{Cacoa$plotOntologyTerms()}}
\item \href{#method-plotOntologyBarplot}{\code{Cacoa$plotOntologyBarplot()}}
\item \href{#method-plotOntologyDotplot}{\code{Cacoa$plotOntologyDotplot()}}
\item \href{#method-plotOntologyHeatmap}{\code{Cacoa$plotOntologyHeatmap()}}
\item \href{#method-plotOntologySimilarities}{\code{Cacoa$plotOntologySimilarities()}}
\item \href{#method-plotProportions}{\code{Cacoa$plotProportions()}}
\item \href{#method-plotCellNumbers}{\code{Cacoa$plotCellNumbers()}}
\item \href{#method-clone}{\code{Cacoa$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$new(
  data.object,
  sample.groups = NULL,
  cell.groups = NULL,
  sample.per.cell = NULL,
  ref.level = NULL,
  target.level = NULL,
  n.cores = 1,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftMagnitudes}{}}}
\subsection{Method \code{estimateExpressionShiftMagnitudes()}}{
Calculate expression shift magnitudes of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftMagnitudes(
  cell.groups = self$cell.groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  sample.groups = self$sample.groups,
  n.cores = self$n.cores,
  verbose = self$verbose,
  name = "expression.shifts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{dist}}{'JS' - Jensen Shannon divergence, or 'cor' - correlation distance (default="JS")}

\item{\code{within.group.normalization}}{Normalize the shift magnitude by the mean magnitude of within-group variation (default=T)}

\item{\code{valid.comparisons}}{A logical matrix (rows and columns are samples) specifying valid between-sample comparisons. Note that if within.group.normalization=T, the method will automatically include all within-group comparisons of the samples for which at least one valid pair is included in the valid.comparisons (default=NULL)}

\item{\code{n.cells}}{Number of cells to subsmaple across all samples (if not specified, defaults to the total size of the smallest cell cluster)}

\item{\code{n.top.genes}}{Number of top highest-expressed genes to consider (default: all genes)}

\item{\code{n.subsamples}}{Number of samples to draw (default=100)}

\item{\code{min.cells}}{Minimum number of cells per cluster/per sample to be included in the analysis (default=10)}

\item{\code{sample.groups}}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{n.cores}}{Number of cores (default: stored integer)}

\item{\code{verbose}}{Print progress}

\item{\code{name}}{Test name (default="expression.shifts")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list include \itemize{
  \item{df: a table with cluster distances (normalized if within.gorup.normalization=T), cell type, number of cells}
  \item{ctdml: raw list of `n.subsamples` sampled distance matrices (cells were subsampled)}
  \item{sample.groups: same as the provided variable}
  \item{valid.comparisons: a matrix of valid comparisons (in this case all should be valid, since we're not restricting samples that should be compared)}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftMagnitudes"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftMagnitudes}{}}}
\subsection{Method \code{plotExpressionShiftMagnitudes()}}{
Plot results from cao$estimateExpressionShiftMagnitudes()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftMagnitudes(
  name = "expression.shifts",
  size.norm = F,
  normalized.distance = F,
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Test results to plot (default=expression.shifts)}

\item{\code{size.norm}}{Plot size normalized results. Requires cell.groups, and sample.per.cell (default=F)}

\item{\code{normalized.distance}}{Plot the absolute median distance (default=F)}

\item{\code{cell.groups}}{Named factor with cell names defining groups/clusters (default: stored vector)}

\item{\code{sample.per.cell}}{Named sample factor with cell names (default: stored vector)}

\item{\code{label}}{Plot labels on size normalized plots (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-estimateExpressionShiftZScores}{}}}
\subsection{Method \code{estimateExpressionShiftZScores()}}{
Calculate expression shift Z scores of different clusters between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateExpressionShiftZScores(
  cell.groups = self$cell.groups,
  ref.level = self$ref.level,
  sample.groups = self$sample.groups,
  sample.per.cell = self$sample.per.cell,
  n.od.genes = 1000,
  n.pcs = 100,
  pca.maxit = 1000,
  ignore.cache = F,
  name = "expression.z.scores",
  verbose = self$verbose
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{Named cell group factor with cell names (default: stored vector)}

\item{\code{ref.level}}{Reference sample group, e.g., ctrl, healthy, or untreated. (default: stored value)}

\item{\code{sample.groups}}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{\code{sample.per.cell}}{(default: stored vector)}

\item{\code{n.od.genes}}{(default=1000)}

\item{\code{n.pcs}}{(default=100)}

\item{\code{pca.maxit}}{(default=1000)}

\item{\code{ignore.cache}}{(default=F)}

\item{\code{name}}{(default="expression.z.scores")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotExpressionShiftZScores"></a>}}
\if{latex}{\out{\hypertarget{method-plotExpressionShiftZScores}{}}}
\subsection{Method \code{plotExpressionShiftZScores()}}{
Plot results from cao$estimateExpressionShiftZScores()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotExpressionShiftZScores(
  type.order = NULL,
  name = "expression.z.scores"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type.order}}{(default=NULL)}

\item{\code{name}}{Test results to plot (default=expression.z.shifts)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimatePerCellTypeDE"></a>}}
\if{latex}{\out{\hypertarget{method-estimatePerCellTypeDE}{}}}
\subsection{Method \code{estimatePerCellTypeDE()}}{
Estimate differential gene expression per cell type between conditions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimatePerCellTypeDE(
  cell.groups = self$cell.groups,
  sample.groups = self$sample.groups,
  ref.level = self$ref.level,
  common.genes = F,
  n.cores = self$n.cores,
  cooks.cutoff = FALSE,
  min.cell.count = 10,
  independent.filtering = FALSE,
  cluster.sep.chr = "<!!>",
  return.matrix = F,
  verbose = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cell.groups}}{factor specifying cell types (default=NULL)}

\item{\code{sample.groups}}{a list of two character vector specifying the app groups to compare (default=NULL)}

\item{\code{ref.level}}{Reference level in 'sample.groups', e.g., ctrl, healthy, wt (default=NULL)}

\item{\code{common.genes}}{Only investigate common genes across cell groups (default=F)}

\item{\code{n.cores}}{Number of cores (default=1)}

\item{\code{cooks.cutoff}}{cooksCutoff for DESeq2 (default=F)}

\item{\code{min.cell.count}}{(default=10)}

\item{\code{independent.filtering}}{independentFiltering for DESeq2 (default=F)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}

\item{\code{return.matrix}}{Return merged matrix of results (default=F)}

\item{\code{verbose}}{Show progress (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotDEGenes}{}}}
\subsection{Method \code{plotDEGenes()}}{
Plot number of significant DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotDEGenes(
  de.raw = self$test.results$de,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T,
  p.adjust.cutoff = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.raw}}{List with differentially expressed genes per cell group (default: stored list, results from estimatePerCellTypeDE)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}

\item{\code{p.adjust.cutoff}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDEasJSON"></a>}}
\if{latex}{\out{\hypertarget{method-saveDEasJSON}{}}}
\subsection{Method \code{saveDEasJSON()}}{
Save DE results as JSON files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$saveDEasJSON(
  saveprefix = NULL,
  dir.name = "JSON",
  de.raw = self$test.results$de,
  gene.metadata = NULL,
  cluster.sep.chr = "<!!>"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{saveprefix}}{Prefix for created files (default=NULL)}

\item{\code{dir.name}}{Name for directory with results (default="JSON")}

\item{\code{de.raw}}{List of DE results}

\item{\code{gene.metadata}}{(default=NULL)}

\item{\code{cluster.sep.chr}}{character string of length 1 specifying a delimiter to separate cluster and app names (default="<!!>")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepareOntologyData"></a>}}
\if{latex}{\out{\hypertarget{method-prepareOntologyData}{}}}
\subsection{Method \code{prepareOntologyData()}}{
Filter and prepare DE genes for ontology calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$prepareOntologyData(
  org.db,
  n.top.genes = 100,
  p.adj.cutoff = 0.05,
  expr.cutoff = 0.05,
  de.raw = self$test.results$de,
  cell.groups = self$cell.groups,
  universe = NULL,
  transposed = T,
  verbose = T,
  n.cores = self$n.cores
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{n.top.genes}}{Number of most different genes to take as input. If less are left after filtering for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}

\item{\code{p.adj.cutoff}}{Cutoff for filtering highly-expressed DE genes (default=0.05)}

\item{\code{expr.cutoff}}{Cutoff for cells per group expressing a DE gene, i.e., cutoff for highly-expressed genes (default=0.05)}

\item{\code{de.raw}}{Differentially expressed genes per cell group, results from estimatePerCellTypeDE (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{universe}}{Only set this if a common background gene set is desired for all cell groups (default: NULL)}

\item{\code{transposed}}{Whether count matrices should be transposed (default=T)}

\item{\code{verbose}}{Print progress (default=T)}

\item{\code{n.cores}}{Number of cores to use (default: stored integer)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFilteredDEGenes"></a>}}
\if{latex}{\out{\hypertarget{method-plotFilteredDEGenes}{}}}
\subsection{Method \code{plotFilteredDEGenes()}}{
Plot number of highly-expressed DE genes as a function of number of cells
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotFilteredDEGenes(
  de.filter = self$test.results$ontology$de.filter,
  cell.groups = self$cell.groups,
  legend.position = "none",
  label = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{\code{label}}{Show labels on plot (default=T)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimateOntology"></a>}}
\if{latex}{\out{\hypertarget{method-estimateOntology}{}}}
\subsection{Method \code{estimateOntology()}}{
Estimate ontology terms based on DEs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$estimateOntology(
  type = "GO",
  org.db,
  de.gene.ids = self$test.results$ontology$de.gene.ids,
  go.environment = self$test.results$GO$go.environment,
  p.adj = 0.05,
  p.adjust.method = "BH",
  readable = T,
  verbose = T,
  qvalueCutoff = 0.2,
  minGSSize = 5,
  maxGSSize = 500,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information (default="GO")}

\item{\code{org.db}}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{\code{de.gene.ids}}{List containing DE gene IDs, and filtered DE genes (default: stored list, results from prepareOntologyData)}

\item{\code{go.environment}}{Extracted GO environment. If set to NULL, the environment will be re-extracted (default: stored environment)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}

\item{\code{p.adjust.method}}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{\code{readable}}{Mapping gene ID to gene name (default=T)}

\item{\code{verbose}}{Print progress (default=T)}

\item{\code{qvalueCutoff}}{Q value cutoff, please see clusterProfiler package for more information (default=0.2)}

\item{\code{minGSSize}}{Minimal geneset size, please see clusterProfiler package for more information (default=5)}

\item{\code{maxGSSize}}{Minimal geneset size, please see clusterProfiler package for more information (default=5e2)}

\item{\code{...}}{Additional parameters for sccore:::plapply function}

\item{\code{universe}}{All measured genes (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing a list of terms per ontology, and a data frame with merged results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyDistribution}{}}}
\subsection{Method \code{plotOntologyDistribution()}}{
Bar plot of ontology terms per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyDistribution(
  genes = c("up", "down"),
  type = "GO",
  cell.groups = self$cell.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyTerms"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyTerms}{}}}
\subsection{Method \code{plotOntologyTerms()}}{
Plot ontology terms as a function of both number of DE genes, and number of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyTerms(
  genes = c("up", "down"),
  type = "GO",
  de.filter = self$test.results$ontology$de.filter,
  cell.groups = self$cell.groups,
  label.x.pos = 0.01,
  label.y.pos = 1,
  scale = 0.93
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up', 'all' or a combination of these (default=c("up","down"))}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}

\item{\code{de.filter}}{Filtered DE genes, results from prepareOntologyData (default: stored list)}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{label.x.pos}}{Plot label position on x axis (default=0.01)}

\item{\code{label.y.pos}}{Plot label position on y axis (default=1)}

\item{\code{scale}}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (default=0.93)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyBarplot"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyBarplot}{}}}
\subsection{Method \code{plotOntologyBarplot()}}{
Plot a barplot of ontology terms with adj. P values for a specific cell subgroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyBarplot(
  genes = "up",
  type = "GO",
  cell.subgroups = NULL,
  n = 20,
  p.adj = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{cell.subgroups}}{Cell group to plot (default=NULL)}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyDotplot"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyDotplot}{}}}
\subsection{Method \code{plotOntologyDotplot()}}{
Plot a dotplot of ontology terms with adj. P values for a specific cell subgroup
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyDotplot(
  genes = "up",
  type = "GO",
  cell.subgroups = NULL,
  n = 20,
  p.adj = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{cell.subgroups}}{Cell group to plot (default=NULL)}

\item{\code{n}}{Number of ontology terms to show. Not applicable when order is 'unique' or 'unique-max-row' (default=10)}

\item{\code{p.adj}}{Adjusted P cutoff (default=0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologyHeatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologyHeatmap}{}}}
\subsection{Method \code{plotOntologyHeatmap()}}{
Plot a heatmap of ontology P values per cell type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologyHeatmap(
  genes = "up",
  type = "GO",
  legend.position = "left",
  selection = "all",
  n = 10,
  cell.subgroups = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down' for downregulated genes, 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "BP", "CC", or "MF" (GO types) or "DO" (default="GO")}

\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="left")}

\item{\code{selection}}{Order of rows in heatmap. Can be 'unique' (only show terms that are unique for any cell type); 'common' (only show terms that are present in at least two cell types); 'all' (all ontology terms) (default="all")}

\item{\code{n}}{Number of terms to show (default=10)}

\item{\code{cell.subgroups}}{Cell groups to plot (default=NULL)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotOntologySimilarities"></a>}}
\if{latex}{\out{\hypertarget{method-plotOntologySimilarities}{}}}
\subsection{Method \code{plotOntologySimilarities()}}{
Plot correlation matrix for ontology terms between cell types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotOntologySimilarities(genes = "up", type = "GO")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Specify which genes to plot, can either be 'down', 'up' or 'all' (default="up")}

\item{\code{type}}{Ontology, must be either "GO" or "DO" (default="GO")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotProportions"></a>}}
\if{latex}{\out{\hypertarget{method-plotProportions}{}}}
\subsection{Method \code{plotProportions()}}{
Plot the cell group proportions per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotProportions(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{Vector indicating sample groups with sample names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCellNumbers"></a>}}
\if{latex}{\out{\hypertarget{method-plotCellNumbers}{}}}
\subsection{Method \code{plotCellNumbers()}}{
Plot the cell numbers per sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$plotCellNumbers(
  legend.position = "right",
  cell.groups = self$cell.groups,
  sample.per.cell = self$sample.per.cell,
  sample.groups = self$sample.groups
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{legend.position}}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{\code{cell.groups}}{Vector indicating cell groups with cell names (default: stored vector)}

\item{\code{sample.per.cell}}{Vector indicating sample name with cell names (default: stored vector)}

\item{\code{sample.groups}}{Vector indicating sample groups with sample names (default: stored vector)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cacoa$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
