% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clusteredOntologyDotplot}
\alias{clusteredOntologyDotplot}
\title{Clustered Ontology Dotplot}
\usage{
clusteredOntologyDotplot(
  ont.res,
  p.adj = 0.05,
  min.genes = 1,
  cut.h = 0.66,
  top.n = Inf,
  ...
)
}
\description{
Performs ontology clustering by genes and then shows medoids of the clusters with
enrichplot::dotplot
}
