% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_tree_new.R
\name{constructBestPartitionTree}
\alias{constructBestPartitionTree}
\title{Construct the canonical tree}
\usage{
constructBestPartitionTree(cnts, groups, partition.thresh = 0)
}
\arguments{
\item{cnts}{Table with cell type counts}

\item{groups}{Groups variable for samples}
}
\value{
phylo tree
}
\description{
Construct the canonical tree
}
\keyword{internal}
