% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_trees.R
\name{constructCanonicalTree}
\alias{constructCanonicalTree}
\title{#' @description Get balances for all inner nodes of the tree
#'
#' @param t A phylo object
#' @param log.freq Logarithms of frequnencies
#' @return Balances, contrast matrix and names on cell types
#' @details We use ellipsis because a previous version of this function
#'          contains additional and insignificant parameters.
getNodeBalances <- function(t, log.freq, ...){
checkDataTree(log.freq, t)---------}
\usage{
constructCanonicalTree(d.counts, d.groups)
}
\arguments{
\item{d.counts}{Table with cell type counts}

\item{d.groups}{Groups variable for samples}
}
\value{
phylo tree
}
\description{
Construct the canonical tree
}
\details{
sample.names <- rownames(log.freq)
}
\section{Initialisation of balances}{
n.cells <- ncol(log.freq)
n.samples <- nrow(log.freq)
log.freq.ext <- cbind(log.freq\link{,t$tip.label}, matrix(0, nrow=n.samples, ncol=n.cells-1))  # log-frequencoes extended for inner nodes
n.in.clade <- c(rep(1,n.cells), rep(0,n.cells-1)) # number of cell-types in a clade

balances <- matrix(0, nrow=n.samples, ncol=n.cells-1)
rownames(balances) <- sample.names
sbp.my <- rbind(diag(x = 1, n.cells, n.cells), matrix(0, n.cells - 1,n.cells))

edges <- t$edge
id.edges <- getNodeOrder(t)
}

\section{Calculate balances in traversing the tree}{

}

