% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{constructTree}
\alias{constructTree}
\title{Construct the canonical tree}
\usage{
constructTree(cnts, groups, partition.thresh = 0)
}
\arguments{
\item{cnts}{Table with cell type counts}

\item{groups}{Groups variable for samples}
}
\value{
phylo tree
}
\description{
Construct the canonical tree
}
\keyword{internal}
