% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{distanceBetweenTerms}
\alias{distanceBetweenTerms}
\title{Distance between terms}
\usage{
distanceBetweenTerms(genes.per.go)
}
\arguments{
\item{genes.per.go}{named list of genes per GO}
}
\value{
Distance matrix
}
\description{
Calculate distance matrix between ontology terms
}
