% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{estimateCellDensityGraph}
\alias{estimateCellDensityGraph}
\title{Estimate graph smooth based cell density}
\usage{
estimateCellDensityGraph(
  graph,
  sample.per.cell,
  sample.groups,
  n.cores = 1,
  beta = 30,
  m = 50,
  verbose = TRUE
)
}
\arguments{
\item{sample.per.cell}{Named sample factor with cell names (default: stored vector)}

\item{sample.groups}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{n.cores}{number of cores}

\item{m}{numeric Maximum order of Chebyshev coeff to compute (default=50)}
}
\description{
Estimate graph smooth based cell density
}
\keyword{internal}
