% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{estimateCellDensityKde}
\alias{estimateCellDensityKde}
\title{Estimate cell density in giving embedding, Density will estimated for indivisual sample}
\usage{
estimateCellDensityKde(
  emb,
  sample.per.cell,
  sample.groups,
  bins,
  bandwidth = 0.05,
  expansion.mult = 0.05
)
}
\arguments{
\item{emb}{cell embedding matrix}

\item{sample.per.cell}{Named sample factor with cell names (default: stored vector)}

\item{sample.groups}{A two-level factor on the sample names describing the conditions being compared (default: stored vector)}

\item{bins}{number of bins for density estimation, default 400}
}
\description{
Estimate cell density in giving embedding, Density will estimated for indivisual sample
}
\keyword{internal}
