% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeDESeq}
\alias{estimateDEForTypeDESeq}
\title{estimate DE using DESeq2}
\usage{
estimateDEForTypeDESeq(
  cm,
  meta,
  design.formula,
  ref.level,
  target.level,
  test.type,
  ...
)
}
\arguments{
\item{cm}{count matrix}

\item{meta}{data frame containing meta data}

\item{design.formula}{design formula according to}

\item{ref.level}{reference level, e.g. controls}

\item{target.level}{target level, e.g. disease}

\item{test.type}{test type incorporated in DESeq2, either "Wald" or "LRT"}

\item{...}{additional parameters forwarded to DESeq2}
}
\description{
estimate DE using DESeq2
}
\keyword{internal}
