% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeEdgeR}
\alias{estimateDEForTypeEdgeR}
\title{Estimate DE using edgeR}
\usage{
estimateDEForTypeEdgeR(cm, meta, design.formula)
}
\arguments{
\item{cm}{count matrix}

\item{meta}{data frame containing metadata}

\item{design.formula}{design formula according to}
}
\description{
Estimate DE using edgeR
}
\keyword{internal}
