% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{estimateDEForTypeLimma}
\alias{estimateDEForTypeLimma}
\title{Estimate DE using limma}
\usage{
estimateDEForTypeLimma(cm, meta, design.formula, target.level)
}
\arguments{
\item{cm}{count matrix}

\item{meta}{data frame containing metadata}

\item{design.formula}{design formula according to}

\item{target.level}{target level, e.g. disease}
}
\description{
Estimate DE using limma
}
\keyword{internal}
