% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_distance.R
\name{estimateExpressionShiftMagnitudes}
\alias{estimateExpressionShiftMagnitudes}
\title{Expression shift magnitudes per cluster between conditions}
\usage{
estimateExpressionShiftMagnitudes(
  count.matrices,
  sample.groups,
  cell.groups,
  dist = "JS",
  within.group.normalization = TRUE,
  valid.comparisons = NULL,
  n.cells = NULL,
  n.top.genes = Inf,
  n.subsamples = 100,
  min.cells = 10,
  n.cores = 1,
  verbose = FALSE,
  transposed.matrices = FALSE
)
}
\arguments{
\item{count.matrices}{List of count matrices}

\item{sample.groups}{Named factor with cell names indicating condition/sample, e.g., ctrl/disease}

\item{cell.groups}{Named clustering/annotation factor with cell names}

\item{dist}{what distance measure to use: 'JS' - Jensen-Shannon divergence (default), 'cor' - Pearson's linear correlation on log transformed values}

\item{within.group.normalization}{(default=T)}

\item{valid.comparisons}{(default=NULL)}

\item{n.cells}{Number of cells per group (default=NULL)}

\item{n.top.genes}{(default=Inf)}

\item{n.subsamples}{(default=100)}

\item{min.cells}{(default=10)}

\item{n.cores}{number of cores (default=1)}

\item{verbose}{(default=F)}

\item{transposed.matrices}{(default=F)}
}
\description{
Expression shift magnitudes per cluster between conditions
}
