% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_distance.R
\name{estimateExpressionShiftMagnitudes}
\alias{estimateExpressionShiftMagnitudes}
\title{Expression shift magnitudes per cluster between conditions}
\usage{
estimateExpressionShiftMagnitudes(
  cm.per.type,
  sample.groups,
  cell.groups,
  sample.per.cell,
  dist = NULL,
  normalize.both = TRUE,
  verbose = FALSE,
  ref.level = NULL,
  n.permutations = 1000,
  p.adjust.method = "BH",
  top.n.genes = NULL,
  gene.selection = "wilcox",
  n.pcs = NULL,
  trim = 0.2,
  n.cores = 1,
  norm.type = NULL,
  ...
)
}
\arguments{
\item{cm.per.type}{List of normalized count matrices per cell type}

\item{sample.groups}{Named factor with cell names indicating condition/sample, e.g., ctrl/disease}

\item{cell.groups}{Named clustering/annotation factor with cell names}

\item{dist}{what distance measure to use: 'JS' - Jensen-Shannon divergence (default), 'cor' - Pearson's linear correlation on log transformed values}

\item{normalize.both}{(default=T)}

\item{verbose}{(default=F)}

\item{n.cores}{number of cores (default=1)}

\item{transposed.matrices}{(default=F)}
}
\description{
Expression shift magnitudes per cluster between conditions
}
