% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free.R
\name{estimateGeneProgramsFabia}
\alias{estimateGeneProgramsFabia}
\title{Estimate Gene Programs with FABIA}
\usage{
estimateGeneProgramsFabia(
  z.scores,
  n.programs,
  n.sampled.cells = 15000,
  cyc = 1500,
  alpha = 0.2,
  random = -1,
  ...
)
}
\arguments{
\item{n.programs}{maximal number of gene programs to find (parameter \code{p} for fabia).}

\item{n.sampled.cells}{number of sub-sampled cells for estimating the gene programs. If 0, all cells are used.
it is interpreted as a vector of cell}

\item{...}{
  Arguments passed on to \code{\link[fabia:fabia]{fabia::fabia}}
  \describe{
    \item{\code{spl}}{sparseness prior loadings (0 - 2.0); default = 0 (Laplace).}
    \item{\code{spz}}{sparseness factors (0.5 - 2.0); default = 0.5 (Laplace).}
    \item{\code{non_negative}}{Non-negative factors and loadings if non_negative
    > 0; default = 0.}
    \item{\code{center}}{data centering: 1 (mean), 2
    (median), > 2 (mode), 0 (no); default = 2.}
    \item{\code{norm}}{data normalization:   1 (0.75-0.25 quantile), >1 (var=1),
    0 (no); default = 1.}
    \item{\code{scale}}{loading vectors are scaled in each iteration to the given
    variance. 0.0 indicates non scaling; default = 0.0.}
    \item{\code{lap}}{minimal value of the variational parameter; default = 1.0}
    \item{\code{nL}}{maximal number of biclusters at which a row element can participate; default
    = 0 (no limit)}
    \item{\code{lL}}{maximal number of row elements per bicluster; default
    = 0 (no limit)}
    \item{\code{bL}}{cycle at which the nL or lL maximum starts; default
    = 0 (start at the beginning)}
  }}
}
\description{
Estimate Gene Programs with FABIA
}
