% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntology}
\alias{estimateOntology}
\title{Estimate ontology}
\usage{
estimateOntology(
  type = "GO",
  org.db,
  de.gene.ids,
  go.environment = NULL,
  p.adj = 0.05,
  p.adjust.method = "BH",
  readable = T,
  verbose = T,
  ...
)
}
\arguments{
\item{type}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information.}

\item{p.adj}{Adjusted P cutoff (default=0.05)}

\item{p.adjust.method}{Method for calculating adj. P. Please see DOSE package for more information (default="BH")}

\item{readable}{Mapping gene ID to gene name (default=T)}

\item{verbose}{Print progress (default=T)}

\item{...}{Additional parameters for sccore:::plapply function}

\item{ont.data}{List containing DE gene IDs, and filtered DE genes}

\item{org}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{n.cores}{Number of cores used (default: stored vector)}
}
\value{
A list containing a list of ontologies per type of ontology, and a data frame with merged results
}
\description{
Calculate ontologies based on DEs
}
