% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{estimateOntologyFromIds}
\alias{estimateOntologyFromIds}
\title{Estimate ontology}
\usage{
estimateOntologyFromIds(
  de.gene.ids,
  go.environment,
  type = "GO",
  org.db = NULL,
  n.top.genes = 500,
  keep.gene.sets = FALSE,
  verbose = TRUE,
  qvalue.cutoff = 0.2,
  ...
)
}
\arguments{
\item{type}{Ontology type, either GO (gene ontology) or DO (disease ontology). Please see DOSE package for more information.}

\item{org.db}{Organism database, e.g., org.Hs.eg.db for human or org.Ms.eg.db for mouse. Input must be of class 'OrgDb'}

\item{n.top.genes}{Number of most different genes to take as input. If less are left after filtering for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}

\item{verbose}{Print progress (default=T)}

\item{qvalue.cutoff}{Q value cutoff, please see clusterProfiler package for more information (default=0.2)}

\item{...}{Additional parameters for DO/GO/GSEA functions}
}
\value{
A list containing a list of ontologies per type of ontology, and a data frame with merged results
}
\description{
Calculate ontologies based on DEs
}
