% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getCoreSamples}
\alias{getCoreSamples}
\title{Get core samples from model}
\usage{
getCoreSamples(model, samples)
}
\arguments{
\item{model}{list returned by buildDesignMatrices (must contain F and core.rows)}

\item{samples}{character vector of sample IDs}
}
\value{
character vector of core sample IDs
}
\description{
Get core samples from model
}
\keyword{internal}
