% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getGenePalette}
\alias{getGenePalette}
\title{Get Gene Scale}
\usage{
getGenePalette(
  genes = c("up", "down", "all"),
  neutral.col = "white",
  bidirectional = FALSE
)
}
\arguments{
\item{genes}{character vector (default=c("up", "down" or "all")) Type of genes}

\item{high}{color for the highest value (default: "gray80")}
}
\value{
palette function
}
\description{
Get Gene Scale
}
