% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{getGenePalette}
\alias{getGenePalette}
\title{Get Gene Scale}
\usage{
getGenePalette(
  genes = c("up", "down", "all"),
  neutral.col = "white",
  bidirectional = FALSE
)
}
\arguments{
\item{genes}{character vector Type of genes (default=c("up", "down" or "all"))}

\item{neutral.col}{character string (default="white")}

\item{bidirectional}{boolean (default=FALSE)}
}
\value{
palette function
}
\description{
Get Gene Scale
}
