% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_new.R
\name{getLoadings}
\alias{getLoadings}
\title{Get loadings by different methods}
\usage{
getLoadings(cnts, groups, criteria = "lda", ref.cell.type = NULL)
}
\arguments{
\item{cnts}{Counts of cell typer in samples. Rows - samples, columns - cell types}

\item{groups}{Vector with boolean values. TRUE - sample in the case group, FALSE - sample in the control group}

\item{criteria}{Method to get loadings}

\item{ref.cell.type}{Reference cell type}
}
\value{
Updated data frame with Z scores
}
\description{
Get loadings by different methods
}
\keyword{internal}
