% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{getNodeBalances}
\alias{getNodeBalances}
\title{Get balances for all inner nodes of the tree}
\usage{
getNodeBalances(log.f, sbp.my)
}
\arguments{
\item{log.f}{Logarithms of frequencies}

\item{t}{A phylo object}
}
\value{
Balances, contrast matrix and names on cell types
}
\description{
Get balances for all inner nodes of the tree
}
\keyword{internal}
