% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getPerCellTypeDEmat}
\alias{getPerCellTypeDEmat}
\title{Do differential expression for each cell type in a conos object between the specified subsets of apps}
\usage{
getPerCellTypeDEmat(
  raw.mats,
  groups = NULL,
  sample.groups = NULL,
  cooks.cutoff = FALSE,
  ref.level = NULL,
  min.cell.count = 10,
  independent.filtering = FALSE,
  n.cores = 1,
  cluster.sep.chr = "<!!>",
  return.details = TRUE
)
}
\arguments{
\item{raw.mats}{list of counts matrix; row for gene and column for cell}

\item{groups}{factor specifying cell types}

\item{sample.groups}{a list of two character vector specifying the app groups to compare}

\item{cooks.cutoff}{cooksCutoff for DESeq2}

\item{independent.filtering}{independentFiltering for DESeq2}

\item{n.cores}{number of cores}

\item{cluster.sep.chr}{character string of length 1 specifying a delimiter to separate cluster and app names}

\item{return.details}{return detals}
}
\description{
Do differential expression for each cell type in a conos object between the specified subsets of apps
}
