% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_balances.R
\name{getRndPartition}
\alias{getRndPartition}
\title{Generate one random partition of given length}
\usage{
getRndPartition(bal.len)
}
\arguments{
\item{bal.len}{length of partiotion}
}
\value{
random partition - verctor with values from {1, -1}
}
\description{
Generate one random partition of given length
}
