% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free_de.R
\name{localZScores}
\alias{localZScores}
\title{Local Z Scores}
\usage{
localZScores(
  graph,
  condition.per.cell,
  genes,
  count.matrix.transposed,
  reference.level,
  min.expr = 1e-10,
  min.std = 1e-10,
  n.cores = 2,
  verbose = 1
)
}
\arguments{
\item{graph}{alignment graph}

\item{condition.per.cell}{conditions per cell. Must have two levels.}

\item{genes}{genes to be tested. Procedure is slow, so it's reasonable to restrict amount of genes.}

\item{count.matrix.transposed}{joint count matrix with cells by rows and genes by columns}

\item{reference.level}{reference condition level}
}
\description{
Local Z Scores
}
