% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_free_de.R
\name{localZScores}
\alias{localZScores}
\title{Local Z Scores}
\usage{
localZScores(
  graph,
  condition.per.cell,
  genes,
  count.matrix.transposed,
  ref.level,
  min.expr = 1e-10,
  min.std = 1e-10,
  n.cores = 1,
  verbose = T
)
}
\arguments{
\item{graph}{Alignment graph (embedding)}

\item{condition.per.cell}{Named group factor with cell names. Must have exactly two levels.}

\item{genes}{Genes to be tested. Procedure is slow, so it's reasonable to restrict amount of genes.}

\item{count.matrix.transposed}{Joint count matrix with cells by rows and genes by columns}

\item{ref.level}{Reference condition level, e.g., wt, ctrl, or healthy}
}
\description{
Local Z Scores
}
