% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellLoadings}
\alias{plotCellLoadings}
\title{Internal plotting function for cell loadings}
\usage{
plotCellLoadings(
  loadings,
  pval,
  ref.level,
  target.level,
  signif.threshold = 0.05,
  jitter.alpha = 0.1,
  palette = NULL,
  show.pvals = FALSE,
  plot.theme = theme_get(),
  jitter.size = 1,
  ordering = c("pvalue", "loading"),
  ref.load.level = 0,
  annotation.x = NULL,
  annotation.y = 1
)
}
\description{
Internal plotting function for cell loadings
}
\keyword{internal}
