% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCountBoxplotsPerType}
\alias{plotCountBoxplotsPerType}
\title{Plot Count Boxplots Per Type}
\usage{
plotCountBoxplotsPerType(
  count.df,
  y.lab = "count",
  x.lab = "",
  y.expand = 1.05,
  show.significance = FALSE,
  jitter.width = 0.15,
  notch = FALSE,
  legend.position = "right",
  alpha = 0.2,
  size = 0.5,
  palette = NULL,
  adjust.pvalues = TRUE,
  plot.theme = theme_get(),
  label.y.npc = 0.92
)
}
\arguments{
\item{count.df}{data.frame with columns \code{group}, \code{variable} and \code{value}}

\item{jitter.width}{width of the point jitter (default: 0.15)}

\item{notch}{Whether to show notch in the boxplots}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{alpha}{Transparency level on the data points (default: 0.2)}

\item{size}{marker size (default: 0.5)}
}
\description{
Plot Count Boxplots Per Type
}
