% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDEGenes}
\alias{plotDEGenes}
\title{Plot raw DE genes}
\usage{
plotDEGenes(
  de.raw,
  cell.groups,
  legend.position = "none",
  p.adjust.cutoff = 0.05,
  label = T
)
}
\arguments{
\item{de.raw}{List of differentially expressed genes per cell type, results from getPerCellTypeDE (default: stored list)}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="none")}

\item{p.adjust.cutoff}{Adjusted P cutoff (default=0.05)}
}
\value{
A ggplot2 object
}
\description{
Plot number of DE genes as a function of number of cells
}
