% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{plotDensity}
\alias{plotDensity}
\title{cell embedding density}
\usage{
plotDensity(emb, anoSample, sample.groups, bins = 200)
}
\arguments{
\item{emb}{cell embedding matrix}

\item{anoSample}{sample annotation  #Could do with better explanation}

\item{sample.groups}{sample groups factor}

\item{bins}{number of bins for density
@import dplyr
@import tibble
@import tidyr
@importFrom MASS kde2d}
}
\description{
Evaluate cell density between different sample groups.  Density is calculated per sample and normlized by quantiles normlization.
}
