% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap}
\usage{
plotHeatmap(
  df,
  color.per.group = NULL,
  row.order = TRUE,
  col.order = TRUE,
  legend.position = "right",
  size.df = NULL,
  size.range = c(1, 5),
  size.legend.title = "size",
  legend.key.width = unit(8, "pt"),
  legend.title = "-log10(p-value)",
  x.axis.position = "top",
  color.range = NULL,
  plot.theme = ggplot2::theme_get(),
  symmetric = FALSE,
  palette = NULL,
  font.size = 8,
  distance = "manhattan",
  clust.method = "complete",
  grid.color = "gray50"
)
}
\arguments{
\item{df}{Data frame with plot data}

\item{color.per.group}{Colors per cell group (default=NULL)}

\item{row.order}{Forced row order (default=NULL)}

\item{col.order}{Forced column order (default=NULL)}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="right")}

\item{size.df}{(default=NULL)}

\item{size.range}{(default=c(1, 5))}

\item{size.legend.title}{(default="size")}

\item{legend.key.width}{(default=unit(8, "pt))}

\item{legend.title}{Title on plot (default="-log10(p-value)")}

\item{x.axis.position}{Position of x axis (default="top")}

\item{color.range}{Range for filling colors}

\item{plot.theme}{(default=ggplot2::theme_get())}

\item{symmetric}{boolean (default=FALSE)}

\item{palette}{(default=NULL)}

\item{font.size}{integer (default=8)}

\item{distance}{character string (default="manhattan")}

\item{clust.method}{character string (default="complete")}

\item{grid.color}{Color of the grid. Set to "transparent" to disable the grid. (default: "gray50")}
}
\value{
A ggplot2 object
}
\description{
Plot heatmap
}
