% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMeanMedValuesPerCellType}
\alias{plotMeanMedValuesPerCellType}
\title{Plot bar, point or boxplots showing mean/median values per cell type}
\usage{
plotMeanMedValuesPerCellType(
  df,
  pvalues = NULL,
  type = c("box", "point", "bar"),
  show.jitter = TRUE,
  notch = TRUE,
  jitter.alpha = 0.05,
  palette = NULL,
  ylab = "expression distance",
  yline = 1,
  plot.theme = theme_get(),
  jitter.size = 1,
  line.size = 0.75,
  trim = 0,
  order.x = TRUE,
  pvalue.y = NULL,
  y.max = NULL
)
}
\arguments{
\item{df}{\itemize{
\item data frame containing the results, including $value and $Type slots which will be summarized
}}

\item{type}{\itemize{
\item type of a plot "bar" (default), "point" (mean + sd), or "box" for boxplot
}}

\item{show.jitter}{whether to show individual data points (default: FALSE)}

\item{notch}{\itemize{
\item whether to show notches in the boxplot version (default=TRUE)
}}

\item{jitter.alpha}{transparency value for the data points (default: 0.05)}

\item{palette}{\itemize{
\item cell type palette
}}
}
\value{
A ggplot2 object
}
\description{
Generic function for plotting mean or median values per cell type (used for expression shift distances and others)
}
