% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNCellRegression}
\alias{plotNCellRegression}
\title{Plot Number of Cells Regression}
\usage{
plotNCellRegression(
  n,
  n.total,
  x.lab = "Number of cells",
  y.lab = "N",
  legend.position = "right",
  label = TRUE,
  size = 4,
  palette = NULL,
  plot.line = TRUE,
  line.width = 0.5
)
}
\arguments{
\item{n}{the regressed variable}

\item{n.total}{number of cells}

\item{x.lab}{x axis label (default: "Number of cells").}

\item{y.lab}{y axis label (default: "N")}

\item{legend.position}{legend position (default: "right")}

\item{label}{show cell type labels (default: TRUE)}

\item{size}{label text size (default: 4)}

\item{palette}{color palette}

\item{plot.line}{plot the robust regression (default: TRUE)}

\item{line.width}{regression line width (default: 0.5)}
}
\description{
Plot Number of Cells Regression
}
