% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOnthologyTerms}
\alias{plotOnthologyTerms}
\title{Plot onthology terms}
\usage{
plotOnthologyTerms(
  type = NULL,
  ont.res,
  de.genes.filtered,
  cell.groups,
  legend.position = "bottom",
  label.x.pos = 0.01,
  label.y.pos = 1,
  rel_heights = c(2.5, 0.5),
  scale = 0.93,
  show.legend = T
)
}
\arguments{
\item{type}{Onthology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Onthology results in a data frame}

\item{cell.groups}{Vector indicating cell group sizes with cell group names}

\item{legend.position}{Position of legend in plot. See ggplot2::theme (default="bottom")}

\item{label.x.pos}{Plot label position on x axis (default=0.01)}

\item{label.y.pos}{Plot label position on y axis (default=1)}

\item{rel_heights}{Vector indicating relative heights for plots. Only relevant if show.legend=T. See cowplot::plot_grid for more info (default=c(2.5, 0.5))}

\item{scale}{Scaling of plots, adjust if e.g. label is misplaced. See cowplot::plot_grid for more info (default=0.93)}

\item{show.legend}{Include legend in plot (default=T)}
}
\value{
A ggplot2 object
}
\description{
Plot onthology terms as a function of both number of DE genes, and number of cells.
}
