% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologyFamily}
\alias{plotOntologyFamily}
\title{Plot related ontologies in one hierarchical network plot}
\usage{
plotOntologyFamily(
  fam,
  data,
  plot.type = "complete",
  show.ids = FALSE,
  string.length = 18,
  legend.label.size = 1,
  legend.position = "topright",
  verbose = TRUE,
  n.cores = 1,
  reduce.edges = FALSE,
  font.size = 24
)
}
\arguments{
\item{fam}{List of ontology IDs for the chosen family}

\item{data}{Data frane if raw ontology data for the chosen cell type}

\item{plot.type}{character (default="complete") How much of the family network should be plotted. Can be "complete" (entire network), "dense" (show 1 parent for each significant term), or "minimal" (only show significant terms)}

\item{show.ids}{boolean (default=FALSE) Whether to show ontology IDs instead of names}

\item{string.length}{integer (default=18) Length of strings for wrapping in order to fit text within boxes}

\item{legend.label.size}{integer (default=1) Size of legend labels}

\item{legend.position}{character vecotr (default="topright") Position of legend}

\item{verbose}{boolean (default=TRUE) Print messages}

\item{n.cores}{integer (default=1) Number of cores to use}

\item{reduce.edges}{boolean (default=FALSE) Remove redundant edges in network}

\item{font.size}{integer (default=24) Size of the font}
}
\value{
Rgraphviz object
}
\description{
Plot related ontologies in one hierarchical network plot
}
\keyword{internal}
