% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotOntologySimilarities}
\alias{plotOntologySimilarities}
\title{Plot ontology correlations}
\usage{
plotOntologySimilarities(type = NULL, ont.res, genes = NULL)
}
\arguments{
\item{type}{Ontology, must be either "GO" or "DO" (default=NULL)}

\item{ont.res}{Data frame with ontology resuls from estimateOntology}

\item{genes}{Specify which genes are plotted, can either be 'down', 'up' or 'all' (default=NULL)}
}
\value{
A ggplot2 object
}
\description{
Plot correlation matrix for ontologies between cell types
}
