% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onthology.R
\name{prepareOnthologyData}
\alias{prepareOnthologyData}
\title{Prepare onthology data}
\usage{
prepareOnthologyData(
  cms,
  de.raw,
  cell.groups,
  universe = NULL,
  org = "human",
  n.top.genes = Inf,
  stat.cutoff = 3,
  transposed = T,
  verbose = T,
  n.cores = 1
)
}
\arguments{
\item{cms}{list of counts matrices; column for gene and row for cell}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{org}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{stat.cutoff}{Cutoff for filtering highly-expressed DE genes (default=3)}

\item{verbose}{Print progress (default=T)}

\item{n.cores}{Number of cores to use (default=1)}

\item{de}{List with differentially expressed genes per cell group}
}
\value{
A list containing DE gene IDs, filtered DE genes, and input DE genes
}
\description{
Filter and prepare DE genes for onthology calculations
}
