% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{prepareOntologyData}
\alias{prepareOntologyData}
\title{Prepare ontology data}
\usage{
prepareOntologyData(
  cms,
  de.raw,
  cell.groups,
  org.db,
  universe = NULL,
  n.top.genes = 100,
  p.adj.cutoff = 0.05,
  transposed = T,
  verbose = T,
  n.cores = 1
)
}
\arguments{
\item{cms}{list of counts matrices; column for gene and row for cell}

\item{de.raw}{List with differentially expressed genes per cell group}

\item{cell.groups}{Vector indicating cell groups with cell names (default: stored vector)}

\item{org.db}{Organism, can be "human", "mouse", "zebrafish", "worm", or "fly" (default="human")}

\item{universe}{Background gene list, NULL will take all input genes in de.raw (default=NULL)}

\item{n.top.genes}{Number of most different genes to take as input. If less are left after filtering for p.adj.cutoff, additional genes are included. To disable, set n.top.genes=0 (default=1e2)}

\item{p.adj.cutoff}{Cutoff for filtering highly-expressed DE genes (default=0.05)}

\item{transposed}{Indicate whether raw counts are transposed, i.e., cells as rows, genes as columns (default=T)}

\item{verbose}{Print progress (default=T)}

\item{n.cores}{Number of cores to use (default=1)}
}
\value{
A list containing DE ENSEMBL gene IDs, and filtered DE genes
}
\description{
Filter and prepare DE genes for ontology calculations
}
