% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{prepareSamplesForDE}
\alias{prepareSamplesForDE}
\title{Prepare samples for DE analysis}
\usage{
prepareSamplesForDE(
  sample.groups,
  resampling.method = c("loo", "bootstrap", "fix.cells", "fix.samples"),
  n.resamplings = 30
)
}
\arguments{
\item{sample.groups}{named list containing sample names}

\item{resampling.method}{one of "loo" (leave-one-out, remove one sample per iteration), "bootstrap", "fix.cells" (fixed number of cells per subsample), or "fix.samples" (fixed number of samples per iteration)}

\item{n.resamplings}{number of iterations (default=30)}
}
\description{
Prepare samples for DE analysis
}
\keyword{internal}
