% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{saveDEasJSON}
\alias{saveDEasJSON}
\title{Save DE results as JSON files}
\usage{
saveDEasJSON(
  de.raw,
  saveprefix = NULL,
  dir.name = "JSON",
  gene.metadata = NULL,
  sample.groups = NULL,
  verbose = TRUE
)
}
\arguments{
\item{de.raw}{List of DE results}

\item{saveprefix}{Prefix for created files (default=NULL)}

\item{dir.name}{Name for directory with results. If it doesn't exist, it will be created. To disable, set as NULL (default="JSON")}

\item{gene.metadata}{(default=NULL)}

\item{sample.groups}{Sample groups as named list, each element containing a vector of samples}

\item{verbose}{Show progress (default=TRUE)}
}
\description{
Save DE results as JSON files
}
