% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{summarizeDEResamplingResults}
\alias{summarizeDEResamplingResults}
\title{Summarize DE Resampling Results}
\usage{
summarizeDEResamplingResults(de.list, var.to.sort = "pvalue")
}
\arguments{
\item{var.to.sort}{Variable to calculate ranks}
}
\description{
Summarize DE Resampling Results
}
