% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{validateDEPerCellTypeParams}
\alias{validateDEPerCellTypeParams}
\title{Validate parameters per cell type}
\usage{
validateDEPerCellTypeParams(raw.mats, cell.groups, sample.groups, ref.level)
}
\arguments{
\item{raw.mats}{List of raw count matrices}

\item{cell.groups}{Named clustering/annotation factor with cell names}

\item{sample.groups}{Named factor with cell names indicating condition/sample, e.g., ctrl/disease}

\item{ref.level}{Reference cluster level in 'sample.groups', e.g., ctrl, healthy, wt}
}
\description{
Validate parameters per cell type
}
